Find Pathway WPIDs By Xref

Retrieve list of pathway WPIDs containing the query Xref by identifier and system code.

findPathwayIdsByXref(identifier, systemCode)

Arguments

identifier
(character) The official ID specified by a data source or system

systemCode
(character) The BridgeDb code associated with the data source or system, e.g., En (Ensembl), L (Entrez), Ch (HMDB), etc. See column two of https://github.com/bridgedb/BridgeDb/blob/master/org.bridgedb.bio/src/main/resources/org/bridgedb/bio/datasources.tsv.

Value

A list of WPIDs

Details

Note: there will be multiple listings of the same pathway if the Xref is present mutiple times.

See also

Examples

{
findPathwayIdsByXref('ENSG00000232810','En')
}

#>  [1] "WP3893" "WP2882" "WP1584" "WP2431" "WP4673" "WP2059" "WP3599" "WP3863"
#>  [9] "WP4008" "WP3865" "WP3624" "WP4341" "WP4298" "WP4666" "WP3941" "WP2813"
#> [17] "WP262"  "WP2873" "WP2571" "WP4329" "WP4136" "WP4864" "WP3287" "WP4066"
#> [25] "WP4063" "WP4969" "WP2795" "WP4847" "WP129"  "WP314"  "WP2513" "WP1449"
#> [33] "WP4481" "WP1544" "WP4482" "WP3297" "WP530"  "WP4891" "WP236"  "WP4747"
#> [41] "WP176"  "WP727"  "WP4496" "WP254"  "WP1533" "WP75"   "WP4478" "WP2328"
#> [49] "WP3851" "WP231"  "WP3617" "WP2036" "WP3611" "WP560"  "WP4816" "WP2864"
#> [57] "WP4304" "WP4857" "WP15"   "WP2586" "WP4493" "WP4396" "WP4480" "WP706" 
#> [65] "WP2447" "WP382"  "WP5044" "WP2751" "WP3380" "WP3398" "WP4066" "WP4063"
#> [73] "WP4754" "WP5039"