Find Pathway Names By Literature

Retrieve list of pathway names containing the query citation.

findPathwayNamesByLiterature(query)

Arguments

query
The character string to search for, e.g., a PMID, title keyword or author name.

Value

A list of lists

Details

Note: there will be multiple listings of the same pathway name if copies exist for multiple species.

See also

Examples

{
findPathwayNamesByLiterature('19649250')
findPathwayNamesByLiterature('smith')
findPathwayNamesByLiterature('cancer')
}

#>   [1] "Regulation of mRNA stability by proteins that bind AU-rich elements"                                                        
#>   [2] "SUMOylation of DNA damage response and repair proteins"                                                                     
#>   [3] "PI3K Cascade"                                                                                                               
#>   [4] "Opioid Signalling"                                                                                                          
#>   [5] "DAG and IP3 signaling"                                                                                                      
#>   [6] "MAP kinase activation"                                                                                                      
#>   [7] "Regulation of Insulin-like Growth Factor (IGF) transport and uptake by Insulin-like Growth Factor Binding Proteins (IGFBPs)"
#>   [8] "Sialic acid metabolism"                                                                                                     
#>   [9] "Inhibition of exosome biogenesis and secretion by Manumycin A in CRPC cells"                                                
#>  [10] "SUMOylation of transcription cofactors"                                                                                     
#>  [11] "Cell migration and invasion through p75NTR"                                                                                 
#>  [12] "Neural Crest Cell Migration during Development"                                                                             
#>  [13] "Neural Crest Cell Migration in Cancer"                                                                                      
#>  [14] "Anti-inflammatory response favouring Leishmania parasite infection"                                                         
#>  [15] "Pentose Phosphate Pathway in senescent cells"                                                                               
#>  [16] "Codeine and Morphine Metabolism"                                                                                            
#>  [17] "DNA Damage Reversal"                                                                                                        
#>  [18] "GPVI-mediated activation cascade"                                                                                           
#>  [19] "Peroxisomal lipid metabolism"                                                                                               
#>  [20] "Transport of inorganic cations/anions and amino acids/oligopeptides"                                                        
#>  [21] "Alzheimers Disease"                                                                                                         
#>  [22] "Serotonin Receptor 2 and STAT3 Signaling"                                                                                   
#>  [23] "Gastric Cancer Network 1"                                                                                                   
#>  [24] "Gastric Cancer Network 2"                                                                                                   
#>  [25] "Nanoparticle triggered autophagic cell death"                                                                               
#>  [26] "Aryl Hydrocarbon Receptor Netpath"                                                                                          
#>  [27] "Structural Pathway of Interleukin 1 (IL-1)"                                                                                 
#>  [28] "Vasopressin regulates renal water homeostasis via Aquaporins"                                                               
#>  [29] "Signaling by NODAL"                                                                                                         
#>  [30] "Extracellular matrix organization"                                                                                          
#>  [31] "Glycerophospholipid biosynthesis"                                                                                           
#>  [32] "Metabolism of steroid hormones"                                                                                             
#>  [33] "Transcriptional regulation of white adipocyte differentiation"                                                              
#>  [34] "ATF4 activates genes in response to endoplasmic reticulum  stress"                                                          
#>  [35] "Transcriptional activity of SMAD2/SMAD3:SMAD4 heterotrimer"                                                                 
#>  [36] "Signaling by BMP"                                                                                                           
#>  [37] "Gamma carboxylation, hypusine formation and arylsulfatase activation"                                                       
#>  [38] "MyD88 dependent cascade initiated on endosome"                                                                              
#>  [39] "Degradation of beta-catenin by the destruction complex"                                                                     
#>  [40] "Toll-like Receptor Cascades"                                                                                                
#>  [41] "Signaling by Activin"                                                                                                       
#>  [42] "Transcriptional regulation of pluripotent stem cells"                                                                       
#>  [43] "TCA Cycle Nutrient Utilization and Invasiveness of Ovarian Cancer"                                                          
#>  [44] "Serotonin Receptor 2 and STAT3 Signaling"                                                                                   
#>  [45] "Gastric cancer network 2"                                                                                                   
#>  [46] "Gastric Cancer Network 1"                                                                                                   
#>  [47] "Structural Pathway of Interleukin 1 (IL-1)"                                                                                 
#>  [48] "Aryl Hydrocarbon Receptor"                                                                                                  
#>  [49] "Serotonin Receptor 2 and ELK-SRF/GATA4 signaling"                                                                           
#>  [50] "TCA Cycle Nutrient Utilization and Invasiveness of Ovarian Cancer"                                                          
#>  [51] "Codeine and Morphine Metabolism"                                                                                            
#>  [52] "Serotonin Receptor 4/6/7 and NR3C Signaling"                                                                                
#>  [53] "Melatonin metabolism and effects"                                                                                           
#>  [54] "MFAP5-mediated ovarian cancer cell motility and invasiveness"                                                               
#>  [55] "NoRC negatively regulates rRNA expression"                                                                                  
#>  [56] "RHO GTPases activate IQGAPs"                                                                                                
#>  [57] "POU5F1 (OCT4), SOX2, NANOG activate genes related to proliferation"                                                         
#>  [58] "BMAL1:CLOCK,NPAS2 activates circadian gene expression"                                                                      
#>  [59] "TNF signaling"                                                                                                              
#>  [60] "tRNA modification in the nucleus and cytosol"                                                                               
#>  [61] "Gene Silencing by RNA"                                                                                                      
#>  [62] "Class I MHC mediated antigen processing &amp; presentation"                                                                 
#>  [63] "mir-124 predicted interactions with cell cycle and differentiation "                                                        
#>  [64] "miR-517 relationship with ARCN1 and USP1"                                                                                   
#>  [65] "Simplified Interaction Map Between LOXL4 and Oxidative Stress Pathway"                                                      
#>  [66] "Renin - Angiotensin System"                                                                                                 
#>  [67] "Regulation of TP53 Activity through Association with Co-factors"                                                            
#>  [68] "TP53 Regulates Transcription of Cell Death Genes"                                                                           
#>  [69] "Interleukin-20 family signaling"                                                                                            
#>  [70] "Signaling by MST1"                                                                                                          
#>  [71] "Interleukin-17 signaling"                                                                                                   
#>  [72] "Toll-like Receptor Signaling related to MyD88"                                                                              
#>  [73] "miR-509-3p alteration of YAP1/ECM axis"                                                                                     
#>  [74] "H19 action Rb-E2F1 signaling and CDK-Beta-catenin activity"                                                                 
#>  [75] "Prader-Willi and Angelman Syndrome"                                                                                         
#>  [76] "Liver steatosis AOP"                                                                                                        
#>  [77] "Cellular hexose transport"                                                                                                  
#>  [78] "Interleukin-4 and Interleukin-13 signaling"                                                                                 
#>  [79] "Nicotinate metabolism"                                                                                                      
#>  [80] "Deubiquitination"                                                                                                           
#>  [81] "Fatty acyl-CoA biosynthesis"                                                                                                
#>  [82] "Tumor suppressor activity of SMARCB1"                                                                                       
#>  [83] "Pyrimidine metabolism and related diseases"                                                                                 
#>  [84] "Transcriptional Regulation by E2F6"                                                                                         
#>  [85] "Regulation of RUNX2 expression and activity"                                                                                
#>  [86] "Class A/1 (Rhodopsin-like receptors)"                                                                                       
#>  [87] "SUMOylation of intracellular receptors"                                                                                     
#>  [88] "Glycosylation and related congenital defects"                                                                               
#>  [89] "Head and Neck Squamous Cell Carcinoma"                                                                                      
#>  [90] "Thyroid hormones production and their peripheral downstream signalling effects"                                             
#>  [91] "Regulation of Pgc1a expression by a Gsk3b-Tfeb signaling axis in skeletal muscle"                                           
#>  [92] "Hijack of Ubiquitination by SARS-CoV-2"                                                                                     
#>  [93] "Type I Interferon Induction and Signaling During SARS-CoV-2 Infection"                                                      
#>  [94] "Response of EIF2AK4 (GCN2) to amino acid deficiency"                                                                        
#>  [95] "Cell recruitment (pro-inflammatory response)"                                                                               
#>  [96] "WNT5:FZD7-mediated leishmania damping"                                                                                      
#>  [97] "RIPK1-mediated regulated necrosis"                                                                                          
#>  [98] "ATM Signaling Pathway"                                                                                                      
#>  [99] "Serotonin Receptor 2 and ELK-SRF/GATA4 signaling"                                                                           
#> [100] "Serotonin Receptor 2 and STAT3 Signaling"                                                                                   
#> [101] "Serotonin Receptor 4/6/7 and NR3C Signaling"                                                                                
#> [102] "Prostaglandin Synthesis and Regulation"                                                                                     
#> [103] "FAS pathway and Stress induction of HSP regulation"                                                                         
#> [104] "Selenium Metabolism and Selenoproteins"                                                                                     
#> [105] "Estrogen metabolism"                                                                                                        
#> [106] "Selenium metabolism Selenoproteins"                                                                                         
#> [107] "Estrogen metabolism"                                                                                                        
#> [108] "Estrogen metabolism"                                                                                                        
#> [109] "Selenium metabolism Selenoproteins"                                                                                         
#> [110] "Estrogen metabolism"                                                                                                        
#> [111] "Selenium metabolism Selenoproteins"                                                                                         
#> [112] "Influenza A virus infection"                                                                                                
#> [113] "miRNAs involved in DNA damage response"                                                                                     
#> [114] "Apoptosis Modulation and Signaling"                                                                                         
#> [115] "Gap junction trafficking and regulation"                                                                                    
#> [116] "Intrinsic Pathway for Apoptosis"                                                                                            
#> [117] "Metabolism of porphyrins"                                                                                                   
#> [118] "Mitotic G2-G2/M phases"                                                                                                     
#> [119] "Netrin-1 signaling"                                                                                                         
#> [120] "Neurotransmitter release cycle"                                                                                             
#> [121] "Signaling by ROBO receptors"                                                                                                
#> [122] "Telomere Maintenance"                                                                                                       
#> [123] "Unfolded Protein Response (UPR)"                                                                                            
#> [124] "Genetic alterations of lung cancer"                                                                                         
#> [125] "Thyroxine (Thyroid Hormone) Production"                                                                                     
#> [126] "miRNAs involved in DNA damage response"                                                                                     
#> [127] "Quercetin and Nf-kB/ AP-1 Induced Cell Apoptosis"                                                                           
#> [128] "Iron uptake and transport"                                                                                                  
#> [129] "Abacavir transport and metabolism"                                                                                          
#> [130] "Signaling by NOTCH2"                                                                                                        
#> [131] "Signaling by NOTCH1"                                                                                                        
#> [132] "Signaling by NOTCH4"                                                                                                        
#> [133] "Glycosaminoglycan metabolism"                                                                                               
#> [134] "Degradation of the extracellular matrix"                                                                                    
#> [135] "Visual phototransduction"                                                                                                   
#> [136] "Pre-NOTCH Expression and Processing"                                                                                        
#> [137] "Sphingolipid metabolism"                                                                                                    
#> [138] "WNT ligand biogenesis and trafficking"                                                                                      
#> [139] "Selenium Metabolism and Selenoproteins"                                                                                     
#> [140] "Vitamin D Receptor Pathway"                                                                                                 
#> [141] "Nuclear Receptors Meta-Pathway"                                                                                             
#> [142] "NRF2 pathway"                                                                                                               
#> [143] "miRNA targets in ECM and membrane receptors"                                                                                
#> [144] "Fas Ligand (FasL) pathway and Stress induction of Heat Shock Proteins (HSP) regulation"                                     
#> [145] "Apoptosis Modulation and Signaling"                                                                                         
#> [146] "Quercetin and Nf-kB/ AP-1 Induced Cell Apoptosis"                                                                           
#> [147] "NRF2 pathway"                                                                                                               
#> [148] "Neurotransmitter Release Cycle"                                                                                             
#> [149] "Influenza A virus infection"                                                                                                
#> [150] "Iron uptake and transport"                                                                                                  
#> [151] "Thyroxine (Thyroid Hormone) Production"                                                                                     
#> [152] "MAPK6/MAPK4 signaling"                                                                                                      
#> [153] "Hedgehog 'on' state"                                                                                                        
#> [154] "Miscellaneous transport and binding events"                                                                                 
#> [155] "RAF-independent MAPK1/3 activation"                                                                                         
#> [156] "RHO GTPases activate PKNs"                                                                                                  
#> [157] "RHO GTPases Activate Formins"                                                                                               
#> [158] "RHO GTPases Activate WASPs and WAVEs"                                                                                       
#> [159] "Activated PKN1 stimulates transcription of AR (androgen receptor) regulated genes KLK2 and KLK3"                            
#> [160] "Base-Excision Repair, AP Site Formation"                                                                                    
#> [161] "DNA Double Strand Break Response"                                                                                           
#> [162] "tRNA processing in the nucleus"                                                                                             
#> [163] "Sulfur amino acid metabolism"                                                                                               
#> [164] "Surfactant metabolism"                                                                                                      
#> [165] "Metabolism of folate and pterines"                                                                                          
#> [166] "Activation of anterior HOX genes in hindbrain development during early embryogenesis"                                       
#> [167] "SUMOylation of transcription factors"                                                                                       
#> [168] "PI3K-AKT-mTOR signaling pathway and therapeutic opportunities"                                                              
#> [169] "Lipid Metabolism Pathway"                                                                                                   
#> [170] "PDGFR-beta pathway"                                                                                                         
#> [171] "mir-193a and MVP in colon cancer metastasis"                                                                                
#> [172] "miRNA regulation of prostate cancer signaling pathways"                                                                     
#> [173] "miRNA regulation of p53 pathway in prostate cancer"                                                                         
#> [174] "Nucleobase catabolism"                                                                                                      
#> [175] "PTEN Regulation"                                                                                                            
#> [176] "Phase I - Functionalization of compounds"                                                                                   
#> [177] "Transcriptional regulation by the AP-2 (TFAP2) family of transcription factors"                                             
#> [178] "Chromatin organization"                                                                                                     
#> [179] "Antimicrobial peptides"                                                                                                     
#> [180] "Transcriptional regulation by RUNX3"                                                                                        
#> [181] "Neddylation"                                                                                                                
#> [182] "RUNX1 regulates genes involved in megakaryocyte differentiation and platelet function"                                      
#> [183] "LncRNA involvement in canonical Wnt signaling and colorectal cancer"                                                        
#> [184] "Pancreatic adenocarcinoma pathway"                                                                                          
#> [185] "Genotoxicity pathway"                                                                                                       
#> [186] "Malate-aspartate shuttle"                                                                                                   
#> [187] "Eicosanoid metabolism via Cyclo Oxygenases (COX)"                                                                           
#> [188] "Eicosanoid metabolism via Lipo Oxygenases (LOX)"                                                                            
#> [189] "Transcriptional Regulation by MECP2"                                                                                        
#> [190] "Elongation of (very) long chain fatty acids"                                                                                
#> [191] "Cancer immunotherapy by CTLA4 blockade"                                                                                     
#> [192] "22q11.2 copy number variation syndrome"                                                                                     
#> [193] "Eicosanoid metabolism via Cyclo Oxygenases (COX)"                                                                           
#> [194] "Eicosanoid metabolism via Lipo Oxygenases (LOX)"                                                                            
#> [195] "Netrin-UNC5B signaling Pathway"                                                                                             
#> [196] "Endothelin Pathway"                                                                                                         
#> [197] "CAMKK2  Pathway"                                                                                                            
#> [198] "1q21.1 copy number variation syndrome"                                                                                      
#> [199] "3q29 copy number variation syndrome"                                                                                        
#> [200] "7q11.23 copy number variation syndrome"                                                                                     
#> [201] "DNA Repair Pathways Full Network"                                                                                           
#> [202] "16p11.2 proximal deletion syndrome"                                                                                         
#> [203] "Galanin receptor pathway"                                                                                                   
#> [204] "Aspartate and asparagine metabolism"                                                                                        
#> [205] "FOXO-mediated transcription of cell death genes"                                                                            
#> [206] "FOXO-mediated transcription of cell cycle genes"                                                                            
#> [207] "Glutamate and glutamine metabolism"                                                                                         
#> [208] "Transcriptional Regulation by VENTX"                                                                                        
#> [209] "Transcriptional regulation of granulopoiesis"                                                                               
#> [210] "FOXO-mediated transcription"                                                                                                
#> [211] "SARS-CoV-1 Infection"                                                                                                       
#> [212] "SARS-CoV-2 Infection"                                                                                                       
#> [213] "Glycolysis in senescence"                                                                                                   
#> [214] "RHOBTB GTPase Cycle"                                                                                                        
#> [215] "Glycolysis in senescence"                                                                                                   
#> [216] "Estrogen metabolism"                                                                                                        
#> [217] "Serotonin HTR1 Group and FOS Pathway"                                                                                       
#> [218] "Estrogen metabolism"                                                                                                        
#> [219] "Estrogen metabolism"                                                                                                        
#> [220] "Transport of bile salts and organic acids, metal ions and amine compounds"                                                  
#> [221] "MicroRNA for Targeting Cancer Growth and Vascularization in Glioblastoma"                                                   
#> [222] "Hereditary leiomyomatosis and renal cell carcinoma pathway"                                                                 
#> [223] "Modified nucleosides derived from t-RNA as urinary cancer markers"                                                          
#> [224] "MFAP5 effect on permeability and motility of endothelial cells via cytoskeleton rearrangement"                              
#> [225] "ErbB Signaling Pathway"                                                                                                     
#> [226] "TGF-beta Receptor Signaling"                                                                                                
#> [227] "Folate Metabolism"                                                                                                          
#> [228] "TGF Beta Signaling Pathway"                                                                                                 
#> [229] "Transforming Growth Factor (TGF) Beta Signaling Pathway"                                                                    
#> [230] "Transforming Growth Factor (TGF) Beta Signaling Pathway"                                                                    
#> [231] "Selenium Micronutrient Network"                                                                                             
#> [232] "Vitamin B12 Metabolism"                                                                                                     
#> [233] "Fluoropyrimidine Activity"                                                                                                  
#> [234] "Folate Metabolism"                                                                                                          
#> [235] "PluriNetWork: mechanisms associated with pluripotency"                                                                      
#> [236] "Asparagine N-linked glycosylation"                                                                                          
#> [237] "Costimulation by the CD28 family"                                                                                           
#> [238] "Metabolism of water-soluble vitamins and cofactors"                                                                         
#> [239] "Mitotic G1 phase and G1/S transition"                                                                                       
#> [240] "Platelet Aggregation (Plug Formation)"                                                                                      
#> [241] "Processing of Capped Intron-Containing Pre-mRNA"                                                                            
#> [242] "Integrated Cancer Pathway"                                                                                                  
#> [243] "Nucleotide Excision Repair"                                                                                                 
#> [244] "Integrated Breast Cancer Pathway"                                                                                           
#> [245] "Deregulation of Rab and Rab Effector Genes in Bladder Cancer"                                                               
#> [246] "Cellular response to hypoxia"                                                                                               
#> [247] "Inositol phosphate metabolism"                                                                                              
#> [248] "S Phase"                                                                                                                    
#> [249] "Mammary gland development pathway - Embryonic development (Stage 1 of 4)"                                                   
#> [250] "Deregulation of Rab and Rab Effector Genes in Bladder Cancer"                                                               
#> [251] "Vitamin B12 Metabolism"                                                                                                     
#> [252] "Senescence and Autophagy in Cancer"                                                                                         
#> [253] "Integrated Cancer pathway"                                                                                                  
#> [254] "Selenium Micronutrient Network"                                                                                             
#> [255] "Integrated Breast Cancer Pathway"                                                                                           
#> [256] "Tamoxifen metabolism"                                                                                                       
#> [257] "Fluoropyrimidine Activity"                                                                                                  
#> [258] "RAC1/PAK1/p38/MMP2 Pathway"                                                                                                 
#> [259] "Caspase activation via Dependence Receptors in the absence of ligand"                                                       
#> [260] "DNA methylation"                                                                                                            
#> [261] "TET1,2,3 and TDG demethylate DNA"                                                                                           
#> [262] "Mismatch Repair"                                                                                                            
#> [263] "Abnormal conversion of 2-oxoglutarate to 2-hydroxyglutarate"                                                                
#> [264] "Zinc homeostasis"                                                                                                           
#> [265] "Beta-catenin independent WNT signaling"                                                                                     
#> [266] "TP53 Regulates Transcription of Cell Cycle Genes"                                                                           
#> [267] "Regulation of TP53 Expression and Degradation"                                                                              
#> [268] "Regulation of TP53 Activity through Phosphorylation"                                                                        
#> [269] "Mevalonate pathway"                                                                                                         
#> [270] "Pathways in clear cell renal cell carcinoma"                                                                                
#> [271] "Pentose phosphate pathway"                                                                                                  
#> [272] "RUNX1 and FOXP3 control the development of regulatory T lymphocytes (Tregs)"                                                
#> [273] "Regulation of RUNX1 Expression and Activity"                                                                                
#> [274] "Protein ubiquitination"                                                                                                     
#> [275] "Endometrial cancer"                                                                                                         
#> [276] "PI3K-Akt Signaling Pathway"                                                                                                 
#> [277] "MET in type 1 papillary renal cell carcinoma"                                                                               
#> [278] "Ebola Virus Pathway on Host"                                                                                                
#> [279] "Non-small cell lung cancer"                                                                                                 
#> [280] "Breast cancer pathway"                                                                                                      
#> [281] "Rett syndrome causing genes"                                                                                                
#> [282] "FABP4 in ovarian cancer"                                                                                                    
#> [283] "Interleukin-15 signaling"                                                                                                   
#> [284] "Integrin signaling"                                                                                                         
#> [285] "Signaling by Erythropoietin"                                                                                                
#> [286] "Oxysterols derived from cholesterol"                                                                                        
#> [287] "Fragile X Syndrome "                                                                                                        
#> [288] "Interactions between immune cells and microRNAs in tumor microenvironment"                                                  
#> [289] "Neurodegeneration with brain iron accumulation (NBIA) subtypes pathway"                                                     
#> [290] "EGFR Tyrosine Kinase Inhibitor Resistance"                                                                                  
#> [291] "TGF-beta Receptor Signalling in Skeletal Dysplasias"                                                                        
#> [292] "LDLRAD4 and what we know about it"                                                                                          
#> [293] "NGF-stimulated transcription"                                                                                               
#> [294] "Metabolism of cofactors"                                                                                                    
#> [295] "TGF Beta Signaling Pathway"                                                                                                 
#> [296] "TGF-beta Receptor Signaling"                                                                                                
#> [297] "Senescence and Autophagy in Cancer"                                                                                         
#> [298] "p53 pathway"                                                                                                                
#> [299] "ErbB Signaling Pathway"                                                                                                     
#> [300] "Tamoxifen metabolism"                                                                                                       
#> [301] "Transforming Growth Factor (TGF) Beta Signaling Pathway"                                                                    
#> [302] "Tamoxifen metabolism"                                                                                                       
#> [303] "Transforming Growth Factor (TGF) Beta Signaling Pathway"                                                                    
#> [304] "Aflatoxin B1 metabolism"                                                                                                    
#> [305] "Aflatoxin B1 metabolism"                                                                                                    
#> [306] "Aflatoxin B1 metabolism"                                                                                                    
#> [307] "Generic Transcription Pathway"                                                                                              
#> [308] "Interleukin-3, Interleukin-5 and GM-CSF signaling"                                                                          
#> [309] "NCAM signaling for neurite out-growth"                                                                                      
#> [310] "Signaling by NTRK1 (TRKA)"                                                                                                  
#> [311] "Signaling by EGFR"                                                                                                          
#> [312] "Signaling by Insulin receptor"                                                                                              
#> [313] "Signaling by PDGF"                                                                                                          
#> [314] "Signaling by VEGF"                                                                                                          
#> [315] "Glycerophospholipid Biosynthetic Pathway"                                                                                   
#> [316] "PIP3 activates AKT signaling"                                                                                               
#> [317] "Gastrin-CREB signalling pathway via PKC and MAPK"                                                                           
#> [318] "Signaling by Type 1 Insulin-like Growth Factor 1 Receptor (IGF1R)"                                                          
#> [319] "DAP12 interactions"                                                                                                         
#> [320] "Signaling by SCF-KIT"                                                                                                       
#> [321] "Signaling by Hippo"                                                                                                         
#> [322] "Signaling by NOTCH3"                                                                                                        
#> [323] "Interleukin-2 family signaling"                                                                                             
#> [324] "YAP1- and WWTR1 (TAZ)-stimulated gene expression"                                                                           
#> [325] "PI Metabolism"                                                                                                              
#> [326] "Fc epsilon receptor (FCERI) signaling"                                                                                      
#> [327] "Signaling by ERBB2"                                                                                                         
#> [328] "Signaling by ERBB4"                                                                                                         
#> [329] "exRNA mechanism of action and biogenesis"                                                                                   
#> [330] "Apoptosis-related network due to altered Notch3 in ovarian cancer"                                                          
#> [331] "Apoptosis-related network due to altered Notch3 in ovarian cancer"                                                          
#> [332] "RNA interference"                                                                                                           
#> [333] "Glycerophospholipid Biosynthetic Pathway"                                                                                   
#> [334] "Aflatoxin B1 metabolism"                                                                                                    
#> [335] "Oncogene Induced Senescence"                                                                                                
#> [336] "O-linked glycosylation"                                                                                                     
#> [337] "TCF dependent signaling in response to WNT"                                                                                 
#> [338] "Signaling by Leptin"                                                                                                        
#> [339] "Senescence-Associated Secretory Phenotype (SASP)"                                                                           
#> [340] "Oxidative Stress Induced Senescence"                                                                                        
#> [341] "DNA Damage/Telomere Stress Induced Senescence"                                                                              
#> [342] "HDR through Homologous Recombination (HRR) or Single Strand Annealing (SSA)"                                                
#> [343] "Fanconi Anemia Pathway"                                                                                                     
#> [344] "TP53 Regulates Transcription of DNA Repair Genes"                                                                           
#> [345] "RAB GEFs exchange GTP for GDP on RABs"                                                                                      
#> [346] "RET signaling"                                                                                                              
#> [347] "Transcriptional regulation by RUNX1"                                                                                        
#> [348] "Phase II - Conjugation of compounds"                                                                                        
#> [349] "Signaling by MET"                                                                                                           
#> [350] "Chromosomal and microsatellite instability in colorectal cancer "                                                           
#> [351] "Epithelial to mesenchymal transition in colorectal cancer"                                                                  
#> [352] "Regulation of sister chromatid separation at the metaphase-anaphase transition"                                             
#> [353] "Type 2 papillary renal cell carcinoma"                                                                                      
#> [354] "Transcriptional regulation by RUNX2"                                                                                        
#> [355] "ESR-mediated signaling"                                                                                                     
#> [356] "Signaling by NTRK2 (TRKB)"                                                                                                  
#> [357] "Signaling by PTK6"                                                                                                          
#> [358] "Signaling by NTRK3 (TRKC)"                                                                                                  
#> [359] "Cancer immunotherapy by PD-1 blockade"                                                                                      
#> [360] "Small cell lung cancer"                                                                                                     
#> [361] "AXL signaling pathway"                                                                                                      
#> [362] "MAPK pathway in congenital thyroid cancer"                                                                                  
#> [363] "Extra-nuclear estrogen signaling"                                                                                           
#> [364] "Blood group systems biosynthesis"                                                                                           
#> [365] "Aflatoxin B1 metabolism"                                                                                                    
#> [366] "Aflatoxin B1 metabolism"                                                                                                    
#> [367] "miRNA Regulation of DNA Damage Response"                                                                                    
#> [368] "DNA damage response"                                                                                                        
#> [369] "DNA Damage Response"                                                                                                        
#> [370] "DNA Damage Response"                                                                                                        
#> [371] "Integrin-mediated Cell Adhesion"                                                                                            
#> [372] "Keap1-Nrf2"                                                                                                                 
#> [373] "Keap1-Nrf2"                                                                                                                 
#> [374] "Keap1-Nrf2"                                                                                                                 
#> [375] "Keap1-Nrf2"                                                                                                                 
#> [376] "Integrin-mediated Cell Adhesion"                                                                                            
#> [377] "Signaling by Rho GTPases"                                                                                                   
#> [378] "Metastatic brain tumor"                                                                                                     
#> [379] "Parkinsons Disease Pathway"                                                                                                 
#> [380] "ATM Signaling Pathway"                                                                                                      
#> [381] "Energy dependent regulation of mTOR by LKB1-AMPK"                                                                           
#> [382] "Bladder Cancer"                                                                                                             
#> [383] "Phytochemical activity on NRF2 transcriptional activation"                                                                  
#> [384] "Bladder Cancer"                                                                                                             
#> [385] "ATM Signaling Pathway"                                                                                                      
#> [386] "Metastatic brain tumor"                                                                                                     
#> [387] "MTOR signalling"                                                                                                            
#> [388] "Signaling by FGFR4"                                                                                                         
#> [389] "Signaling by FGFR3"                                                                                                         
#> [390] "Signaling by FGFR2"                                                                                                         
#> [391] "TP53 Regulates Metabolic Genes"                                                                                             
#> [392] "Hedgehog ligand biogenesis"                                                                                                 
#> [393] "Parkinsons Disease Pathway"                                                                                                 
#> [394] "DNA IR-Double Strand Breaks (DSBs) and cellular response via ATM"                                                           
#> [395] "DNA IR-damage and cellular response via ATR"                                                                                
#> [396] "NRF2-ARE regulation"                                                                                                        
#> [397] "Overview of leukocyte-intrinsic Hippo pathway functions"                                                                    
#> [398] "FLT3 Signaling"                                                                                                             
#> [399] "Keap1-Nrf2"                                                                                                                 
#> [400] "Retinol metabolism"                                                                                                         
#> [401] "Keap1-Nrf2"                                                                                                                 
#> [402] "Transcriptional activation by NRF2"                                                                                         
#> [403] "Irinotecan Pathway"                                                                                                         
#> [404] "Irinotecan Pathway"                                                                                                         
#> [405] "Irinotecan Pathway"                                                                                                         
#> [406] "Irinotecan Pathway"                                                                                                         
#> [407] "Irinotecan Pathway"                                                                                                         
#> [408] "Irinotecan Pathway"                                                                                                         
#> [409] "RAF/MAP kinase cascade"                                                                                                     
#> [410] "E3 ubiquitin ligases ubiquitinate target proteins"                                                                          
#> [411] "Signaling by FGFR1"