Find Pathway Names By Text

Retrieve list of pathway names containing the query text.

findPathwayNamesByText(query)

Arguments

query
A character string to search for, e.g., "cancer"

Value

A list of lists

Details

Note: there will be multiple listings of the same pathway name if copies exist for multiple species.

Examples

{
findPathwayNamesByText('cancer')
}

#>   [1] "Integrated Breast Cancer Pathway"                                                                    
#>   [2] "Integrated Breast Cancer Pathway"                                                                    
#>   [3] "Breast cancer pathway"                                                                               
#>   [4] "TCA Cycle Nutrient Utilization and Invasiveness of Ovarian Cancer"                                   
#>   [5] "TCA Cycle Nutrient Utilization and Invasiveness of Ovarian Cancer"                                   
#>   [6] "MFAP5-mediated ovarian cancer cell motility and invasiveness"                                        
#>   [7] "Integrated Cancer pathway"                                                                           
#>   [8] "Integrated Cancer Pathway"                                                                           
#>   [9] "Bladder Cancer"                                                                                      
#>  [10] "Integrated Lung Cancer Pathway"                                                                      
#>  [11] "Genetic alterations of lung cancer"                                                                  
#>  [12] "Retinoblastoma Gene in Cancer"                                                                       
#>  [13] "Retinoblastoma Gene in Cancer"                                                                       
#>  [14] "Endometrial cancer"                                                                                  
#>  [15] "Androgen Receptor Network in Prostate Cancer"                                                        
#>  [16] "miRNA regulation of prostate cancer signaling pathways"                                              
#>  [17] "miRNA regulation of p53 pathway in prostate cancer"                                                  
#>  [18] "Gastric Cancer Network 2"                                                                            
#>  [19] "FABP4 in ovarian cancer"                                                                             
#>  [20] "Senescence and Autophagy in Cancer"                                                                  
#>  [21] "Bladder Cancer"                                                                                      
#>  [22] "Folate-Alcohol and Cancer Pathway"                                                                   
#>  [23] "Folate-Alcohol and Cancer Pathway Hypotheses"                                                        
#>  [24] "Gastric cancer network 2"                                                                            
#>  [25] "Gastric Cancer Network 1"                                                                            
#>  [26] "Gastric Cancer Network 1"                                                                            
#>  [27] "Cancer immunotherapy by PD-1 blockade"                                                               
#>  [28] "Metabolic reprogramming in colon cancer"                                                             
#>  [29] "Small cell lung cancer"                                                                              
#>  [30] "Ultraconserved region 339 modulation of tumor suppressor microRNAs in cancer"                        
#>  [31] "Deregulation of Rab and Rab Effector Genes in Bladder Cancer"                                        
#>  [32] "Deregulation of Rab and Rab Effector Genes in Bladder Cancer"                                        
#>  [33] "MicroRNA for Targeting Cancer Growth and Vascularization in Glioblastoma"                            
#>  [34] "Epithelial to mesenchymal transition in colorectal cancer"                                           
#>  [35] "mir-193a and MVP in colon cancer metastasis"                                                         
#>  [36] "Non-small cell lung cancer"                                                                          
#>  [37] "Cancer immunotherapy by CTLA4 blockade"                                                              
#>  [38] "Chromosomal and microsatellite instability in colorectal cancer "                                    
#>  [39] " Lipids measured in liver metastasis from breast cancer"                                             
#>  [40] "Modified nucleosides derived from t-RNA as urinary cancer markers"                                   
#>  [41] "Neural Crest Cell Migration in Cancer"                                                               
#>  [42] "LncRNA involvement in canonical Wnt signaling and colorectal cancer"                                 
#>  [43] "MFAP5 effect on permeability and motility of endothelial cells via cytoskeleton rearrangement"       
#>  [44] "Apoptosis-related network due to altered Notch3 in ovarian cancer"                                   
#>  [45] "Apoptosis-related network due to altered Notch3 in ovarian cancer"                                   
#>  [46] "4-hydroxytamoxifen, Dexamethasone, and Retinoic Acids Regulation of p27 Expression"                  
#>  [47] "Fluoropyrimidine Activity"                                                                           
#>  [48] "Fluoropyrimidine Activity"                                                                           
#>  [49] "Model for regulation of MSMP expression in cancer cells and its proangiogenic role in ovarian tumors"
#>  [50] "cancer"                                                                                              
#>  [51] "ncRNAs involved in STAT3 signaling in hepatocellular carcinoma"                                      
#>  [52] "ErbB Signaling Pathway"                                                                              
#>  [53] "Melatonin metabolism and effects"                                                                    
#>  [54] "ErbB signaling pathway"                                                                              
#>  [55] "ErbB signaling pathway"                                                                              
#>  [56] "ErbB signaling pathway"                                                                              
#>  [57] "ErbB Signaling Pathway"                                                                              
#>  [58] "ErbB signaling pathway"                                                                              
#>  [59] "ErbB signaling pathway"                                                                              
#>  [60] "Association Between Physico-Chemical Features and Toxicity Associated Pathways"                      
#>  [61] "LncRNA-mediated mechanisms of therapeutic resistance"                                                
#>  [62] "ErbB Signaling Pathway"                                                                              
#>  [63] "inflammation and cancer"                                                                             
#>  [64] "Netrin-UNC5B signaling Pathway"                                                                      
#>  [65] "ncRNAs involved in Wnt signaling in hepatocellular carcinoma"                                        
#>  [66] "Interactions between immune cells and microRNAs in tumor microenvironment"                           
#>  [67] "H19 action Rb-E2F1 signaling and CDK-Beta-catenin activity"                                          
#>  [68] "TNF-alpha NF-kB Signaling Pathway"                                                                   
#>  [69] "TNF-alpha NF-kB Signaling Pathway"                                                                   
#>  [70] "TNF-alpha NF-kB Signaling Pathway"                                                                   
#>  [71] "TNF-alpha NF-kB Signaling Pathway"                                                                   
#>  [72] "Senescence and Autophagy in Cancer"                                                                  
#>  [73] "inflammation leading to cancer"                                                                      
#>  [74] "miR-509-3p alteration of YAP1/ECM axis"                                                              
#>  [75] "Tamoxifen metabolism"                                                                                
#>  [76] "PDGFR-beta pathway"                                                                                  
#>  [77] "RIOK1 and RIOK2 in EGFR- and PI3K-mediated tumorigenesis"                                            
#>  [78] "TP53 Network"                                                                                        
#>  [79] "Photodynamic therapy-induced AP-1 survival signaling."                                               
#>  [80] "TP53 Network"                                                                                        
#>  [81] "TGF-B Signaling in Thyroid Cells for Epithelial-Mesenchymal Transition"                              
#>  [82] "Regulation of Apoptosis by Parathyroid Hormone-related Protein"                                      
#>  [83] "Photodynamic therapy-induced unfolded protein response"                                              
#>  [84] "Photodynamic therapy-induced HIF-1 survival signaling"                                               
#>  [85] "Photodynamic therapy-induced NF-kB survival signaling"                                               
#>  [86] "Photodynamic therapy-induced NFE2L2 (NRF2) survival signaling"                                       
#>  [87] "RAC1/PAK1/p38/MMP2 Pathway"                                                                          
#>  [88] "Hereditary leiomyomatosis and renal cell carcinoma pathway"                                          
#>  [89] "Amplification and Expansion of Oncogenic Pathways as Metastatic Traits"                              
#>  [90] "Toll-like Receptor Signaling related to MyD88"                                                       
#>  [91] "Inhibition of exosome biogenesis and secretion by Manumycin A in CRPC cells"                         
#>  [92] "Extracellular vesicle-mediated signaling in recipient cells"                                         
#>  [93] "Mechanoregulation and pathology of YAP/TAZ via Hippo and non-Hippo mechanisms"                       
#>  [94] "Extracellular vesicle-mediated signaling in recipient cells"                                         
#>  [95] "Aflatoxin B1 metabolism"                                                                             
#>  [96] "Tamoxifen metabolism"                                                                                
#>  [97] "Simplified Interaction Map Between LOXL4 and Oxidative Stress Pathway"                               
#>  [98] "Tumor suppressor activity of SMARCB1"                                                                
#>  [99] "Irinotecan Pathway"                                                                                  
#> [100] "Overview of leukocyte-intrinsic Hippo pathway functions"                                             
#> [101] "Irinotecan Pathway"                                                                                  
#> [102] "PI3K-Akt Signaling Pathway"                                                                          
#> [103] "PDGF Pathway"                                                                                        
#> [104] "PDGF Pathway"                                                                                        
#> [105] "aspirin and miRNAs"                                                                                  
#> [106] "DNA IR-damage and cellular response via ATR"                                                         
#> [107] "Pancreatic adenocarcinoma pathway"                                                                   
#> [108] "Signaling Pathways in Glioblastoma"                                                                  
#> [109] "Signaling Pathways in Glioblastoma"                                                                  
#> [110] "Overview of nanoparticle effects"                                                                    
#> [111] "Gastrin Signaling Pathway"                                                                           
#> [112] "Canonical wnt signaling"                                                                             
#> [113] "Corticotropin-releasing hormone signaling pathway"                                                   
#> [114] "Corticotropin-releasing hormone"                                                                     
#> [115] "IL-18 signaling pathway"