Find Pathway Names By Xref

Retrieve list of pathway names containing the query Xref by identifier and system code.

findPathwayNamesByXref(identifier, systemCode)

Arguments

identifier
(character) The official ID specified by a data source or system

systemCode
(character) The BridgeDb code associated with the data source or system, e.g., En (Ensembl), L (Entrez), Ch (HMDB), etc. See column two of https://github.com/bridgedb/BridgeDb/blob/master/org.bridgedb.bio/resources/org/bridgedb/bio/datasources.txt.

Value

A list of lists

Details

Note: there will be multiple listings of the same pathway if the Xref is present mutiple times.

See also

Examples

{
findPathwayNamesByXref('ENSG00000232810','En')
}

#>  [1] "Development and heterogeneity of the ILC family"                                        
#>  [2] "Nuclear Receptors Meta-Pathway"                                                         
#>  [3] "Type II diabetes mellitus"                                                              
#>  [4] "Spinal Cord Injury"                                                                     
#>  [5] "Genes involved in male infertility"                                                     
#>  [6] "Alzheimers Disease"                                                                     
#>  [7] "Transcription factor regulation in adipogenesis"                                        
#>  [8] "T-Cell antigen Receptor (TCR) pathway during Staphylococcus aureus infection"           
#>  [9] "NO/cGMP/PKG mediated Neuroprotection"                                                   
#> [10] "Novel intracellular components of RIG-I-like receptor (RLR) pathway"                    
#> [11] "Lung fibrosis"                                                                          
#> [12] "Non-genomic actions of 1,25 dihydroxyvitamin D3"                                        
#> [13] "Viral Acute Myocarditis"                                                                
#> [14] "Hepatitis B infection"                                                                  
#> [15] "Oxidative Damage"                                                                       
#> [16] "Mammary gland development pathway - Embryonic development (Stage 1 of 4)"               
#> [17] "EBV LMP1 signaling"                                                                     
#> [18] "Aryl Hydrocarbon Receptor Pathway"                                                      
#> [19] "Polycystic Kidney Disease Pathway"                                                      
#> [20] "miRNAs involvement in the immune response in sepsis"                                    
#> [21] "Fibrin Complement Receptor 3 Signaling Pathway"                                         
#> [22] "Host-pathogen interaction of human corona viruses - apoptosis"                          
#> [23] "Overview of nanoparticle effects"                                                       
#> [24] "Interleukin-4 and Interleukin-13 signaling"                                             
#> [25] "Interleukin-10 signaling"                                                               
#> [26] "RAS and Bradykinin pathways in COVID-19"                                                
#> [27] "Cardiac Hypertrophic Response"                                                          
#> [28] "AXL signaling pathway"                                                                  
#> [29] "Matrix Metalloproteinases"                                                              
#> [30] "Fas Ligand (FasL) pathway and Stress induction of Heat Shock Proteins (HSP) regulation" 
#> [31] "Nanoparticle triggered regulated necrosis"                                              
#> [32] "Regulation of toll-like receptor signaling pathway"                                     
#> [33] "Resistin as a regulator of inflammation"                                                
#> [34] "MicroRNAs in cardiomyocyte hypertrophy"                                                 
#> [35] "Vitamin D in inflammatory diseases"                                                     
#> [36] "EV release from cardiac cells and their functional effects"                             
#> [37] "Cytokines and Inflammatory Response"                                                    
#> [38] "COVID-19 Adverse Outcome Pathway"                                                       
#> [39] "Adipogenesis"                                                                           
#> [40] "Netrin-UNC5B signaling Pathway"                                                         
#> [41] "Folate Metabolism"                                                                      
#> [42] "Monoamine Transport"                                                                    
#> [43] "Signal transduction through IL1R"                                                       
#> [44] "Apoptosis"                                                                              
#> [45] "Vitamin B12 Metabolism"                                                                 
#> [46] "Toll-like Receptor Signaling Pathway"                                                   
#> [47] "LTF danger signal response pathway"                                                     
#> [48] "Allograft Rejection"                                                                    
#> [49] "TLR4 Signaling and Tolerance"                                                           
#> [50] "TNF alpha Signaling Pathway"                                                            
#> [51] "Photodynamic therapy-induced NF-kB survival signaling"                                  
#> [52] "TNF related weak inducer of apoptosis (TWEAK) Signaling Pathway"                        
#> [53] "Photodynamic therapy-induced AP-1 survival signaling."                                  
#> [54] "TGF-beta Receptor Signaling"                                                            
#> [55] "TGF-beta Receptor Signalling in Skeletal Dysplasias"                                    
#> [56] "Apoptosis-related network due to altered Notch3 in ovarian cancer"                      
#> [57] "Oligodendrocyte specification and differentiation, leading to Myelin Components for CNS"
#> [58] "Endothelin Pathway"                                                                     
#> [59] "Selenium Micronutrient Network"                                                         
#> [60] "Aryl Hydrocarbon Receptor Netpath"                                                      
#> [61] "Cells and Molecules involved in local acute inflammatory response "                     
#> [62] "Nonalcoholic fatty liver disease"                                                       
#> [63] "Role of Altered Glycolysation of MUC1 in Tumour Microenvironment"                       
#> [64] "Sudden Infant Death Syndrome (SIDS) Susceptibility Pathways"                            
#> [65] "Amyotrophic lateral sclerosis (ALS)"                                                    
#> [66] "MAPK Signaling Pathway"                                                                 
#> [67] "Kynurenine Pathway and links to Cellular Senescence"                                    
#> [68] "Transcriptional regulation of white adipocyte differentiation"                          
#> [69] "TNF signaling"                                                                          
#> [70] "TNFR2 non-canonical NF-kB pathway"                                                      
#> [71] "Interleukin-4 and Interleukin-13 signaling"                                             
#> [72] "Interleukin-10 signaling"                                                               
#> [73] "IL-18 signaling pathway"                                                                
#> [74] "SARS-CoV-2 Innate Immunity Evasion and Cell-specific immune response"