Find Pathway URLs By Xref

Retrieve list of pathway URLs containing the query Xref by identifier and system code.

findPathwayUrlsByXref(identifier, systemCode)

Arguments

identifier
(character) The official ID specified by a data source or system

systemCode
(character) The BridgeDb code associated with the data source or system, e.g., En (Ensembl), L (Entrez), Ch (HMDB), etc. See column two of https://github.com/bridgedb/BridgeDb/blob/master/org.bridgedb.bio/resources/org/bridgedb/bio/datasources.txt.

Value

A list of lists

Details

Note: there will be multiple listings of the same pathway if the Xref is present mutiple times.

See also

Examples

{
findPathwayUrlsByXref('ENSG00000232810','En')
}

#>  [1] "https://www.wikipathways.org/index.php/Pathway:WP3893"
#>  [2] "https://www.wikipathways.org/index.php/Pathway:WP2882"
#>  [3] "https://www.wikipathways.org/index.php/Pathway:WP1584"
#>  [4] "https://www.wikipathways.org/index.php/Pathway:WP2431"
#>  [5] "https://www.wikipathways.org/index.php/Pathway:WP4673"
#>  [6] "https://www.wikipathways.org/index.php/Pathway:WP2059"
#>  [7] "https://www.wikipathways.org/index.php/Pathway:WP3599"
#>  [8] "https://www.wikipathways.org/index.php/Pathway:WP3863"
#>  [9] "https://www.wikipathways.org/index.php/Pathway:WP4008"
#> [10] "https://www.wikipathways.org/index.php/Pathway:WP3865"
#> [11] "https://www.wikipathways.org/index.php/Pathway:WP3624"
#> [12] "https://www.wikipathways.org/index.php/Pathway:WP4341"
#> [13] "https://www.wikipathways.org/index.php/Pathway:WP4298"
#> [14] "https://www.wikipathways.org/index.php/Pathway:WP4666"
#> [15] "https://www.wikipathways.org/index.php/Pathway:WP3941"
#> [16] "https://www.wikipathways.org/index.php/Pathway:WP2813"
#> [17] "https://www.wikipathways.org/index.php/Pathway:WP262" 
#> [18] "https://www.wikipathways.org/index.php/Pathway:WP2873"
#> [19] "https://www.wikipathways.org/index.php/Pathway:WP2571"
#> [20] "https://www.wikipathways.org/index.php/Pathway:WP4329"
#> [21] "https://www.wikipathways.org/index.php/Pathway:WP4136"
#> [22] "https://www.wikipathways.org/index.php/Pathway:WP4864"
#> [23] "https://www.wikipathways.org/index.php/Pathway:WP3287"
#> [24] "https://www.wikipathways.org/index.php/Pathway:WP4066"
#> [25] "https://www.wikipathways.org/index.php/Pathway:WP4063"
#> [26] "https://www.wikipathways.org/index.php/Pathway:WP4969"
#> [27] "https://www.wikipathways.org/index.php/Pathway:WP2795"
#> [28] "https://www.wikipathways.org/index.php/Pathway:WP4847"
#> [29] "https://www.wikipathways.org/index.php/Pathway:WP129" 
#> [30] "https://www.wikipathways.org/index.php/Pathway:WP314" 
#> [31] "https://www.wikipathways.org/index.php/Pathway:WP2513"
#> [32] "https://www.wikipathways.org/index.php/Pathway:WP1449"
#> [33] "https://www.wikipathways.org/index.php/Pathway:WP4481"
#> [34] "https://www.wikipathways.org/index.php/Pathway:WP1544"
#> [35] "https://www.wikipathways.org/index.php/Pathway:WP4482"
#> [36] "https://www.wikipathways.org/index.php/Pathway:WP3297"
#> [37] "https://www.wikipathways.org/index.php/Pathway:WP530" 
#> [38] "https://www.wikipathways.org/index.php/Pathway:WP4891"
#> [39] "https://www.wikipathways.org/index.php/Pathway:WP236" 
#> [40] "https://www.wikipathways.org/index.php/Pathway:WP4747"
#> [41] "https://www.wikipathways.org/index.php/Pathway:WP176" 
#> [42] "https://www.wikipathways.org/index.php/Pathway:WP727" 
#> [43] "https://www.wikipathways.org/index.php/Pathway:WP4496"
#> [44] "https://www.wikipathways.org/index.php/Pathway:WP254" 
#> [45] "https://www.wikipathways.org/index.php/Pathway:WP1533"
#> [46] "https://www.wikipathways.org/index.php/Pathway:WP75"  
#> [47] "https://www.wikipathways.org/index.php/Pathway:WP4478"
#> [48] "https://www.wikipathways.org/index.php/Pathway:WP2328"
#> [49] "https://www.wikipathways.org/index.php/Pathway:WP3851"
#> [50] "https://www.wikipathways.org/index.php/Pathway:WP231" 
#> [51] "https://www.wikipathways.org/index.php/Pathway:WP3617"
#> [52] "https://www.wikipathways.org/index.php/Pathway:WP2036"
#> [53] "https://www.wikipathways.org/index.php/Pathway:WP3611"
#> [54] "https://www.wikipathways.org/index.php/Pathway:WP560" 
#> [55] "https://www.wikipathways.org/index.php/Pathway:WP4816"
#> [56] "https://www.wikipathways.org/index.php/Pathway:WP2864"
#> [57] "https://www.wikipathways.org/index.php/Pathway:WP4304"
#> [58] "https://www.wikipathways.org/index.php/Pathway:WP4857"
#> [59] "https://www.wikipathways.org/index.php/Pathway:WP15"  
#> [60] "https://www.wikipathways.org/index.php/Pathway:WP2586"
#> [61] "https://www.wikipathways.org/index.php/Pathway:WP4493"
#> [62] "https://www.wikipathways.org/index.php/Pathway:WP4396"
#> [63] "https://www.wikipathways.org/index.php/Pathway:WP4480"
#> [64] "https://www.wikipathways.org/index.php/Pathway:WP706" 
#> [65] "https://www.wikipathways.org/index.php/Pathway:WP2447"
#> [66] "https://www.wikipathways.org/index.php/Pathway:WP382" 
#> [67] "https://www.wikipathways.org/index.php/Pathway:WP5044"
#> [68] "https://www.wikipathways.org/index.php/Pathway:WP2751"
#> [69] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#> [70] "https://www.wikipathways.org/index.php/Pathway:WP3398"
#> [71] "https://www.wikipathways.org/index.php/Pathway:WP4066"
#> [72] "https://www.wikipathways.org/index.php/Pathway:WP4063"
#> [73] "https://www.wikipathways.org/index.php/Pathway:WP4754"
#> [74] "https://www.wikipathways.org/index.php/Pathway:WP5039"