Find Pathways By Literature

Retrieve pathways containing the query citation.

findPathwaysByLiterature(query)

Arguments

query
The character string to search for, e.g., a PMID, title keyword or author name.

Value

A dataframe of pathway attributes in addition to query result score and literature details

Details

The score is from a lucene index search engine, ranging from 0 to 1 with higher scores for better matches. The two literature columns are lists of pubmed ids and titles for the citations matching the query per pathway. The graphId column lists the id for any objects in the GPML pathway model that have been spcifically annotated with the matching citations.

Examples

{
findPathwaysByLiterature('19649250')
findPathwaysByLiterature('smith')
findPathwaysByLiterature('cancer')
}

#>          score     id
#>   1: 0.1266941 WP2733
#>   2: 0.1266941 WP3327
#>   3: 0.1266941 WP4448
#>   4: 0.1478098 WP1978
#>   5: 0.1478098 WP2688
#>  ---                 
#> 407: 0.3583451  WP475
#> 408: 0.3583451  WP961
#> 409: 0.4441429 WP2735
#> 410: 0.9936264 WP4055
#> 411: 1.0363334 WP3335
#>                                                                     name
#>   1: Regulation of mRNA stability by proteins that bind AU-rich elements
#>   2:              SUMOylation of DNA damage response and repair proteins
#>   3:                                                        PI3K Cascade
#>   4:                                                   Opioid Signalling
#>   5:                                               DAG and IP3 signaling
#>  ---                                                                    
#> 407:                                                  Irinotecan Pathway
#> 408:                                                  Irinotecan Pathway
#> 409:                                              RAF/MAP kinase cascade
#> 410:                   E3 ubiquitin ligases ubiquitinate target proteins
#> 411:                                                  Signaling by FGFR1
#>                                                        url         species
#>   1: https://www.wikipathways.org/index.php/Pathway:WP2733    Homo sapiens
#>   2: https://www.wikipathways.org/index.php/Pathway:WP3327    Homo sapiens
#>   3: https://www.wikipathways.org/index.php/Pathway:WP4448    Homo sapiens
#>   4: https://www.wikipathways.org/index.php/Pathway:WP1978    Homo sapiens
#>   5: https://www.wikipathways.org/index.php/Pathway:WP2688    Homo sapiens
#>  ---                                                                      
#> 407:  https://www.wikipathways.org/index.php/Pathway:WP475    Mus musculus
#> 408:  https://www.wikipathways.org/index.php/Pathway:WP961 Pan troglodytes
#> 409: https://www.wikipathways.org/index.php/Pathway:WP2735    Homo sapiens
#> 410: https://www.wikipathways.org/index.php/Pathway:WP4055    Homo sapiens
#> 411: https://www.wikipathways.org/index.php/Pathway:WP3335    Homo sapiens
#>      revision                                 graphId
#>   1:   114931                 fb7a3,fa496,f68af,e8096
#>   2:   114993                             c5cf9,d1025
#>   3:   114871                                   a8379
#>   4:   115086                                   d9972
#>   5:   114839                                   a9a82
#>  ---                                                 
#> 407:    93210                                        
#> 408:    93193                                        
#> 409:   114761 ea75c,e31fd,a7708,f6d76,d85c9,c1e57,...
#> 410:   115039                                   edc97
#> 411:   114733 e9df9,dccd3,dc4e2,d8045,ac740,abcf8,...
#>                                              literature.pubmed
#>   1:                                                  15514971
#>   2:                                                  19287951
#>   3:                                                   9679957
#>   4:                                                  19244110
#>   5:                                                  19244110
#>  ---                                                          
#> 407:                                                  12893990
#> 408:                                                  12893990
#> 409: 18060073,28939558,27999210,16778220,28481359,17496923,...
#> 410:                                                  27798111
#> 411: 21160078,24850843,15116089,17496923,17699850,11389083,...
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      literature.title
#>   1:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    IL-1beta induces stabilization of IL-8 mRNA in malignant breast cancer cells via the 3' untranslated region: Involvement of divergent RNA-binding factors HuR, KSRP and TIAR.
#>   2:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         A novel mechanism whereby BRCA1/1a/1b fine tunes the dynamic complex interplay between SUMO-dependent/independent activities of Ubc9 on E2-induced ERalpha activation/repression and degradation in breast cancer cells.
#>   3:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               AKT2, a member of the protein kinase B family, is activated by growth factors, v-Ha-ras, and v-src through phosphatidylinositol 3-kinase in human ovarian epithelial cancer cells.
#>   4:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        RGS17, an overexpressed gene in human lung and prostate cancer, induces tumor cell proliferation through the cyclic AMP-PKA-CREB pathway.
#>   5:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        RGS17, an overexpressed gene in human lung and prostate cancer, induces tumor cell proliferation through the cyclic AMP-PKA-CREB pathway.
#>  ---                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 
#> 407:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      UGT pharmacogenomics: implications for cancer risk and cancer therapeutics.
#> 408:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      UGT pharmacogenomics: implications for cancer risk and cancer therapeutics.
#> 409:               Mutation analysis of BRAF, MEK1 and MEK2 in 15 ovarian cancer cell lines: implications for therapy.,Targeting the MAPK Signaling Pathway in Cancer: Promising Preclinical Activity with the Novel Selective ERK1/2 Inhibitor BVD-523 (Ulixertinib).,LY3009120, a panRAF inhibitor, has significant anti-tumor activity in BRAF and KRAS mutant preclinical models of colorectal cancer.,The ERBB4/HER4 intracellular domain 4ICD is a BH3-only protein promoting apoptosis of breast cancer cells.,Mutational landscape of metastatic cancer revealed from prospective clinical sequencing of 10,000 patients.,Targeting the Raf-MEK-ERK mitogen-activated protein kinase cascade for the treatment of cancer.,...
#> 410:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     The RING finger domain E3 ubiquitin ligases BRCA1 and the RNF20/RNF40 complex in global loss of the chromatin mark histone H2B monoubiquitination (H2Bub1) in cell line models and primary high-grade serous ovarian cancer.
#> 411: Frequent and focal FGFR1 amplification associates with therapeutically tractable FGFR1 dependency in squamous cell lung cancer.,FGFR1/3 tyrosine kinase fusions define a unique molecular subtype of non-small cell lung cancer.,Blocking of FGFR signaling inhibits breast cancer cell proliferation through downregulation of D-type cyclins.,Targeting the Raf-MEK-ERK mitogen-activated protein kinase cascade for the treatment of cancer.,High-resolution single nucleotide polymorphism array analysis of epithelial ovarian cancer reveals numerous microdeletions and amplifications.,High-throughput tissue microarray analysis of 3p25 (RAF1) and 8p12 (FGFR1) copy number alterations in urinary bladder cancer.,...