Find Pathways By Text

Retrieve a list of pathways containing the query text.

findPathwaysByText(query)

Arguments

query
A character string to search for, e.g., "cancer"

Value

A list of lists

Examples

{
findPathwaysByText('cancer')
}

#> [[1]]
#> [[1]]$score
#> [[1]]$score$`0`
#> [1] "1.4718786"
#> 
#> 
#> [[1]]$fields
#> list()
#> 
#> [[1]]$id
#> [1] "WP1984"
#> 
#> [[1]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP1984"
#> 
#> [[1]]$name
#> [1] "Integrated Breast Cancer Pathway"
#> 
#> [[1]]$species
#> [1] "Homo sapiens"
#> 
#> [[1]]$revision
#> [1] "108807"
#> 
#> 
#> [[2]]
#> [[2]]$score
#> [[2]]$score$`0`
#> [1] "1.4718785"
#> 
#> 
#> [[2]]$fields
#> list()
#> 
#> [[2]]$id
#> [1] "WP3260"
#> 
#> [[2]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3260"
#> 
#> [[2]]$name
#> [1] "Integrated Breast Cancer Pathway"
#> 
#> [[2]]$species
#> [1] "Bos taurus"
#> 
#> [[2]]$revision
#> [1] "108100"
#> 
#> 
#> [[3]]
#> [[3]]$score
#> [[3]]$score$`0`
#> [1] "1.2271224"
#> 
#> 
#> [[3]]$fields
#> list()
#> 
#> [[3]]$id
#> [1] "WP4262"
#> 
#> [[3]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4262"
#> 
#> [[3]]$name
#> [1] "Breast cancer pathway"
#> 
#> [[3]]$species
#> [1] "Homo sapiens"
#> 
#> [[3]]$revision
#> [1] "108103"
#> 
#> 
#> [[4]]
#> [[4]]$score
#> [[4]]$score$`0`
#> [1] "1.2051921"
#> 
#> 
#> [[4]]$fields
#> list()
#> 
#> [[4]]$id
#> [1] "WP2868"
#> 
#> [[4]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP2868"
#> 
#> [[4]]$name
#> [1] "TCA Cycle Nutrient Utilization and Invasiveness of Ovarian Cancer"
#> 
#> [[4]]$species
#> [1] "Homo sapiens"
#> 
#> [[4]]$revision
#> [1] "108102"
#> 
#> 
#> [[5]]
#> [[5]]$score
#> [[5]]$score$`0`
#> [1] "1.2051921"
#> 
#> 
#> [[5]]$fields
#> list()
#> 
#> [[5]]$id
#> [1] "WP3252"
#> 
#> [[5]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3252"
#> 
#> [[5]]$name
#> [1] "TCA Cycle Nutrient Utilization and Invasiveness of Ovarian Cancer"
#> 
#> [[5]]$species
#> [1] "Bos taurus"
#> 
#> [[5]]$revision
#> [1] "108105"
#> 
#> 
#> [[6]]
#> [[6]]$score
#> [[6]]$score$`0`
#> [1] "1.1844578"
#> 
#> 
#> [[6]]$fields
#> list()
#> 
#> [[6]]$id
#> [1] "WP3301"
#> 
#> [[6]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3301"
#> 
#> [[6]]$name
#> [1] "MFAP5-mediated ovarian cancer cell motility and invasiveness"
#> 
#> [[6]]$species
#> [1] "Homo sapiens"
#> 
#> [[6]]$revision
#> [1] "108104"
#> 
#> 
#> [[7]]
#> [[7]]$score
#> [[7]]$score$`0`
#> [1] "0.97223824"
#> 
#> 
#> [[7]]$fields
#> list()
#> 
#> [[7]]$id
#> [1] "WP3244"
#> 
#> [[7]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3244"
#> 
#> [[7]]$name
#> [1] "Integrated Cancer pathway"
#> 
#> [[7]]$species
#> [1] "Bos taurus"
#> 
#> [[7]]$revision
#> [1] "80873"
#> 
#> 
#> [[8]]
#> [[8]]$score
#> [[8]]$score$`0`
#> [1] "0.97223824"
#> 
#> 
#> [[8]]$fields
#> list()
#> 
#> [[8]]$id
#> [1] "WP1971"
#> 
#> [[8]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP1971"
#> 
#> [[8]]$name
#> [1] "Integrated Cancer Pathway"
#> 
#> [[8]]$species
#> [1] "Homo sapiens"
#> 
#> [[8]]$revision
#> [1] "106620"
#> 
#> 
#> [[9]]
#> [[9]]$score
#> [[9]]$score$`0`
#> [1] "0.9488409"
#> 
#> 
#> [[9]]$fields
#> list()
#> 
#> [[9]]$id
#> [1] "WP3129"
#> 
#> [[9]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3129"
#> 
#> [[9]]$name
#> [1] "Bladder Cancer"
#> 
#> [[9]]$species
#> [1] "Bos taurus"
#> 
#> [[9]]$revision
#> [1] "89145"
#> 
#> 
#> [[10]]
#> [[10]]$score
#> [[10]]$score$`0`
#> [1] "0.928837"
#> 
#> 
#> [[10]]$fields
#> list()
#> 
#> [[10]]$id
#> [1] "WP2512"
#> 
#> [[10]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP2512"
#> 
#> [[10]]$name
#> [1] "Integrated Lung Cancer Pathway"
#> 
#> [[10]]$species
#> [1] "Homo sapiens"
#> 
#> [[10]]$revision
#> [1] "108111"
#> 
#> 
#> [[11]]
#> [[11]]$score
#> [[11]]$score$`0`
#> [1] "0.928837"
#> 
#> 
#> [[11]]$fields
#> list()
#> 
#> [[11]]$id
#> [1] "WP1968"
#> 
#> [[11]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP1968"
#> 
#> [[11]]$name
#> [1] "Genetic alterations of lung cancer"
#> 
#> [[11]]$species
#> [1] "Rattus norvegicus"
#> 
#> [[11]]$revision
#> [1] "108361"
#> 
#> 
#> [[12]]
#> [[12]]$score
#> [[12]]$score$`0`
#> [1] "0.90562403"
#> 
#> 
#> [[12]]$fields
#> list()
#> 
#> [[12]]$id
#> [1] "WP3206"
#> 
#> [[12]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3206"
#> 
#> [[12]]$name
#> [1] "Retinoblastoma Gene in Cancer"
#> 
#> [[12]]$species
#> [1] "Bos taurus"
#> 
#> [[12]]$revision
#> [1] "108143"
#> 
#> 
#> [[13]]
#> [[13]]$score
#> [[13]]$score$`0`
#> [1] "0.90562403"
#> 
#> 
#> [[13]]$fields
#> list()
#> 
#> [[13]]$id
#> [1] "WP2446"
#> 
#> [[13]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP2446"
#> 
#> [[13]]$name
#> [1] "Retinoblastoma Gene in Cancer"
#> 
#> [[13]]$species
#> [1] "Homo sapiens"
#> 
#> [[13]]$revision
#> [1] "108145"
#> 
#> 
#> [[14]]
#> [[14]]$score
#> [[14]]$score$`0`
#> [1] "0.87549734"
#> 
#> 
#> [[14]]$fields
#> list()
#> 
#> [[14]]$id
#> [1] "WP4155"
#> 
#> [[14]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4155"
#> 
#> [[14]]$name
#> [1] "Endometrial cancer"
#> 
#> [[14]]$species
#> [1] "Homo sapiens"
#> 
#> [[14]]$revision
#> [1] "108106"
#> 
#> 
#> [[15]]
#> [[15]]$score
#> [[15]]$score$`0`
#> [1] "0.87392145"
#> 
#> 
#> [[15]]$fields
#> list()
#> 
#> [[15]]$id
#> [1] "WP2263"
#> 
#> [[15]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP2263"
#> 
#> [[15]]$name
#> [1] "Androgen Receptor Network in Prostate Cancer"
#> 
#> [[15]]$species
#> [1] "Homo sapiens"
#> 
#> [[15]]$revision
#> [1] "108261"
#> 
#> 
#> [[16]]
#> [[16]]$score
#> [[16]]$score$`0`
#> [1] "0.8624071"
#> 
#> 
#> [[16]]$fields
#> list()
#> 
#> [[16]]$id
#> [1] "WP3981"
#> 
#> [[16]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3981"
#> 
#> [[16]]$name
#> [1] "miRNA regulation of prostate cancer signaling pathways"
#> 
#> [[16]]$species
#> [1] "Homo sapiens"
#> 
#> [[16]]$revision
#> [1] "108114"
#> 
#> 
#> [[17]]
#> [[17]]$score
#> [[17]]$score$`0`
#> [1] "0.8624071"
#> 
#> 
#> [[17]]$fields
#> list()
#> 
#> [[17]]$id
#> [1] "WP3982"
#> 
#> [[17]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3982"
#> 
#> [[17]]$name
#> [1] "miRNA regulation of p53 pathway in prostate cancer"
#> 
#> [[17]]$species
#> [1] "Homo sapiens"
#> 
#> [[17]]$revision
#> [1] "108115"
#> 
#> 
#> [[18]]
#> [[18]]$score
#> [[18]]$score$`0`
#> [1] "0.8472516"
#> 
#> 
#> [[18]]$fields
#> list()
#> 
#> [[18]]$id
#> [1] "WP2363"
#> 
#> [[18]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP2363"
#> 
#> [[18]]$name
#> [1] "Gastric Cancer Network 2"
#> 
#> [[18]]$species
#> [1] "Homo sapiens"
#> 
#> [[18]]$revision
#> [1] "108061"
#> 
#> 
#> [[19]]
#> [[19]]$score
#> [[19]]$score$`0`
#> [1] "0.8390098"
#> 
#> 
#> [[19]]$fields
#> list()
#> 
#> [[19]]$id
#> [1] "WP4400"
#> 
#> [[19]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4400"
#> 
#> [[19]]$name
#> [1] "FABP4 in ovarian cancer"
#> 
#> [[19]]$species
#> [1] "Homo sapiens"
#> 
#> [[19]]$revision
#> [1] "108112"
#> 
#> 
#> [[20]]
#> [[20]]$score
#> [[20]]$score$`0`
#> [1] "0.8390098"
#> 
#> 
#> [[20]]$fields
#> list()
#> 
#> [[20]]$id
#> [1] "WP615"
#> 
#> [[20]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP615"
#> 
#> [[20]]$name
#> [1] "Senescence and Autophagy in Cancer"
#> 
#> [[20]]$species
#> [1] "Homo sapiens"
#> 
#> [[20]]$revision
#> [1] "108116"
#> 
#> 
#> [[21]]
#> [[21]]$score
#> [[21]]$score$`0`
#> [1] "0.8347046"
#> 
#> 
#> [[21]]$fields
#> list()
#> 
#> [[21]]$id
#> [1] "WP2828"
#> 
#> [[21]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP2828"
#> 
#> [[21]]$name
#> [1] "Bladder Cancer"
#> 
#> [[21]]$species
#> [1] "Homo sapiens"
#> 
#> [[21]]$revision
#> [1] "108113"
#> 
#> 
#> [[22]]
#> [[22]]$score
#> [[22]]$score$`0`
#> [1] "0.8161514"
#> 
#> 
#> [[22]]$fields
#> list()
#> 
#> [[22]]$id
#> [1] "WP3120"
#> 
#> [[22]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3120"
#> 
#> [[22]]$name
#> [1] "Folate-Alcohol and Cancer Pathway"
#> 
#> [[22]]$species
#> [1] "Bos taurus"
#> 
#> [[22]]$revision
#> [1] "105966"
#> 
#> 
#> [[23]]
#> [[23]]$score
#> [[23]]$score$`0`
#> [1] "0.7840942"
#> 
#> 
#> [[23]]$fields
#> list()
#> 
#> [[23]]$id
#> [1] "WP1589"
#> 
#> [[23]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP1589"
#> 
#> [[23]]$name
#> [1] "Folate-Alcohol and Cancer Pathway Hypotheses"
#> 
#> [[23]]$species
#> [1] "Homo sapiens"
#> 
#> [[23]]$revision
#> [1] "106697"
#> 
#> 
#> [[24]]
#> [[24]]$score
#> [[24]]$score$`0`
#> [1] "0.77239555"
#> 
#> 
#> [[24]]$fields
#> list()
#> 
#> [[24]]$id
#> [1] "WP3139"
#> 
#> [[24]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3139"
#> 
#> [[24]]$name
#> [1] "Gastric cancer network 2"
#> 
#> [[24]]$species
#> [1] "Bos taurus"
#> 
#> [[24]]$revision
#> [1] "87525"
#> 
#> 
#> [[25]]
#> [[25]]$score
#> [[25]]$score$`0`
#> [1] "0.77239555"
#> 
#> 
#> [[25]]$fields
#> list()
#> 
#> [[25]]$id
#> [1] "WP2361"
#> 
#> [[25]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP2361"
#> 
#> [[25]]$name
#> [1] "Gastric Cancer Network 1"
#> 
#> [[25]]$species
#> [1] "Homo sapiens"
#> 
#> [[25]]$revision
#> [1] "108121"
#> 
#> 
#> [[26]]
#> [[26]]$score
#> [[26]]$score$`0`
#> [1] "0.77239555"
#> 
#> 
#> [[26]]$fields
#> list()
#> 
#> [[26]]$id
#> [1] "WP3166"
#> 
#> [[26]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3166"
#> 
#> [[26]]$name
#> [1] "Gastric Cancer Network 1"
#> 
#> [[26]]$species
#> [1] "Bos taurus"
#> 
#> [[26]]$revision
#> [1] "108120"
#> 
#> 
#> [[27]]
#> [[27]]$score
#> [[27]]$score$`0`
#> [1] "0.77239555"
#> 
#> 
#> [[27]]$fields
#> list()
#> 
#> [[27]]$id
#> [1] "WP4585"
#> 
#> [[27]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4585"
#> 
#> [[27]]$name
#> [1] "Cancer immunotherapy by PD-1 blockade"
#> 
#> [[27]]$species
#> [1] "Homo sapiens"
#> 
#> [[27]]$revision
#> [1] "108119"
#> 
#> 
#> [[28]]
#> [[28]]$score
#> [[28]]$score$`0`
#> [1] "0.77239555"
#> 
#> 
#> [[28]]$fields
#> list()
#> 
#> [[28]]$id
#> [1] "WP4290"
#> 
#> [[28]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4290"
#> 
#> [[28]]$name
#> [1] "Metabolic reprogramming in colon cancer"
#> 
#> [[28]]$species
#> [1] "Homo sapiens"
#> 
#> [[28]]$revision
#> [1] "108376"
#> 
#> 
#> [[29]]
#> [[29]]$score
#> [[29]]$score$`0`
#> [1] "0.7656662"
#> 
#> 
#> [[29]]$fields
#> list()
#> 
#> [[29]]$id
#> [1] "WP4658"
#> 
#> [[29]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4658"
#> 
#> [[29]]$name
#> [1] "Small cell lung cancer"
#> 
#> [[29]]$species
#> [1] "Homo sapiens"
#> 
#> [[29]]$revision
#> [1] "108148"
#> 
#> 
#> [[30]]
#> [[30]]$score
#> [[30]]$score$`0`
#> [1] "0.74087733"
#> 
#> 
#> [[30]]$fields
#> list()
#> 
#> [[30]]$id
#> [1] "WP4284"
#> 
#> [[30]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4284"
#> 
#> [[30]]$name
#> [1] "Ultraconserved region 339 modulation of tumor suppressor microRNAs in cancer"
#> 
#> [[30]]$species
#> [1] "Homo sapiens"
#> 
#> [[30]]$revision
#> [1] "108123"
#> 
#> 
#> [[31]]
#> [[31]]$score
#> [[31]]$score$`0`
#> [1] "0.73742044"
#> 
#> 
#> [[31]]$fields
#> list()
#> 
#> [[31]]$id
#> [1] "WP3171"
#> 
#> [[31]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3171"
#> 
#> [[31]]$name
#> [1] "Deregulation of Rab and Rab Effector Genes in Bladder Cancer"
#> 
#> [[31]]$species
#> [1] "Bos taurus"
#> 
#> [[31]]$revision
#> [1] "108117"
#> 
#> 
#> [[32]]
#> [[32]]$score
#> [[32]]$score$`0`
#> [1] "0.73742044"
#> 
#> 
#> [[32]]$fields
#> list()
#> 
#> [[32]]$id
#> [1] "WP2291"
#> 
#> [[32]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP2291"
#> 
#> [[32]]$name
#> [1] "Deregulation of Rab and Rab Effector Genes in Bladder Cancer"
#> 
#> [[32]]$species
#> [1] "Homo sapiens"
#> 
#> [[32]]$revision
#> [1] "108118"
#> 
#> 
#> [[33]]
#> [[33]]$score
#> [[33]]$score$`0`
#> [1] "0.72917867"
#> 
#> 
#> [[33]]$fields
#> list()
#> 
#> [[33]]$id
#> [1] "WP3593"
#> 
#> [[33]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3593"
#> 
#> [[33]]$name
#> [1] "MicroRNA for Targeting Cancer Growth and Vascularization in Glioblastoma"
#> 
#> [[33]]$species
#> [1] "Homo sapiens"
#> 
#> [[33]]$revision
#> [1] "108122"
#> 
#> 
#> [[34]]
#> [[34]]$score
#> [[34]]$score$`0`
#> [1] "0.71748"
#> 
#> 
#> [[34]]$fields
#> list()
#> 
#> [[34]]$id
#> [1] "WP4239"
#> 
#> [[34]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4239"
#> 
#> [[34]]$name
#> [1] "Epithelial to mesenchymal transition in colorectal cancer"
#> 
#> [[34]]$species
#> [1] "Homo sapiens"
#> 
#> [[34]]$revision
#> [1] "108129"
#> 
#> 
#> [[35]]
#> [[35]]$score
#> [[35]]$score$`0`
#> [1] "0.71748"
#> 
#> 
#> [[35]]$fields
#> list()
#> 
#> [[35]]$id
#> [1] "WP3979"
#> 
#> [[35]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3979"
#> 
#> [[35]]$name
#> [1] "mir-193a and MVP in colon cancer metastasis"
#> 
#> [[35]]$species
#> [1] "Mus musculus"
#> 
#> [[35]]$revision
#> [1] "108126"
#> 
#> 
#> [[36]]
#> [[36]]$score
#> [[36]]$score$`0`
#> [1] "0.71075064"
#> 
#> 
#> [[36]]$fields
#> list()
#> 
#> [[36]]$id
#> [1] "WP4255"
#> 
#> [[36]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4255"
#> 
#> [[36]]$name
#> [1] "Non-small cell lung cancer"
#> 
#> [[36]]$species
#> [1] "Homo sapiens"
#> 
#> [[36]]$revision
#> [1] "108128"
#> 
#> 
#> [[37]]
#> [[37]]$score
#> [[37]]$score$`0`
#> [1] "0.70578134"
#> 
#> 
#> [[37]]$fields
#> list()
#> 
#> [[37]]$id
#> [1] "WP4582"
#> 
#> [[37]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4582"
#> 
#> [[37]]$name
#> [1] "Cancer immunotherapy by CTLA4 blockade"
#> 
#> [[37]]$species
#> [1] "Homo sapiens"
#> 
#> [[37]]$revision
#> [1] "108125"
#> 
#> 
#> [[38]]
#> [[38]]$score
#> [[38]]$score$`0`
#> [1] "0.6699579"
#> 
#> 
#> [[38]]$fields
#> list()
#> 
#> [[38]]$id
#> [1] "WP4216"
#> 
#> [[38]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4216"
#> 
#> [[38]]$name
#> [1] "Chromosomal and microsatellite instability in colorectal cancer "
#> 
#> [[38]]$species
#> [1] "Homo sapiens"
#> 
#> [[38]]$revision
#> [1] "108124"
#> 
#> 
#> [[39]]
#> [[39]]$score
#> [[39]]$score$`0`
#> [1] "0.6592169"
#> 
#> 
#> [[39]]$fields
#> list()
#> 
#> [[39]]$id
#> [1] "WP4627"
#> 
#> [[39]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4627"
#> 
#> [[39]]$name
#> [1] " Lipids measured in liver metastasis from breast cancer"
#> 
#> [[39]]$species
#> [1] "Mus musculus"
#> 
#> [[39]]$revision
#> [1] "108158"
#> 
#> 
#> [[40]]
#> [[40]]$score
#> [[40]]$score$`0`
#> [1] "0.65140474"
#> 
#> 
#> [[40]]$fields
#> list()
#> 
#> [[40]]$id
#> [1] "WP4485"
#> 
#> [[40]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4485"
#> 
#> [[40]]$name
#> [1] "Modified nucleosides derived from t-RNA as urinary cancer markers"
#> 
#> [[40]]$species
#> [1] "Homo sapiens"
#> 
#> [[40]]$revision
#> [1] "108161"
#> 
#> 
#> [[41]]
#> [[41]]$score
#> [[41]]$score$`0`
#> [1] "0.6508658"
#> 
#> 
#> [[41]]$fields
#> list()
#> 
#> [[41]]$id
#> [1] "WP4565"
#> 
#> [[41]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4565"
#> 
#> [[41]]$name
#> [1] "Neural Crest Cell Migration in Cancer"
#> 
#> [[41]]$species
#> [1] "Homo sapiens"
#> 
#> [[41]]$revision
#> [1] "108154"
#> 
#> 
#> [[42]]
#> [[42]]$score
#> [[42]]$score$`0`
#> [1] "0.6150423"
#> 
#> 
#> [[42]]$fields
#> list()
#> 
#> [[42]]$id
#> [1] "WP4258"
#> 
#> [[42]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4258"
#> 
#> [[42]]$name
#> [1] "LncRNA involvement in canonical Wnt signaling and colorectal cancer"
#> 
#> [[42]]$species
#> [1] "Homo sapiens"
#> 
#> [[42]]$revision
#> [1] "108136"
#> 
#> 
#> [[43]]
#> [[43]]$score
#> [[43]]$score$`0`
#> [1] "0.6093468"
#> 
#> 
#> [[43]]$fields
#> list()
#> 
#> [[43]]$id
#> [1] "WP4560"
#> 
#> [[43]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4560"
#> 
#> [[43]]$name
#> [1] "MFAP5 effect on permeability and motility of endothelial cells via cytoskeleton rearrangement"
#> 
#> [[43]]$species
#> [1] "Homo sapiens"
#> 
#> [[43]]$revision
#> [1] "108134"
#> 
#> 
#> [[44]]
#> [[44]]$score
#> [[44]]$score$`0`
#> [1] "0.60764885"
#> 
#> 
#> [[44]]$fields
#> list()
#> 
#> [[44]]$id
#> [1] "WP3126"
#> 
#> [[44]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3126"
#> 
#> [[44]]$name
#> [1] "Apoptosis-related network due to altered Notch3 in ovarian cancer"
#> 
#> [[44]]$species
#> [1] "Bos taurus"
#> 
#> [[44]]$revision
#> [1] "108132"
#> 
#> 
#> [[45]]
#> [[45]]$score
#> [[45]]$score$`0`
#> [1] "0.60764885"
#> 
#> 
#> [[45]]$fields
#> list()
#> 
#> [[45]]$id
#> [1] "WP2864"
#> 
#> [[45]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP2864"
#> 
#> [[45]]$name
#> [1] "Apoptosis-related network due to altered Notch3 in ovarian cancer"
#> 
#> [[45]]$species
#> [1] "Homo sapiens"
#> 
#> [[45]]$revision
#> [1] "108133"
#> 
#> 
#> [[46]]
#> [[46]]$score
#> [[46]]$score$`0`
#> [1] "0.59810776"
#> 
#> 
#> [[46]]$fields
#> list()
#> 
#> [[46]]$id
#> [1] "WP3879"
#> 
#> [[46]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3879"
#> 
#> [[46]]$name
#> [1] "4-hydroxytamoxifen, Dexamethasone, and Retinoic Acids Regulation of p27 Expression"
#> 
#> [[46]]$species
#> [1] "Homo sapiens"
#> 
#> [[46]]$revision
#> [1] "108131"
#> 
#> 
#> [[47]]
#> [[47]]$score
#> [[47]]$score$`0`
#> [1] "0.48954025"
#> 
#> 
#> [[47]]$fields
#> list()
#> 
#> [[47]]$id
#> [1] "WP1601"
#> 
#> [[47]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP1601"
#> 
#> [[47]]$name
#> [1] "Fluoropyrimidine Activity"
#> 
#> [[47]]$species
#> [1] "Homo sapiens"
#> 
#> [[47]]$revision
#> [1] "108138"
#> 
#> 
#> [[48]]
#> [[48]]$score
#> [[48]]$score$`0`
#> [1] "0.48954025"
#> 
#> 
#> [[48]]$fields
#> list()
#> 
#> [[48]]$id
#> [1] "WP3275"
#> 
#> [[48]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3275"
#> 
#> [[48]]$name
#> [1] "Fluoropyrimidine Activity"
#> 
#> [[48]]$species
#> [1] "Bos taurus"
#> 
#> [[48]]$revision
#> [1] "108137"
#> 
#> 
#> [[49]]
#> [[49]]$score
#> [[49]]$score$`0`
#> [1] "0.48611912"
#> 
#> 
#> [[49]]$fields
#> list()
#> 
#> [[49]]$id
#> [1] "WP4397"
#> 
#> [[49]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4397"
#> 
#> [[49]]$name
#> [1] "Model for regulation of MSMP expression in cancer cells and its proangiogenic role in ovarian tumors"
#> 
#> [[49]]$species
#> [1] "Homo sapiens"
#> 
#> [[49]]$revision
#> [1] "108140"
#> 
#> 
#> [[50]]
#> [[50]]$score
#> [[50]]$score$`0`
#> [1] "0.4393245"
#> 
#> 
#> [[50]]$fields
#> list()
#> 
#> [[50]]$id
#> [1] "WP4794"
#> 
#> [[50]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4794"
#> 
#> [[50]]$name
#> [1] "cancer"
#> 
#> [[50]]$species
#> [1] "Homo sapiens"
#> 
#> [[50]]$revision
#> [1] "108946"
#> 
#> 
#> [[51]]
#> [[51]]$score
#> [[51]]$score$`0`
#> [1] "0.35532343"
#> 
#> 
#> [[51]]$fields
#> list()
#> 
#> [[51]]$id
#> [1] "WP4337"
#> 
#> [[51]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4337"
#> 
#> [[51]]$name
#> [1] "ncRNAs involved in STAT3 signaling in hepatocellular carcinoma"
#> 
#> [[51]]$species
#> [1] "Homo sapiens"
#> 
#> [[51]]$revision
#> [1] "108190"
#> 
#> 
#> [[52]]
#> [[52]]$score
#> [[52]]$score$`0`
#> [1] "0.3381717"
#> 
#> 
#> [[52]]$fields
#> list()
#> 
#> [[52]]$id
#> [1] "WP1044"
#> 
#> [[52]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP1044"
#> 
#> [[52]]$name
#> [1] "ErbB Signaling Pathway"
#> 
#> [[52]]$species
#> [1] "Bos taurus"
#> 
#> [[52]]$revision
#> [1] "108379"
#> 
#> 
#> [[53]]
#> [[53]]$score
#> [[53]]$score$`0`
#> [1] "0.3253718"
#> 
#> 
#> [[53]]$fields
#> list()
#> 
#> [[53]]$id
#> [1] "WP3298"
#> 
#> [[53]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3298"
#> 
#> [[53]]$name
#> [1] "Melatonin metabolism and effects"
#> 
#> [[53]]$species
#> [1] "Homo sapiens"
#> 
#> [[53]]$revision
#> [1] "107142"
#> 
#> 
#> [[54]]
#> [[54]]$score
#> [[54]]$score$`0`
#> [1] "0.3181079"
#> 
#> 
#> [[54]]$fields
#> list()
#> 
#> [[54]]$id
#> [1] "WP924"
#> 
#> [[54]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP924"
#> 
#> [[54]]$name
#> [1] "ErbB signaling pathway"
#> 
#> [[54]]$species
#> [1] "Pan troglodytes"
#> 
#> [[54]]$revision
#> [1] "108172"
#> 
#> 
#> [[55]]
#> [[55]]$score
#> [[55]]$score$`0`
#> [1] "0.3181079"
#> 
#> 
#> [[55]]$fields
#> list()
#> 
#> [[55]]$id
#> [1] "WP807"
#> 
#> [[55]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP807"
#> 
#> [[55]]$name
#> [1] "ErbB signaling pathway"
#> 
#> [[55]]$species
#> [1] "Gallus gallus"
#> 
#> [[55]]$revision
#> [1] "108178"
#> 
#> 
#> [[56]]
#> [[56]]$score
#> [[56]]$score$`0`
#> [1] "0.3181079"
#> 
#> 
#> [[56]]$fields
#> list()
#> 
#> [[56]]$id
#> [1] "WP1160"
#> 
#> [[56]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP1160"
#> 
#> [[56]]$name
#> [1] "ErbB signaling pathway"
#> 
#> [[56]]$species
#> [1] "Canis familiaris"
#> 
#> [[56]]$revision
#> [1] "108181"
#> 
#> 
#> [[57]]
#> [[57]]$score
#> [[57]]$score$`0`
#> [1] "0.3181079"
#> 
#> 
#> [[57]]$fields
#> list()
#> 
#> [[57]]$id
#> [1] "WP1366"
#> 
#> [[57]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP1366"
#> 
#> [[57]]$name
#> [1] "ErbB Signaling Pathway"
#> 
#> [[57]]$species
#> [1] "Danio rerio"
#> 
#> [[57]]$revision
#> [1] "108184"
#> 
#> 
#> [[58]]
#> [[58]]$score
#> [[58]]$score$`0`
#> [1] "0.3181079"
#> 
#> 
#> [[58]]$fields
#> list()
#> 
#> [[58]]$id
#> [1] "WP1261"
#> 
#> [[58]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP1261"
#> 
#> [[58]]$name
#> [1] "ErbB signaling pathway"
#> 
#> [[58]]$species
#> [1] "Mus musculus"
#> 
#> [[58]]$revision
#> [1] "108380"
#> 
#> 
#> [[59]]
#> [[59]]$score
#> [[59]]$score$`0`
#> [1] "0.3181079"
#> 
#> 
#> [[59]]$fields
#> list()
#> 
#> [[59]]$id
#> [1] "WP1299"
#> 
#> [[59]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP1299"
#> 
#> [[59]]$name
#> [1] "ErbB signaling pathway"
#> 
#> [[59]]$species
#> [1] "Rattus norvegicus"
#> 
#> [[59]]$revision
#> [1] "108381"
#> 
#> 
#> [[60]]
#> [[60]]$score
#> [[60]]$score$`0`
#> [1] "0.31039596"
#> 
#> 
#> [[60]]$fields
#> list()
#> 
#> [[60]]$id
#> [1] "WP3680"
#> 
#> [[60]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3680"
#> 
#> [[60]]$name
#> [1] "Association Between Physico-Chemical Features and Toxicity Associated Pathways"
#> 
#> [[60]]$species
#> [1] "Homo sapiens"
#> 
#> [[60]]$revision
#> [1] "106306"
#> 
#> 
#> [[61]]
#> [[61]]$score
#> [[61]]$score$`0`
#> [1] "0.30822006"
#> 
#> 
#> [[61]]$fields
#> list()
#> 
#> [[61]]$id
#> [1] "WP3672"
#> 
#> [[61]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3672"
#> 
#> [[61]]$name
#> [1] "LncRNA-mediated mechanisms of therapeutic resistance"
#> 
#> [[61]]$species
#> [1] "Homo sapiens"
#> 
#> [[61]]$revision
#> [1] "108191"
#> 
#> 
#> [[62]]
#> [[62]]$score
#> [[62]]$score$`0`
#> [1] "0.29324424"
#> 
#> 
#> [[62]]$fields
#> list()
#> 
#> [[62]]$id
#> [1] "WP673"
#> 
#> [[62]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP673"
#> 
#> [[62]]$name
#> [1] "ErbB Signaling Pathway"
#> 
#> [[62]]$species
#> [1] "Homo sapiens"
#> 
#> [[62]]$revision
#> [1] "108898"
#> 
#> 
#> [[63]]
#> [[63]]$score
#> [[63]]$score$`0`
#> [1] "0.27457783"
#> 
#> 
#> [[63]]$fields
#> list()
#> 
#> [[63]]$id
#> [1] "WP4827"
#> 
#> [[63]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4827"
#> 
#> [[63]]$name
#> [1] "inflammation and cancer"
#> 
#> [[63]]$species
#> [1] "Anopheles gambiae"
#> 
#> [[63]]$revision
#> [1] "109192"
#> 
#> 
#> [[64]]
#> [[64]]$score
#> [[64]]$score$`0`
#> [1] "0.26645687"
#> 
#> 
#> [[64]]$fields
#> list()
#> 
#> [[64]]$id
#> [1] "WP4747"
#> 
#> [[64]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4747"
#> 
#> [[64]]$name
#> [1] "Netrin-UNC5B signaling Pathway"
#> 
#> [[64]]$species
#> [1] "Homo sapiens"
#> 
#> [[64]]$revision
#> [1] "108877"
#> 
#> 
#> [[65]]
#> [[65]]$score
#> [[65]]$score$`0`
#> [1] "0.2632926"
#> 
#> 
#> [[65]]$fields
#> list()
#> 
#> [[65]]$id
#> [1] "WP4336"
#> 
#> [[65]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4336"
#> 
#> [[65]]$name
#> [1] "ncRNAs involved in Wnt signaling in hepatocellular carcinoma"
#> 
#> [[65]]$species
#> [1] "Homo sapiens"
#> 
#> [[65]]$revision
#> [1] "108198"
#> 
#> 
#> [[66]]
#> [[66]]$score
#> [[66]]$score$`0`
#> [1] "0.25471672"
#> 
#> 
#> [[66]]$fields
#> list()
#> 
#> [[66]]$id
#> [1] "WP4559"
#> 
#> [[66]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4559"
#> 
#> [[66]]$name
#> [1] "Interactions between immune cells and microRNAs in tumor microenvironment"
#> 
#> [[66]]$species
#> [1] "Homo sapiens"
#> 
#> [[66]]$revision
#> [1] "104316"
#> 
#> 
#> [[67]]
#> [[67]]$score
#> [[67]]$score$`0`
#> [1] "0.23551683"
#> 
#> 
#> [[67]]$fields
#> list()
#> 
#> [[67]]$id
#> [1] "WP3969"
#> 
#> [[67]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3969"
#> 
#> [[67]]$name
#> [1] "H19 action Rb-E2F1 signaling and CDK-Beta-catenin activity"
#> 
#> [[67]]$species
#> [1] "Homo sapiens"
#> 
#> [[67]]$revision
#> [1] "106437"
#> 
#> 
#> [[68]]
#> [[68]]$score
#> [[68]]$score$`0`
#> [1] "0.23314977"
#> 
#> 
#> [[68]]$fields
#> list()
#> 
#> [[68]]$id
#> [1] "WP929"
#> 
#> [[68]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP929"
#> 
#> [[68]]$name
#> [1] "TNF-alpha NF-kB Signaling Pathway"
#> 
#> [[68]]$species
#> [1] "Pan troglodytes"
#> 
#> [[68]]$revision
#> [1] "108150"
#> 
#> 
#> [[69]]
#> [[69]]$score
#> [[69]]$score$`0`
#> [1] "0.23314977"
#> 
#> 
#> [[69]]$fields
#> list()
#> 
#> [[69]]$id
#> [1] "WP1163"
#> 
#> [[69]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP1163"
#> 
#> [[69]]$name
#> [1] "TNF-alpha NF-kB Signaling Pathway"
#> 
#> [[69]]$species
#> [1] "Canis familiaris"
#> 
#> [[69]]$revision
#> [1] "108149"
#> 
#> 
#> [[70]]
#> [[70]]$score
#> [[70]]$score$`0`
#> [1] "0.23314977"
#> 
#> 
#> [[70]]$fields
#> list()
#> 
#> [[70]]$id
#> [1] "WP811"
#> 
#> [[70]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP811"
#> 
#> [[70]]$name
#> [1] "TNF-alpha NF-kB Signaling Pathway"
#> 
#> [[70]]$species
#> [1] "Gallus gallus"
#> 
#> [[70]]$revision
#> [1] "108152"
#> 
#> 
#> [[71]]
#> [[71]]$score
#> [[71]]$score$`0`
#> [1] "0.23314977"
#> 
#> 
#> [[71]]$fields
#> list()
#> 
#> [[71]]$id
#> [1] "WP1369"
#> 
#> [[71]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP1369"
#> 
#> [[71]]$name
#> [1] "TNF-alpha NF-kB Signaling Pathway"
#> 
#> [[71]]$species
#> [1] "Danio rerio"
#> 
#> [[71]]$revision
#> [1] "108316"
#> 
#> 
#> [[72]]
#> [[72]]$score
#> [[72]]$score$`0`
#> [1] "0.21966225"
#> 
#> 
#> [[72]]$fields
#> list()
#> 
#> [[72]]$id
#> [1] "WP3231"
#> 
#> [[72]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3231"
#> 
#> [[72]]$name
#> [1] "Senescence and Autophagy in Cancer"
#> 
#> [[72]]$species
#> [1] "Bos taurus"
#> 
#> [[72]]$revision
#> [1] "106001"
#> 
#> 
#> [[73]]
#> [[73]]$score
#> [[73]]$score$`0`
#> [1] "0.21966225"
#> 
#> 
#> [[73]]$fields
#> list()
#> 
#> [[73]]$id
#> [1] "WP4826"
#> 
#> [[73]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4826"
#> 
#> [[73]]$name
#> [1] "inflammation leading to cancer"
#> 
#> [[73]]$species
#> [1] "Homo sapiens"
#> 
#> [[73]]$revision
#> [1] "109191"
#> 
#> 
#> [[74]]
#> [[74]]$score
#> [[74]]$score$`0`
#> [1] "0.20396355"
#> 
#> 
#> [[74]]$fields
#> list()
#> 
#> [[74]]$id
#> [1] "WP3967"
#> 
#> [[74]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3967"
#> 
#> [[74]]$name
#> [1] "miR-509-3p alteration of YAP1/ECM axis"
#> 
#> [[74]]$species
#> [1] "Homo sapiens"
#> 
#> [[74]]$revision
#> [1] "108139"
#> 
#> 
#> [[75]]
#> [[75]]$score
#> [[75]]$score$`0`
#> [1] "0.20396355"
#> 
#> 
#> [[75]]$fields
#> list()
#> 
#> [[75]]$id
#> [1] "WP3264"
#> 
#> [[75]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3264"
#> 
#> [[75]]$name
#> [1] "Tamoxifen metabolism"
#> 
#> [[75]]$species
#> [1] "Bos taurus"
#> 
#> [[75]]$revision
#> [1] "108155"
#> 
#> 
#> [[76]]
#> [[76]]$score
#> [[76]]$score$`0`
#> [1] "0.20396355"
#> 
#> 
#> [[76]]$fields
#> list()
#> 
#> [[76]]$id
#> [1] "WP3972"
#> 
#> [[76]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3972"
#> 
#> [[76]]$name
#> [1] "PDGFR-beta pathway"
#> 
#> [[76]]$species
#> [1] "Homo sapiens"
#> 
#> [[76]]$revision
#> [1] "108157"
#> 
#> 
#> [[77]]
#> [[77]]$score
#> [[77]]$score$`0`
#> [1] "0.19984265"
#> 
#> 
#> [[77]]$fields
#> list()
#> 
#> [[77]]$id
#> [1] "WP3873"
#> 
#> [[77]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3873"
#> 
#> [[77]]$name
#> [1] "RIOK1 and RIOK2 in EGFR- and PI3K-mediated tumorigenesis"
#> 
#> [[77]]$species
#> [1] "Drosophila melanogaster"
#> 
#> [[77]]$revision
#> [1] "108159"
#> 
#> 
#> [[78]]
#> [[78]]$score
#> [[78]]$score$`0`
#> [1] "0.18841346"
#> 
#> 
#> [[78]]$fields
#> list()
#> 
#> [[78]]$id
#> [1] "WP3239"
#> 
#> [[78]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3239"
#> 
#> [[78]]$name
#> [1] "TP53 Network"
#> 
#> [[78]]$species
#> [1] "Bos taurus"
#> 
#> [[78]]$revision
#> [1] "80865"
#> 
#> 
#> [[79]]
#> [[79]]$score
#> [[79]]$score$`0`
#> [1] "0.18841346"
#> 
#> 
#> [[79]]$fields
#> list()
#> 
#> [[79]]$id
#> [1] "WP3611"
#> 
#> [[79]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3611"
#> 
#> [[79]]$name
#> [1] "Photodynamic therapy-induced AP-1 survival signaling."
#> 
#> [[79]]$species
#> [1] "Homo sapiens"
#> 
#> [[79]]$revision
#> [1] "90923"
#> 
#> 
#> [[80]]
#> [[80]]$score
#> [[80]]$score$`0`
#> [1] "0.18841346"
#> 
#> 
#> [[80]]$fields
#> list()
#> 
#> [[80]]$id
#> [1] "WP1742"
#> 
#> [[80]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP1742"
#> 
#> [[80]]$name
#> [1] "TP53 Network"
#> 
#> [[80]]$species
#> [1] "Homo sapiens"
#> 
#> [[80]]$revision
#> [1] "106445"
#> 
#> 
#> [[81]]
#> [[81]]$score
#> [[81]]$score$`0`
#> [1] "0.18841346"
#> 
#> 
#> [[81]]$fields
#> list()
#> 
#> [[81]]$id
#> [1] "WP3859"
#> 
#> [[81]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3859"
#> 
#> [[81]]$name
#> [1] "TGF-B Signaling in Thyroid Cells for Epithelial-Mesenchymal Transition"
#> 
#> [[81]]$species
#> [1] "Homo sapiens"
#> 
#> [[81]]$revision
#> [1] "106478"
#> 
#> 
#> [[82]]
#> [[82]]$score
#> [[82]]$score$`0`
#> [1] "0.18841346"
#> 
#> 
#> [[82]]$fields
#> list()
#> 
#> [[82]]$id
#> [1] "WP3872"
#> 
#> [[82]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3872"
#> 
#> [[82]]$name
#> [1] "Regulation of Apoptosis by Parathyroid Hormone-related Protein"
#> 
#> [[82]]$species
#> [1] "Homo sapiens"
#> 
#> [[82]]$revision
#> [1] "106498"
#> 
#> 
#> [[83]]
#> [[83]]$score
#> [[83]]$score$`0`
#> [1] "0.18841346"
#> 
#> 
#> [[83]]$fields
#> list()
#> 
#> [[83]]$id
#> [1] "WP3613"
#> 
#> [[83]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3613"
#> 
#> [[83]]$name
#> [1] "Photodynamic therapy-induced unfolded protein response"
#> 
#> [[83]]$species
#> [1] "Homo sapiens"
#> 
#> [[83]]$revision
#> [1] "106539"
#> 
#> 
#> [[84]]
#> [[84]]$score
#> [[84]]$score$`0`
#> [1] "0.18841346"
#> 
#> 
#> [[84]]$fields
#> list()
#> 
#> [[84]]$id
#> [1] "WP3614"
#> 
#> [[84]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3614"
#> 
#> [[84]]$name
#> [1] "Photodynamic therapy-induced HIF-1 survival signaling"
#> 
#> [[84]]$species
#> [1] "Homo sapiens"
#> 
#> [[84]]$revision
#> [1] "106542"
#> 
#> 
#> [[85]]
#> [[85]]$score
#> [[85]]$score$`0`
#> [1] "0.18841346"
#> 
#> 
#> [[85]]$fields
#> list()
#> 
#> [[85]]$id
#> [1] "WP3617"
#> 
#> [[85]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3617"
#> 
#> [[85]]$name
#> [1] "Photodynamic therapy-induced NF-kB survival signaling"
#> 
#> [[85]]$species
#> [1] "Homo sapiens"
#> 
#> [[85]]$revision
#> [1] "106541"
#> 
#> 
#> [[86]]
#> [[86]]$score
#> [[86]]$score$`0`
#> [1] "0.18841346"
#> 
#> 
#> [[86]]$fields
#> list()
#> 
#> [[86]]$id
#> [1] "WP3612"
#> 
#> [[86]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3612"
#> 
#> [[86]]$name
#> [1] "Photodynamic therapy-induced NFE2L2 (NRF2) survival signaling"
#> 
#> [[86]]$species
#> [1] "Homo sapiens"
#> 
#> [[86]]$revision
#> [1] "106540"
#> 
#> 
#> [[87]]
#> [[87]]$score
#> [[87]]$score$`0`
#> [1] "0.16653554"
#> 
#> 
#> [[87]]$fields
#> list()
#> 
#> [[87]]$id
#> [1] "WP3303"
#> 
#> [[87]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3303"
#> 
#> [[87]]$name
#> [1] "RAC1/PAK1/p38/MMP2 Pathway"
#> 
#> [[87]]$species
#> [1] "Homo sapiens"
#> 
#> [[87]]$revision
#> [1] "108160"
#> 
#> 
#> [[88]]
#> [[88]]$score
#> [[88]]$score$`0`
#> [1] "0.16653554"
#> 
#> 
#> [[88]]$fields
#> list()
#> 
#> [[88]]$id
#> [1] "WP4206"
#> 
#> [[88]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4206"
#> 
#> [[88]]$name
#> [1] "Hereditary leiomyomatosis and renal cell carcinoma pathway"
#> 
#> [[88]]$species
#> [1] "Homo sapiens"
#> 
#> [[88]]$revision
#> [1] "108163"
#> 
#> 
#> [[89]]
#> [[89]]$score
#> [[89]]$score$`0`
#> [1] "0.16653554"
#> 
#> 
#> [[89]]$fields
#> list()
#> 
#> [[89]]$id
#> [1] "WP3678"
#> 
#> [[89]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3678"
#> 
#> [[89]]$name
#> [1] "Amplification and Expansion of Oncogenic Pathways as Metastatic Traits"
#> 
#> [[89]]$species
#> [1] "Homo sapiens"
#> 
#> [[89]]$revision
#> [1] "108162"
#> 
#> 
#> [[90]]
#> [[90]]$score
#> [[90]]$score$`0`
#> [1] "0.16317084"
#> 
#> 
#> [[90]]$fields
#> list()
#> 
#> [[90]]$id
#> [1] "WP3858"
#> 
#> [[90]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3858"
#> 
#> [[90]]$name
#> [1] "Toll-like Receptor Signaling related to MyD88"
#> 
#> [[90]]$species
#> [1] "Homo sapiens"
#> 
#> [[90]]$revision
#> [1] "107438"
#> 
#> 
#> [[91]]
#> [[91]]$score
#> [[91]]$score$`0`
#> [1] "0.1413101"
#> 
#> 
#> [[91]]$fields
#> list()
#> 
#> [[91]]$id
#> [1] "WP4301"
#> 
#> [[91]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4301"
#> 
#> [[91]]$name
#> [1] "Inhibition of exosome biogenesis and secretion by Manumycin A in CRPC cells"
#> 
#> [[91]]$species
#> [1] "Homo sapiens"
#> 
#> [[91]]$revision
#> [1] "97800"
#> 
#> 
#> [[92]]
#> [[92]]$score
#> [[92]]$score$`0`
#> [1] "0.1413101"
#> 
#> 
#> [[92]]$fields
#> list()
#> 
#> [[92]]$id
#> [1] "WP3161"
#> 
#> [[92]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3161"
#> 
#> [[92]]$name
#> [1] "Extracellular vesicle-mediated signaling in recipient cells"
#> 
#> [[92]]$species
#> [1] "Bos taurus"
#> 
#> [[92]]$revision
#> [1] "88053"
#> 
#> 
#> [[93]]
#> [[93]]$score
#> [[93]]$score$`0`
#> [1] "0.1413101"
#> 
#> 
#> [[93]]$fields
#> list()
#> 
#> [[93]]$id
#> [1] "WP4534"
#> 
#> [[93]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4534"
#> 
#> [[93]]$name
#> [1] "Mechanoregulation and pathology of YAP/TAZ via Hippo and non-Hippo mechanisms"
#> 
#> [[93]]$species
#> [1] "Homo sapiens"
#> 
#> [[93]]$revision
#> [1] "103303"
#> 
#> 
#> [[94]]
#> [[94]]$score
#> [[94]]$score$`0`
#> [1] "0.1413101"
#> 
#> 
#> [[94]]$fields
#> list()
#> 
#> [[94]]$id
#> [1] "WP2870"
#> 
#> [[94]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP2870"
#> 
#> [[94]]$name
#> [1] "Extracellular vesicle-mediated signaling in recipient cells"
#> 
#> [[94]]$species
#> [1] "Homo sapiens"
#> 
#> [[94]]$revision
#> [1] "106411"
#> 
#> 
#> [[95]]
#> [[95]]$score
#> [[95]]$score$`0`
#> [1] "0.1413101"
#> 
#> 
#> [[95]]$fields
#> list()
#> 
#> [[95]]$id
#> [1] "WP699"
#> 
#> [[95]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP699"
#> 
#> [[95]]$name
#> [1] "Aflatoxin B1 metabolism"
#> 
#> [[95]]$species
#> [1] "Homo sapiens"
#> 
#> [[95]]$revision
#> [1] "106828"
#> 
#> 
#> [[96]]
#> [[96]]$score
#> [[96]]$score$`0`
#> [1] "0.13322844"
#> 
#> 
#> [[96]]$fields
#> list()
#> 
#> [[96]]$id
#> [1] "WP691"
#> 
#> [[96]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP691"
#> 
#> [[96]]$name
#> [1] "Tamoxifen metabolism"
#> 
#> [[96]]$species
#> [1] "Homo sapiens"
#> 
#> [[96]]$revision
#> [1] "106824"
#> 
#> 
#> [[97]]
#> [[97]]$score
#> [[97]]$score$`0`
#> [1] "0.13322844"
#> 
#> 
#> [[97]]$fields
#> list()
#> 
#> [[97]]$id
#> [1] "WP3670"
#> 
#> [[97]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3670"
#> 
#> [[97]]$name
#> [1] "Simplified Interaction Map Between LOXL4 and Oxidative Stress Pathway"
#> 
#> [[97]]$species
#> [1] "Homo sapiens"
#> 
#> [[97]]$revision
#> [1] "108164"
#> 
#> 
#> [[98]]
#> [[98]]$score
#> [[98]]$score$`0`
#> [1] "0.13322844"
#> 
#> 
#> [[98]]$fields
#> list()
#> 
#> [[98]]$id
#> [1] "WP4204"
#> 
#> [[98]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4204"
#> 
#> [[98]]$name
#> [1] "Tumor suppressor activity of SMARCB1"
#> 
#> [[98]]$species
#> [1] "Homo sapiens"
#> 
#> [[98]]$revision
#> [1] "108297"
#> 
#> 
#> [[99]]
#> [[99]]$score
#> [[99]]$score$`0`
#> [1] "0.117758416"
#> 
#> 
#> [[99]]$fields
#> list()
#> 
#> [[99]]$id
#> [1] "WP1080"
#> 
#> [[99]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP1080"
#> 
#> [[99]]$name
#> [1] "Irinotecan Pathway"
#> 
#> [[99]]$species
#> [1] "Bos taurus"
#> 
#> [[99]]$revision
#> [1] "101977"
#> 
#> 
#> [[100]]
#> [[100]]$score
#> [[100]]$score$`0`
#> [1] "0.117758416"
#> 
#> 
#> [[100]]$fields
#> list()
#> 
#> [[100]]$id
#> [1] "WP4542"
#> 
#> [[100]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4542"
#> 
#> [[100]]$name
#> [1] "Overview of leukocyte-intrinsic Hippo pathway functions"
#> 
#> [[100]]$species
#> [1] "Homo sapiens"
#> 
#> [[100]]$revision
#> [1] "104788"
#> 
#> 
#> [[101]]
#> [[101]]$score
#> [[101]]$score$`0`
#> [1] "0.117758416"
#> 
#> 
#> [[101]]$fields
#> list()
#> 
#> [[101]]$id
#> [1] "WP229"
#> 
#> [[101]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP229"
#> 
#> [[101]]$name
#> [1] "Irinotecan Pathway"
#> 
#> [[101]]$species
#> [1] "Homo sapiens"
#> 
#> [[101]]$revision
#> [1] "106624"
#> 
#> 
#> [[102]]
#> [[102]]$score
#> [[102]]$score$`0`
#> [1] "0.09992132"
#> 
#> 
#> [[102]]$fields
#> list()
#> 
#> [[102]]$id
#> [1] "WP4172"
#> 
#> [[102]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4172"
#> 
#> [[102]]$name
#> [1] "PI3K-Akt Signaling Pathway"
#> 
#> [[102]]$species
#> [1] "Homo sapiens"
#> 
#> [[102]]$revision
#> [1] "106781"
#> 
#> 
#> [[103]]
#> [[103]]$score
#> [[103]]$score$`0`
#> [1] "0.09420673"
#> 
#> 
#> [[103]]$fields
#> list()
#> 
#> [[103]]$id
#> [1] "WP3200"
#> 
#> [[103]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3200"
#> 
#> [[103]]$name
#> [1] "PDGF Pathway"
#> 
#> [[103]]$species
#> [1] "Bos taurus"
#> 
#> [[103]]$revision
#> [1] "107066"
#> 
#> 
#> [[104]]
#> [[104]]$score
#> [[104]]$score$`0`
#> [1] "0.09420673"
#> 
#> 
#> [[104]]$fields
#> list()
#> 
#> [[104]]$id
#> [1] "WP2526"
#> 
#> [[104]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP2526"
#> 
#> [[104]]$name
#> [1] "PDGF Pathway"
#> 
#> [[104]]$species
#> [1] "Homo sapiens"
#> 
#> [[104]]$revision
#> [1] "107132"
#> 
#> 
#> [[105]]
#> [[105]]$score
#> [[105]]$score$`0`
#> [1] "0.08471607"
#> 
#> 
#> [[105]]$fields
#> list()
#> 
#> [[105]]$id
#> [1] "WP4707"
#> 
#> [[105]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4707"
#> 
#> [[105]]$name
#> [1] "aspirin and miRNAs"
#> 
#> [[105]]$species
#> [1] "Homo sapiens"
#> 
#> [[105]]$revision
#> [1] "108258"
#> 
#> 
#> [[106]]
#> [[106]]$score
#> [[106]]$score$`0`
#> [1] "0.08326777"
#> 
#> 
#> [[106]]$fields
#> list()
#> 
#> [[106]]$id
#> [1] "WP4016"
#> 
#> [[106]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4016"
#> 
#> [[106]]$name
#> [1] "DNA IR-damage and cellular response via ATR"
#> 
#> [[106]]$species
#> [1] "Homo sapiens"
#> 
#> [[106]]$revision
#> [1] "108188"
#> 
#> 
#> [[107]]
#> [[107]]$score
#> [[107]]$score$`0`
#> [1] "0.082430884"
#> 
#> 
#> [[107]]$fields
#> list()
#> 
#> [[107]]$id
#> [1] "WP4263"
#> 
#> [[107]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4263"
#> 
#> [[107]]$name
#> [1] "Pancreatic adenocarcinoma pathway"
#> 
#> [[107]]$species
#> [1] "Homo sapiens"
#> 
#> [[107]]$revision
#> [1] "105482"
#> 
#> 
#> [[108]]
#> [[108]]$score
#> [[108]]$score$`0`
#> [1] "0.06661422"
#> 
#> 
#> [[108]]$fields
#> list()
#> 
#> [[108]]$id
#> [1] "WP3266"
#> 
#> [[108]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3266"
#> 
#> [[108]]$name
#> [1] "Signaling Pathways in Glioblastoma"
#> 
#> [[108]]$species
#> [1] "Bos taurus"
#> 
#> [[108]]$revision
#> [1] "108195"
#> 
#> 
#> [[109]]
#> [[109]]$score
#> [[109]]$score$`0`
#> [1] "0.06661422"
#> 
#> 
#> [[109]]$fields
#> list()
#> 
#> [[109]]$id
#> [1] "WP2261"
#> 
#> [[109]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP2261"
#> 
#> [[109]]$name
#> [1] "Signaling Pathways in Glioblastoma"
#> 
#> [[109]]$species
#> [1] "Homo sapiens"
#> 
#> [[109]]$revision
#> [1] "108196"
#> 
#> 
#> [[110]]
#> [[110]]$score
#> [[110]]$score$`0`
#> [1] "0.04492748"
#> 
#> 
#> [[110]]$fields
#> list()
#> 
#> [[110]]$id
#> [1] "WP3287"
#> 
#> [[110]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3287"
#> 
#> [[110]]$name
#> [1] "Overview of nanoparticle effects"
#> 
#> [[110]]$species
#> [1] "Homo sapiens"
#> 
#> [[110]]$revision
#> [1] "107141"
#> 
#> 
#> [[111]]
#> [[111]]$score
#> [[111]]$score$`0`
#> [1] "0.031768527"
#> 
#> 
#> [[111]]$fields
#> list()
#> 
#> [[111]]$id
#> [1] "WP4659"
#> 
#> [[111]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4659"
#> 
#> [[111]]$name
#> [1] "Gastrin Signaling Pathway"
#> 
#> [[111]]$species
#> [1] "Homo sapiens"
#> 
#> [[111]]$revision
#> [1] "107967"
#> 
#> 
#> [[112]]
#> [[112]]$score
#> [[112]]$score$`0`
#> [1] "0.026207695"
#> 
#> 
#> [[112]]$fields
#> list()
#> 
#> [[112]]$id
#> [1] "WP566"
#> 
#> [[112]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP566"
#> 
#> [[112]]$name
#> [1] "Canonical wnt signaling"
#> 
#> [[112]]$species
#> [1] "Danio rerio"
#> 
#> [[112]]$revision
#> [1] "96484"
#> 
#> 
#> [[113]]
#> [[113]]$score
#> [[113]]$score$`0`
#> [1] "0.02246374"
#> 
#> 
#> [[113]]$fields
#> list()
#> 
#> [[113]]$id
#> [1] "WP2355"
#> 
#> [[113]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP2355"
#> 
#> [[113]]$name
#> [1] "Corticotropin-releasing hormone signaling pathway"
#> 
#> [[113]]$species
#> [1] "Homo sapiens"
#> 
#> [[113]]$revision
#> [1] "108254"
#> 
#> 
#> [[114]]
#> [[114]]$score
#> [[114]]$score$`0`
#> [1] "0.02246374"
#> 
#> 
#> [[114]]$fields
#> list()
#> 
#> [[114]]$id
#> [1] "WP3265"
#> 
#> [[114]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3265"
#> 
#> [[114]]$name
#> [1] "Corticotropin-releasing hormone"
#> 
#> [[114]]$species
#> [1] "Bos taurus"
#> 
#> [[114]]$revision
#> [1] "108255"
#> 
#> 
#> [[115]]
#> [[115]]$score
#> [[115]]$score$`0`
#> [1] "0.018719783"
#> 
#> 
#> [[115]]$fields
#> list()
#> 
#> [[115]]$id
#> [1] "WP4754"
#> 
#> [[115]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4754"
#> 
#> [[115]]$name
#> [1] "IL-18 signaling pathway"
#> 
#> [[115]]$species
#> [1] "Homo sapiens"
#> 
#> [[115]]$revision
#> [1] "108627"
#> 
#>