Find Pathways By Xref

Retrieve pathways containing the query Xref by identifier and system code.

findPathwaysByXref(identifier, systemCode)

Arguments

identifier
(character) The official ID specified by a data source or system

systemCode
(character) The BridgeDb code associated with the data source or system, e.g., En (Ensembl), L (Entrez), Ch (HMDB), etc. See column two of https://github.com/bridgedb/BridgeDb/blob/master/org.bridgedb.bio/resources/org/bridgedb/bio/datasources.txt.

Value

A dataframe of pathway attributes in addition to query result score

Details

Note: there will be multiple listings of the same pathway if the Xref is present mutiple times.

The score is from a lucene index search engine, ranging from 0 to 1 with higher scores for better matches.

Examples

{
findPathwaysByXref('ENSG00000232810','En')
}

#>         score     id                                                   url
#>  1: 0.3951413 WP3893 https://www.wikipathways.org/index.php/Pathway:WP3893
#>  2: 0.3951413 WP2882 https://www.wikipathways.org/index.php/Pathway:WP2882
#>  3: 0.3951413 WP1584 https://www.wikipathways.org/index.php/Pathway:WP1584
#>  4: 0.3951413 WP2431 https://www.wikipathways.org/index.php/Pathway:WP2431
#>  5: 0.3951413 WP4673 https://www.wikipathways.org/index.php/Pathway:WP4673
#>  6: 0.3951413 WP2059 https://www.wikipathways.org/index.php/Pathway:WP2059
#>  7: 0.3951413 WP3599 https://www.wikipathways.org/index.php/Pathway:WP3599
#>  8: 0.3951413 WP3863 https://www.wikipathways.org/index.php/Pathway:WP3863
#>  9: 0.3951413 WP4008 https://www.wikipathways.org/index.php/Pathway:WP4008
#> 10: 0.3951413 WP3865 https://www.wikipathways.org/index.php/Pathway:WP3865
#> 11: 0.3951413 WP3624 https://www.wikipathways.org/index.php/Pathway:WP3624
#> 12: 0.3951413 WP4341 https://www.wikipathways.org/index.php/Pathway:WP4341
#> 13: 0.3951413 WP4298 https://www.wikipathways.org/index.php/Pathway:WP4298
#> 14: 0.3951413 WP4666 https://www.wikipathways.org/index.php/Pathway:WP4666
#> 15: 0.3951413 WP3941 https://www.wikipathways.org/index.php/Pathway:WP3941
#> 16: 0.3951413 WP2813 https://www.wikipathways.org/index.php/Pathway:WP2813
#> 17: 0.3951413  WP262  https://www.wikipathways.org/index.php/Pathway:WP262
#> 18: 0.3951413 WP2873 https://www.wikipathways.org/index.php/Pathway:WP2873
#> 19: 0.3951413 WP2571 https://www.wikipathways.org/index.php/Pathway:WP2571
#> 20: 0.3951413 WP4329 https://www.wikipathways.org/index.php/Pathway:WP4329
#> 21: 0.3951413 WP4136 https://www.wikipathways.org/index.php/Pathway:WP4136
#> 22: 0.3951413 WP4864 https://www.wikipathways.org/index.php/Pathway:WP4864
#> 23: 0.3951413 WP3287 https://www.wikipathways.org/index.php/Pathway:WP3287
#> 24: 0.3951413 WP4066 https://www.wikipathways.org/index.php/Pathway:WP4066
#> 25: 0.3951413 WP4063 https://www.wikipathways.org/index.php/Pathway:WP4063
#> 26: 0.3951413 WP4969 https://www.wikipathways.org/index.php/Pathway:WP4969
#> 27: 0.2794071 WP2795 https://www.wikipathways.org/index.php/Pathway:WP2795
#> 28: 0.2794071 WP4847 https://www.wikipathways.org/index.php/Pathway:WP4847
#> 29: 0.2794071  WP129  https://www.wikipathways.org/index.php/Pathway:WP129
#> 30: 0.2794071  WP314  https://www.wikipathways.org/index.php/Pathway:WP314
#> 31: 0.2794071 WP2513 https://www.wikipathways.org/index.php/Pathway:WP2513
#> 32: 0.2794071 WP1449 https://www.wikipathways.org/index.php/Pathway:WP1449
#> 33: 0.2794071 WP4481 https://www.wikipathways.org/index.php/Pathway:WP4481
#> 34: 0.2794071 WP1544 https://www.wikipathways.org/index.php/Pathway:WP1544
#> 35: 0.2794071 WP4482 https://www.wikipathways.org/index.php/Pathway:WP4482
#> 36: 0.2794071 WP3297 https://www.wikipathways.org/index.php/Pathway:WP3297
#> 37: 0.2794071  WP530  https://www.wikipathways.org/index.php/Pathway:WP530
#> 38: 0.2794071 WP4891 https://www.wikipathways.org/index.php/Pathway:WP4891
#> 39: 0.2794071  WP236  https://www.wikipathways.org/index.php/Pathway:WP236
#> 40: 0.2794071 WP4747 https://www.wikipathways.org/index.php/Pathway:WP4747
#> 41: 0.2794071  WP176  https://www.wikipathways.org/index.php/Pathway:WP176
#> 42: 0.2794071  WP727  https://www.wikipathways.org/index.php/Pathway:WP727
#> 43: 0.2794071 WP4496 https://www.wikipathways.org/index.php/Pathway:WP4496
#> 44: 0.2794071  WP254  https://www.wikipathways.org/index.php/Pathway:WP254
#> 45: 0.2794071 WP1533 https://www.wikipathways.org/index.php/Pathway:WP1533
#> 46: 0.2794071   WP75   https://www.wikipathways.org/index.php/Pathway:WP75
#> 47: 0.2794071 WP4478 https://www.wikipathways.org/index.php/Pathway:WP4478
#> 48: 0.2794071 WP2328 https://www.wikipathways.org/index.php/Pathway:WP2328
#> 49: 0.2794071 WP3851 https://www.wikipathways.org/index.php/Pathway:WP3851
#> 50: 0.2794071  WP231  https://www.wikipathways.org/index.php/Pathway:WP231
#> 51: 0.2794071 WP3617 https://www.wikipathways.org/index.php/Pathway:WP3617
#> 52: 0.2794071 WP2036 https://www.wikipathways.org/index.php/Pathway:WP2036
#> 53: 0.2794071 WP3611 https://www.wikipathways.org/index.php/Pathway:WP3611
#> 54: 0.2794071  WP560  https://www.wikipathways.org/index.php/Pathway:WP560
#> 55: 0.2794071 WP4816 https://www.wikipathways.org/index.php/Pathway:WP4816
#> 56: 0.2794071 WP2864 https://www.wikipathways.org/index.php/Pathway:WP2864
#> 57: 0.2794071 WP4304 https://www.wikipathways.org/index.php/Pathway:WP4304
#> 58: 0.2794071 WP4857 https://www.wikipathways.org/index.php/Pathway:WP4857
#> 59: 0.2794071   WP15   https://www.wikipathways.org/index.php/Pathway:WP15
#> 60: 0.2794071 WP2586 https://www.wikipathways.org/index.php/Pathway:WP2586
#> 61: 0.2794071 WP4493 https://www.wikipathways.org/index.php/Pathway:WP4493
#> 62: 0.2794071 WP4396 https://www.wikipathways.org/index.php/Pathway:WP4396
#> 63: 0.2794071 WP4480 https://www.wikipathways.org/index.php/Pathway:WP4480
#> 64: 0.2794071  WP706  https://www.wikipathways.org/index.php/Pathway:WP706
#> 65: 0.2794071 WP2447 https://www.wikipathways.org/index.php/Pathway:WP2447
#> 66: 0.2794071  WP382  https://www.wikipathways.org/index.php/Pathway:WP382
#> 67: 0.2794071 WP5044 https://www.wikipathways.org/index.php/Pathway:WP5044
#> 68: 0.2794071 WP2751 https://www.wikipathways.org/index.php/Pathway:WP2751
#> 69: 0.2794071 WP3380 https://www.wikipathways.org/index.php/Pathway:WP3380
#> 70: 0.2794071 WP3398 https://www.wikipathways.org/index.php/Pathway:WP3398
#> 71: 0.2794071 WP4066 https://www.wikipathways.org/index.php/Pathway:WP4066
#> 72: 0.2794071 WP4063 https://www.wikipathways.org/index.php/Pathway:WP4063
#> 73: 0.2794071 WP4754 https://www.wikipathways.org/index.php/Pathway:WP4754
#> 74: 0.2794071 WP5039 https://www.wikipathways.org/index.php/Pathway:WP5039
#>         score     id                                                   url
#>                                                                                        name
#>  1:                                         Development and heterogeneity of the ILC family
#>  2:                                                          Nuclear Receptors Meta-Pathway
#>  3:                                                               Type II diabetes mellitus
#>  4:                                                                      Spinal Cord Injury
#>  5:                                                      Genes involved in male infertility
#>  6:                                                                      Alzheimers Disease
#>  7:                                         Transcription factor regulation in adipogenesis
#>  8:            T-Cell antigen Receptor (TCR) pathway during Staphylococcus aureus infection
#>  9:                                                    NO/cGMP/PKG mediated Neuroprotection
#> 10:                     Novel intracellular components of RIG-I-like receptor (RLR) pathway
#> 11:                                                                           Lung fibrosis
#> 12:                                         Non-genomic actions of 1,25 dihydroxyvitamin D3
#> 13:                                                                 Viral Acute Myocarditis
#> 14:                                                                   Hepatitis B infection
#> 15:                                                                        Oxidative Damage
#> 16:                Mammary gland development pathway - Embryonic development (Stage 1 of 4)
#> 17:                                                                      EBV LMP1 signaling
#> 18:                                                       Aryl Hydrocarbon Receptor Pathway
#> 19:                                                       Polycystic Kidney Disease Pathway
#> 20:                                     miRNAs involvement in the immune response in sepsis
#> 21:                                          Fibrin Complement Receptor 3 Signaling Pathway
#> 22:                           Host-pathogen interaction of human corona viruses - apoptosis
#> 23:                                                        Overview of nanoparticle effects
#> 24:                                              Interleukin-4 and Interleukin-13 signaling
#> 25:                                                                Interleukin-10 signaling
#> 26:                                                 RAS and Bradykinin pathways in COVID-19
#> 27:                                                           Cardiac Hypertrophic Response
#> 28:                                                                   AXL signaling pathway
#> 29:                                                               Matrix Metalloproteinases
#> 30:  Fas Ligand (FasL) pathway and Stress induction of Heat Shock Proteins (HSP) regulation
#> 31:                                               Nanoparticle triggered regulated necrosis
#> 32:                                      Regulation of toll-like receptor signaling pathway
#> 33:                                                 Resistin as a regulator of inflammation
#> 34:                                                  MicroRNAs in cardiomyocyte hypertrophy
#> 35:                                                      Vitamin D in inflammatory diseases
#> 36:                              EV release from cardiac cells and their functional effects
#> 37:                                                     Cytokines and Inflammatory Response
#> 38:                                                        COVID-19 Adverse Outcome Pathway
#> 39:                                                                            Adipogenesis
#> 40:                                                          Netrin-UNC5B signaling Pathway
#> 41:                                                                       Folate Metabolism
#> 42:                                                                     Monoamine Transport
#> 43:                                                        Signal transduction through IL1R
#> 44:                                                                               Apoptosis
#> 45:                                                                  Vitamin B12 Metabolism
#> 46:                                                    Toll-like Receptor Signaling Pathway
#> 47:                                                      LTF danger signal response pathway
#> 48:                                                                     Allograft Rejection
#> 49:                                                            TLR4 Signaling and Tolerance
#> 50:                                                             TNF alpha Signaling Pathway
#> 51:                                   Photodynamic therapy-induced NF-kB survival signaling
#> 52:                         TNF related weak inducer of apoptosis (TWEAK) Signaling Pathway
#> 53:                                   Photodynamic therapy-induced AP-1 survival signaling.
#> 54:                                                             TGF-beta Receptor Signaling
#> 55:                                     TGF-beta Receptor Signalling in Skeletal Dysplasias
#> 56:                       Apoptosis-related network due to altered Notch3 in ovarian cancer
#> 57: Oligodendrocyte specification and differentiation, leading to Myelin Components for CNS
#> 58:                                                                      Endothelin Pathway
#> 59:                                                          Selenium Micronutrient Network
#> 60:                                                       Aryl Hydrocarbon Receptor Netpath
#> 61:                      Cells and Molecules involved in local acute inflammatory response 
#> 62:                                                        Nonalcoholic fatty liver disease
#> 63:                        Role of Altered Glycolysation of MUC1 in Tumour Microenvironment
#> 64:                             Sudden Infant Death Syndrome (SIDS) Susceptibility Pathways
#> 65:                                                     Amyotrophic lateral sclerosis (ALS)
#> 66:                                                                  MAPK Signaling Pathway
#> 67:                                     Kynurenine Pathway and links to Cellular Senescence
#> 68:                           Transcriptional regulation of white adipocyte differentiation
#> 69:                                                                           TNF signaling
#> 70:                                                       TNFR2 non-canonical NF-kB pathway
#> 71:                                              Interleukin-4 and Interleukin-13 signaling
#> 72:                                                                Interleukin-10 signaling
#> 73:                                                                 IL-18 signaling pathway
#> 74:                    SARS-CoV-2 Innate Immunity Evasion and Cell-specific immune response
#>                                                                                        name
#>          species revision
#>  1: Homo sapiens   106386
#>  2: Homo sapiens   103038
#>  3: Homo sapiens   111210
#>  4: Homo sapiens   106712
#>  5: Homo sapiens   106810
#>  6: Homo sapiens   108233
#>  7: Homo sapiens   105845
#>  8: Homo sapiens   106753
#>  9: Homo sapiens   110707
#> 10: Homo sapiens   106524
#> 11: Homo sapiens   110182
#> 12: Homo sapiens   107169
#> 13: Homo sapiens   110705
#> 14: Homo sapiens   107486
#> 15: Homo sapiens   106568
#> 16: Homo sapiens   102405
#> 17: Homo sapiens   106391
#> 18: Homo sapiens   106304
#> 19: Homo sapiens   108308
#> 20: Homo sapiens   101725
#> 21: Homo sapiens   106418
#> 22: Homo sapiens   113801
#> 23: Homo sapiens   113822
#> 24: Homo sapiens   115055
#> 25: Homo sapiens   115095
#> 26: Homo sapiens   115143
#> 27: Homo sapiens   106724
#> 28: Homo sapiens   111812
#> 29: Homo sapiens   106640
#> 30: Homo sapiens   109375
#> 31: Homo sapiens   110691
#> 32: Homo sapiens   106494
#> 33: Homo sapiens   102529
#> 34: Homo sapiens   108800
#> 35: Homo sapiens   112969
#> 36: Homo sapiens   106734
#> 37: Homo sapiens   111164
#> 38: Homo sapiens   111616
#> 39: Homo sapiens   105873
#> 40: Homo sapiens   108877
#> 41: Homo sapiens   107123
#> 42: Homo sapiens   107191
#> 43: Homo sapiens   102693
#> 44: Homo sapiens   106302
#> 45: Homo sapiens   107121
#> 46: Homo sapiens   111250
#> 47: Homo sapiens   103756
#> 48: Homo sapiens   106557
#> 49: Homo sapiens   105835
#> 50: Homo sapiens   111053
#> 51: Homo sapiens   106541
#> 52: Homo sapiens   105837
#> 53: Homo sapiens    90923
#> 54: Homo sapiens   106477
#> 55: Homo sapiens   111739
#> 56: Homo sapiens   108133
#> 57: Homo sapiens   112156
#> 58: Homo sapiens   111629
#> 59: Homo sapiens   107118
#> 60: Homo sapiens   107439
#> 61: Homo sapiens   102660
#> 62: Homo sapiens    98945
#> 63: Homo sapiens   107458
#> 64: Homo sapiens   113813
#> 65: Homo sapiens   113794
#> 66: Homo sapiens   113991
#> 67: Homo sapiens   114613
#> 68: Homo sapiens   114880
#> 69: Homo sapiens   114940
#> 70: Homo sapiens   115048
#> 71: Homo sapiens   115055
#> 72: Homo sapiens   115095
#> 73: Homo sapiens   115147
#> 74: Homo sapiens   115153
#>          species revision