Get Pathway

Retrieve a specific pathway in the GPML format

getPathway(pathway, revision = 0)

Arguments

pathway
WikiPathways identifier (WPID) for the pathway to download, e.g. WP4

revision
(integer, optional) Number to indicate a specific revision to download

Value

GPML

Examples

{
getPathway('WP554')
}

#> [[1]]
#> [1] "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n<Pathway xmlns=\"http://pathvisio.org/GPML/2013a\" Name=\"ACE Inhibitor Pathway\" Organism=\"Homo sapiens\">\n  <Comment Source=\"WikiPathways-description\">The core of this pathway was elucidated over a century ago and involves the conversion of angiotensinogen to angiotensin I (Ang I) by renin, its subsequent conversion to angiotensin II (Ang II) by angiotensin converting enzyme. Ang II activates the angiotensin II receptor type 1 to induce aldosterone synthesis, increasing water and salt resorption and potassium excretion in the kidney and increasing blood pressure.\n\nSource: PharmGKB (https://www.pharmgkb.org/pathway/PA2023)\n\nProteins on this pathway have targeted assays available via the [https://assays.cancer.gov/available_assays?wp_id=WP554 CPTAC Assay Portal]</Comment>\n  <BiopaxRef>b93</BiopaxRef>\n  <Graphics BoardWidth=\"991.0\" BoardHeight=\"651.0\" />\n  <DataNode TextLabel=\"NOS3\" GraphId=\"ab119\" Type=\"GeneProduct\">\n    <Attribute Key=\"org.pathvisio.model.BackpageHead\" Value=\"NOS3 | nitric oxide synthase 3 (endothelial cel...\" />\n    <Graphics CenterX=\"759.0\" CenterY=\"512.3333333333333\" Width=\"60.0\" Height=\"20.0\" ZOrder=\"32768\" FontSize=\"10\" Valign=\"Middle\" />\n    <Xref Database=\"Entrez Gene\" ID=\"4846\" />\n  </DataNode>\n  <DataNode TextLabel=\"AGTR2\" GraphId=\"b1122\" Type=\"GeneProduct\">\n    <BiopaxRef>bc9</BiopaxRef>\n    <Attribute Key=\"org.pathvisio.model.BackpageHead\" Value=\"AGTR2 | angiotensin II receptor type 2\" />\n    <Graphics CenterX=\"351.3333333333333\" CenterY=\"390.3333333333333\" Width=\"60.0\" Height=\"20.0\" ZOrder=\"32768\" FontSize=\"10\" Valign=\"Middle\" />\n    <Xref Database=\"Entrez Gene\" ID=\"186\" />\n  </DataNode>\n  <DataNode TextLabel=\"Ang 1-5\" GraphId=\"b298c\" Type=\"Metabolite\">\n    <Graphics CenterX=\"102.0\" CenterY=\"326.0\" Width=\"54.0\" Height=\"20.0\" ZOrder=\"32768\" FontSize=\"10\" Valign=\"Middle\" Color=\"0000ff\" />\n    <Xref Database=\"KEGG Compound\" ID=\"C15852\" />\n  </DataNode>\n  <DataNode TextLabel=\"BDKRB1\" GraphId=\"b4349\" Type=\"GeneProduct\">\n    <Graphics CenterX=\"681.5\" CenterY=\"388.0\" Width=\"57.0\" Height=\"20.0\" ZOrder=\"32768\" FontSize=\"10\" Valign=\"Middle\" />\n    <Xref Database=\"Entrez Gene\" ID=\"623\" />\n  </DataNode>\n  <DataNode TextLabel=\"TFs\" GraphId=\"b47da\" Type=\"GeneProduct\">\n    <Graphics CenterX=\"395.5\" CenterY=\"469.0\" Width=\"37.0\" Height=\"20.0\" ZOrder=\"32768\" FontSize=\"10\" Valign=\"Middle\" />\n    <Xref Database=\"\" ID=\"\" />\n  </DataNode>\n  <DataNode TextLabel=\"AGTR1\" GraphId=\"b4811\" Type=\"GeneProduct\">\n    <Attribute Key=\"org.pathvisio.model.BackpageHead\" Value=\"AGTR1 | angiotensin II receptor type 1\" />\n    <Graphics CenterX=\"428.3333333333333\" CenterY=\"390.3333333333333\" Width=\"60.0\" Height=\"20.0\" ZOrder=\"32768\" FontSize=\"10\" Valign=\"Middle\" />\n    <Xref Database=\"Entrez Gene\" ID=\"185\" />\n  </DataNode>\n  <DataNode TextLabel=\"Aldosterone\" GraphId=\"b75c3\" Type=\"Metabolite\">\n    <BiopaxRef>ce3</BiopaxRef>\n    <Graphics CenterX=\"563.0\" CenterY=\"530.0\" Width=\"80.0\" Height=\"20.0\" ZOrder=\"32768\" FontSize=\"10\" Valign=\"Middle\" Color=\"0000ff\" />\n    <Xref Database=\"HMDB\" ID=\"HMDB0000037\" />\n  </DataNode>\n  <DataNode TextLabel=\"TGFB1\" GraphId=\"becdd\" Type=\"GeneProduct\">\n    <BiopaxRef>af8</BiopaxRef>\n    <Graphics CenterX=\"192.5\" CenterY=\"481.0\" Width=\"59.0\" Height=\"20.0\" ZOrder=\"32768\" FontSize=\"10\" Valign=\"Middle\" />\n    <Xref Database=\"Entrez Gene\" ID=\"7040\" />\n  </DataNode>\n  <DataNode TextLabel=\"Bradykinin\" GraphId=\"c496e\" Type=\"Metabolite\">\n    <Graphics CenterX=\"717.0\" CenterY=\"319.0\" Width=\"80.0\" Height=\"20.0\" ZOrder=\"32768\" FontSize=\"10\" Valign=\"Middle\" Color=\"0000ff\" />\n    <Xref Database=\"HMDB\" ID=\"HMDB0004246\" />\n  </DataNode>\n  <DataNode TextLabel=\"CTSG\" GraphId=\"c9ca8\" Type=\"GeneProduct\">\n    <Graphics CenterX=\"321.5\" CenterY=\"231.0\" Width=\"60.0\" Height=\"20.0\" ZOrder=\"32768\" FontSize=\"10\" Valign=\"Middle\" />\n    <Xref Database=\"Entrez Gene\" ID=\"1511\" />\n  </DataNode>\n  <DataNode TextLabel=\"Angiotensin II\" GraphId=\"ca3ad\" Type=\"Metabolite\">\n    <Graphics CenterX=\"383.0\" CenterY=\"326.0\" Width=\"80.0\" Height=\"20.0\" ZOrder=\"32768\" FontSize=\"10\" Valign=\"Middle\" Color=\"0000ff\" />\n    <Xref Database=\"HMDB\" ID=\"HMDB0001035\" />\n  </DataNode>\n  <DataNode TextLabel=\"KNG1\" GraphId=\"cc0b4\" Type=\"GeneProduct\">\n    <Attribute Key=\"org.pathvisio.model.BackpageHead\" Value=\"KNG1 | kininogen 1\" />\n    <Graphics CenterX=\"717.6666666666667\" CenterY=\"247.33333333333331\" Width=\"60.0\" Height=\"20.0\" ZOrder=\"32768\" FontSize=\"10\" Valign=\"Middle\" />\n    <Xref Database=\"Entrez Gene\" ID=\"3827\" />\n  </DataNode>\n  <DataNode TextLabel=\"Ca++\" GraphId=\"d519e\" Type=\"Metabolite\">\n    <Graphics CenterX=\"371.0\" CenterY=\"425.0\" Width=\"32.0\" Height=\"20.0\" ZOrder=\"32768\" FontSize=\"10\" Valign=\"Middle\" Color=\"0000ff\" />\n    <Xref Database=\"HMDB\" ID=\"HMDB0000464\" />\n  </DataNode>\n  <DataNode TextLabel=\"Angiotensin I\" GraphId=\"d6a08\" Type=\"Metabolite\">\n    <Graphics CenterX=\"382.0\" CenterY=\"201.0\" Width=\"80.0\" Height=\"20.0\" ZOrder=\"32768\" FontSize=\"10\" Valign=\"Middle\" Color=\"0000ff\" />\n    <Xref Database=\"PubChem-compound\" ID=\"3081372\" />\n  </DataNode>\n  <DataNode TextLabel=\"ACE\" GraphId=\"d6f1e\" Type=\"GeneProduct\">\n    <Attribute Key=\"org.pathvisio.model.BackpageHead\" Value=\"ACE | angiotensin I converting enzyme (peptidyl...\" />\n    <Graphics CenterX=\"652.3333333333333\" CenterY=\"283.66666666666663\" Width=\"60.0\" Height=\"20.0\" ZOrder=\"32768\" FontSize=\"10\" Valign=\"Middle\" />\n    <Xref Database=\"Entrez Gene\" ID=\"1636\" />\n  </DataNode>\n  <DataNode TextLabel=\"ACE Inhibitor\" GraphId=\"d9620\" Type=\"Metabolite\">\n    <Attribute Key=\"org.pathvisio.model.BackpageHead\" Value=\"ACE Inhibitor\" />\n    <Graphics CenterX=\"467.5\" CenterY=\"40.66666666666666\" Width=\"91.0\" Height=\"20.0\" ZOrder=\"32768\" FontSize=\"10\" Valign=\"Middle\" Color=\"cc0000\" />\n    <Xref Database=\"ChEBI\" ID=\"CHEBI:35457\" />\n  </DataNode>\n  <DataNode TextLabel=\"Deoxycorticosterone\" GraphId=\"da9ee\" Type=\"Metabolite\">\n    <Graphics CenterX=\"407.875\" CenterY=\"529.75\" Width=\"113.75\" Height=\"20.0\" ZOrder=\"32768\" FontSize=\"10\" Valign=\"Middle\" Color=\"0000ff\" />\n    <Xref Database=\"HMDB\" ID=\"HMDB0000016\" />\n  </DataNode>\n  <DataNode TextLabel=\"CMA1\" GraphId=\"de1e4\" Type=\"GeneProduct\">\n    <Graphics CenterX=\"301.0\" CenterY=\"257.0\" Width=\"60.0\" Height=\"20.0\" ZOrder=\"32768\" FontSize=\"10\" Valign=\"Middle\" />\n    <Xref Database=\"Entrez Gene\" ID=\"1215\" />\n  </DataNode>\n  <DataNode TextLabel=\"Ang 1-7\" GraphId=\"df262\" Type=\"Metabolite\">\n    <Graphics CenterX=\"204.5\" CenterY=\"326.0\" Width=\"51.0\" Height=\"20.0\" ZOrder=\"32768\" FontSize=\"10\" Valign=\"Middle\" Color=\"0000ff\" />\n    <Xref Database=\"Chemspider\" ID=\"110354\" />\n  </DataNode>\n  <DataNode TextLabel=\"Prostacyclin\" GraphId=\"e01c3\" Type=\"Metabolite\">\n    <Graphics CenterX=\"693.0\" CenterY=\"466.0\" Width=\"70.0\" Height=\"20.0\" ZOrder=\"32768\" FontSize=\"10\" Valign=\"Middle\" Color=\"0000ff\" />\n    <Xref Database=\"PubChem-compound\" ID=\"114805\" />\n  </DataNode>\n  <DataNode TextLabel=\"REN\" GraphId=\"e0a11\" Type=\"GeneProduct\">\n    <Attribute Key=\"org.pathvisio.model.BackpageHead\" Value=\"REN | renin\" />\n    <Graphics CenterX=\"319.08123249299706\" CenterY=\"118.94957983193277\" Width=\"44.66666666666663\" Height=\"20.0\" ZOrder=\"32768\" FontSize=\"10\" Valign=\"Middle\" />\n    <Xref Database=\"Entrez Gene\" ID=\"5972\" />\n  </DataNode>\n  <DataNode TextLabel=\"CYP11B2\" GraphId=\"e14ac\" Type=\"GeneProduct\">\n    <BiopaxRef>ce3</BiopaxRef>\n    <Graphics CenterX=\"491.0\" CenterY=\"479.0\" Width=\"62.0\" Height=\"20.0\" ZOrder=\"32768\" FontSize=\"10\" Valign=\"Middle\" />\n    <Xref Database=\"Entrez Gene\" ID=\"1585\" />\n  </DataNode>\n  <DataNode TextLabel=\"MAS1\" GraphId=\"e3012\" Type=\"GeneProduct\">\n    <BiopaxRef>c74</BiopaxRef>\n    <Graphics CenterX=\"204.0\" CenterY=\"389.0\" Width=\"52.0\" Height=\"20.0\" ZOrder=\"32768\" FontSize=\"10\" Valign=\"Middle\" />\n    <Xref Database=\"Entrez Gene\" ID=\"4142\" />\n  </DataNode>\n  <DataNode TextLabel=\"ACE2\" GraphId=\"e68e0\" Type=\"GeneProduct\">\n    <BiopaxRef>c74</BiopaxRef>\n    <Graphics CenterX=\"282.5\" CenterY=\"168.0\" Width=\"60.0\" Height=\"20.0\" ZOrder=\"32768\" FontSize=\"10\" Valign=\"Middle\" />\n    <Xref Database=\"Entrez Gene\" ID=\"59272\" />\n  </DataNode>\n  <DataNode TextLabel=\"AGT\" GraphId=\"e89d5\" Type=\"GeneProduct\">\n    <Attribute Key=\"org.pathvisio.model.BackpageHead\" Value=\"AGT | angiotensinogen (serine (or cysteine) pro...\" />\n    <Graphics CenterX=\"382.66666666666663\" CenterY=\"88.0\" Width=\"60.0\" Height=\"20.0\" ZOrder=\"32768\" FontSize=\"10\" Valign=\"Middle\" />\n    <Xref Database=\"Entrez Gene\" ID=\"183\" />\n  </DataNode>\n  <DataNode TextLabel=\"NR3C2\" GraphId=\"ec3b5\" Type=\"GeneProduct\">\n    <BiopaxRef>ce3</BiopaxRef>\n    <BiopaxRef>b5d</BiopaxRef>\n    <Graphics CenterX=\"563.5\" CenterY=\"588.0\" Width=\"55.0\" Height=\"20.0\" ZOrder=\"32768\" FontSize=\"10\" Valign=\"Middle\" />\n    <Xref Database=\"Entrez Gene\" ID=\"4306\" />\n  </DataNode>\n  <DataNode TextLabel=\"ATP6AP2\" GraphId=\"ee619\" Type=\"GeneProduct\">\n    <BiopaxRef>c16</BiopaxRef>\n    <Graphics CenterX=\"78.0\" CenterY=\"389.0\" Width=\"64.0\" Height=\"20.0\" ZOrder=\"32768\" FontSize=\"10\" Valign=\"Middle\" />\n    <Xref Database=\"Entrez Gene\" ID=\"10159\" />\n  </DataNode>\n  <DataNode TextLabel=\"ACE\" GraphId=\"ef1f3\" Type=\"GeneProduct\">\n    <BiopaxRef>bd1</BiopaxRef>\n    <Attribute Key=\"org.pathvisio.model.BackpageHead\" Value=\"ACE | angiotensin I converting enzyme (peptidyl...\" />\n    <Graphics CenterX=\"468.0\" CenterY=\"253.66666666666663\" Width=\"60.0\" Height=\"20.0\" ZOrder=\"32768\" FontSize=\"10\" Valign=\"Middle\" />\n    <Xref Database=\"Entrez Gene\" ID=\"1636\" />\n  </DataNode>\n  <DataNode TextLabel=\"BDKRB2\" GraphId=\"f54f5\" Type=\"GeneProduct\">\n    <Attribute Key=\"org.pathvisio.model.BackpageHead\" Value=\"BDKRB2 | bradykinin receptor B2\" />\n    <Graphics CenterX=\"759.3333333333333\" CenterY=\"388.0\" Width=\"60.0\" Height=\"20.0\" ZOrder=\"32768\" FontSize=\"10\" Valign=\"Middle\" />\n    <Xref Database=\"Entrez Gene\" ID=\"624\" />\n  </DataNode>\n  <DataNode TextLabel=\"Ang 1-9\" GraphId=\"fbc8b\" Type=\"Metabolite\">\n    <Graphics CenterX=\"204.5\" CenterY=\"201.0\" Width=\"51.0\" Height=\"20.0\" ZOrder=\"32768\" FontSize=\"10\" Valign=\"Middle\" Color=\"0000ff\" />\n    <Xref Database=\"KEGG Compound\" ID=\"C15851\" />\n  </DataNode>\n  <DataNode TextLabel=\"TFs\" GraphId=\"fd8ec\" Type=\"GeneProduct\">\n    <Graphics CenterX=\"97.5\" CenterY=\"470.0\" Width=\"37.0\" Height=\"20.0\" ZOrder=\"32768\" FontSize=\"10\" Valign=\"Middle\" />\n    <Xref Database=\"\" ID=\"\" />\n  </DataNode>\n  <DataNode TextLabel=\"ACE\" GraphId=\"d046f\" Type=\"GeneProduct\">\n    <Attribute Key=\"org.pathvisio.model.BackpageHead\" Value=\"ACE | angiotensin I converting enzyme (peptidyl...\" />\n    <Graphics CenterX=\"155.0\" CenterY=\"286.66666666666663\" Width=\"60.0\" Height=\"20.0\" ZOrder=\"32770\" FontSize=\"10\" Valign=\"Middle\" />\n    <Xref Database=\"Entrez Gene\" ID=\"1636\" />\n  </DataNode>\n  <DataNode TextLabel=\"ACE2\" GraphId=\"e08c8\" Type=\"GeneProduct\">\n    <Attribute Key=\"org.pathvisio.model.BackpageHead\" Value=\"ACE | angiotensin I converting enzyme (peptidyl...\" />\n    <Graphics CenterX=\"274.0\" CenterY=\"292.66666666666663\" Width=\"60.0\" Height=\"20.0\" ZOrder=\"32770\" FontSize=\"10\" Valign=\"Middle\" />\n    <Xref Database=\"Entrez Gene\" ID=\"59272\" />\n  </DataNode>\n  <Interaction GraphId=\"a0cae\">\n    <BiopaxRef>bc9</BiopaxRef>\n    <Graphics ZOrder=\"12288\" LineThickness=\"1.0\">\n      <Point X=\"331.0\" Y=\"257.0\" GraphRef=\"de1e4\" RelX=\"1.0\" RelY=\"0.0\" />\n      <Point X=\"362.60003627788865\" Y=\"274.0038091783058\" GraphRef=\"b5861\" RelX=\"0.0\" RelY=\"0.0\" ArrowHead=\"mim-catalysis\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"a21fb\">\n    <Graphics ZOrder=\"12288\" LineThickness=\"1.0\">\n      <Point X=\"563.0\" Y=\"540.0\" GraphRef=\"b75c3\" RelX=\"0.0\" RelY=\"1.0\" />\n      <Point X=\"563.5\" Y=\"578.0\" GraphRef=\"ec3b5\" RelX=\"0.0\" RelY=\"-1.0\" ArrowHead=\"mim-binding\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"a24b1\">\n    <Graphics ZOrder=\"12288\" LineThickness=\"1.0\">\n      <Point X=\"342.0\" Y=\"201.0\" GraphRef=\"d6a08\" RelX=\"-1.0\" RelY=\"0.0\" />\n      <Point X=\"230.0\" Y=\"201.0\" GraphRef=\"fbc8b\" RelX=\"1.0\" RelY=\"0.0\" ArrowHead=\"mim-conversion\" />\n      <Anchor Position=\"0.5368421052631579\" Shape=\"None\" GraphId=\"da414\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"a45d0\">\n    <Graphics ZOrder=\"12288\" LineThickness=\"1.0\">\n      <Point X=\"362.0\" Y=\"211.0\" GraphRef=\"d6a08\" RelX=\"-0.5\" RelY=\"1.0\" />\n      <Point X=\"363.0\" Y=\"316.0\" GraphRef=\"ca3ad\" RelX=\"-0.5\" RelY=\"-1.0\" ArrowHead=\"mim-conversion\" />\n      <Anchor Position=\"0.6000362778886266\" Shape=\"None\" GraphId=\"b5861\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"a76b3\">\n    <Graphics ZOrder=\"12288\" LineThickness=\"1.0\">\n      <Point X=\"382.66666666666663\" Y=\"98.0\" GraphRef=\"e89d5\" RelX=\"0.0\" RelY=\"1.0\" />\n      <Point X=\"382.0\" Y=\"191.0\" GraphRef=\"d6a08\" RelX=\"0.0\" RelY=\"-1.0\" ArrowHead=\"Arrow\" />\n      <Anchor Position=\"0.2365983685528935\" Shape=\"None\" GraphId=\"ed881\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"a84e7\">\n    <Graphics ZOrder=\"12288\" LineThickness=\"1.0\">\n      <Point X=\"438.0\" Y=\"253.66666666666663\" GraphRef=\"ef1f3\" RelX=\"-1.0\" RelY=\"0.0\" />\n      <Point X=\"402.4\" Y=\"253.0\" GraphRef=\"da262\" RelX=\"0.0\" RelY=\"0.0\" ArrowHead=\"mim-catalysis\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"aa8ff\">\n    <Graphics ZOrder=\"12288\" LineThickness=\"1.0\">\n      <Point X=\"717.6666666666667\" Y=\"257.3333333333333\" GraphRef=\"cc0b4\" RelX=\"0.0\" RelY=\"1.0\" />\n      <Point X=\"717.0\" Y=\"309.0\" GraphRef=\"c496e\" RelX=\"0.0\" RelY=\"-1.0\" ArrowHead=\"mim-conversion\" />\n      <Anchor Position=\"0.5047596606974556\" Shape=\"None\" GraphId=\"c11bc\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"afbf9\">\n    <Graphics ZOrder=\"12288\" LineStyle=\"Broken\" LineThickness=\"1.0\">\n      <Point X=\"759.3333333333333\" Y=\"398.0\" GraphRef=\"f54f5\" RelX=\"0.0\" RelY=\"1.0\" />\n      <Point X=\"759.0\" Y=\"502.33333333333326\" GraphRef=\"ab119\" RelX=\"0.0\" RelY=\"-1.0\" ArrowHead=\"Arrow\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"bb339\">\n    <Graphics ZOrder=\"12288\" LineThickness=\"1.0\">\n      <Point X=\"491.0\" Y=\"489.0\" GraphRef=\"e14ac\" RelX=\"0.0\" RelY=\"1.0\" />\n      <Point X=\"492.000571007552\" Y=\"529.8669552403758\" GraphRef=\"ff984\" RelX=\"0.0\" RelY=\"0.0\" ArrowHead=\"mim-catalysis\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"bed79\">\n    <Graphics ZOrder=\"12288\" LineThickness=\"1.0\">\n      <Point X=\"717.0\" Y=\"329.0\" GraphRef=\"c496e\" RelX=\"0.0\" RelY=\"1.0\" />\n      <Point X=\"759.3333333333333\" Y=\"378.0\" GraphRef=\"f54f5\" RelX=\"0.0\" RelY=\"-1.0\" ArrowHead=\"mim-binding\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"bf0c1\">\n    <Graphics ZOrder=\"12288\" LineThickness=\"1.0\">\n      <Point X=\"383.0\" Y=\"336.0\" GraphRef=\"ca3ad\" RelX=\"0.0\" RelY=\"1.0\" />\n      <Point X=\"351.3333333333333\" Y=\"380.3333333333333\" GraphRef=\"b1122\" RelX=\"0.0\" RelY=\"-1.0\" ArrowHead=\"mim-binding\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"c353b\">\n    <Graphics ZOrder=\"12288\" LineThickness=\"1.0\" Color=\"ff0000\">\n      <Point X=\"467.5\" Y=\"50.66666666666666\" GraphRef=\"d9620\" RelX=\"0.0\" RelY=\"1.0\" />\n      <Point X=\"468.0\" Y=\"243.66666666666663\" GraphRef=\"ef1f3\" RelX=\"0.0\" RelY=\"-1.0\" ArrowHead=\"mim-inhibition\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"c4468\">\n    <Graphics ZOrder=\"12288\" LineThickness=\"1.0\">\n      <Point X=\"383.0\" Y=\"336.0\" GraphRef=\"ca3ad\" RelX=\"0.0\" RelY=\"1.0\" />\n      <Point X=\"428.3333333333333\" Y=\"380.3333333333333\" GraphRef=\"b4811\" RelX=\"0.0\" RelY=\"-1.0\" ArrowHead=\"mim-binding\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"c6bfc\">\n    <Graphics ZOrder=\"12288\" LineStyle=\"Broken\" LineThickness=\"1.0\">\n      <Point X=\"744.3333333333333\" Y=\"398.0\" GraphRef=\"f54f5\" RelX=\"-0.5\" RelY=\"1.0\" />\n      <Point X=\"710.5\" Y=\"456.0\" GraphRef=\"e01c3\" RelX=\"0.5\" RelY=\"-1.0\" ArrowHead=\"Arrow\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"cc08a\">\n    <Graphics ZOrder=\"12288\" LineStyle=\"Broken\" LineThickness=\"1.0\">\n      <Point X=\"78.0\" Y=\"399.0\" GraphRef=\"ee619\" RelX=\"0.0\" RelY=\"1.0\" />\n      <Point X=\"88.25\" Y=\"460.0\" GraphRef=\"fd8ec\" RelX=\"-0.5\" RelY=\"-1.0\" ArrowHead=\"Arrow\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"d5638\">\n    <Graphics ZOrder=\"12288\" LineThickness=\"1.0\">\n      <Point X=\"274.0\" Y=\"282.66666666666663\" GraphRef=\"e08c8\" RelX=\"0.0\" RelY=\"-1.0\" />\n      <Point X=\"273.5520833333333\" Y=\"326.0\" GraphRef=\"b2c25\" RelX=\"0.0\" RelY=\"0.0\" ArrowHead=\"mim-catalysis\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"d74bc\">\n    <Graphics ZOrder=\"12288\" LineThickness=\"1.0\">\n      <Point X=\"402.0\" Y=\"211.0\" GraphRef=\"d6a08\" RelX=\"0.5\" RelY=\"1.0\" />\n      <Point X=\"403.0\" Y=\"316.0\" GraphRef=\"ca3ad\" RelX=\"0.5\" RelY=\"-1.0\" ArrowHead=\"mim-conversion\" />\n      <Anchor Position=\"0.4\" Shape=\"None\" GraphId=\"da262\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"d8bab\">\n    <Graphics ZOrder=\"12288\" LineThickness=\"1.0\">\n      <Point X=\"155.0\" Y=\"276.66666666666663\" GraphRef=\"d046f\" RelX=\"0.0\" RelY=\"-1.0\" />\n      <Point X=\"155.0\" Y=\"326.0\" GraphRef=\"a547f\" RelX=\"0.0\" RelY=\"0.0\" ArrowHead=\"mim-catalysis\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"d8d51\">\n    <Graphics ZOrder=\"12288\" LineThickness=\"1.0\">\n      <Point X=\"341.4145658263304\" Y=\"118.94957983193277\" GraphRef=\"e0a11\" RelX=\"1.0\" RelY=\"0.0\" />\n      <Point X=\"382.5089344209647\" Y=\"120.0036482754191\" GraphRef=\"ed881\" RelX=\"0.0\" RelY=\"0.0\" ArrowHead=\"mim-catalysis\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"dd7bd\">\n    <Graphics ZOrder=\"12288\" LineThickness=\"1.0\">\n      <Point X=\"204.5\" Y=\"211.0\" GraphRef=\"fbc8b\" RelX=\"0.0\" RelY=\"1.0\" />\n      <Point X=\"204.5\" Y=\"316.0\" GraphRef=\"df262\" RelX=\"0.0\" RelY=\"-1.0\" ArrowHead=\"mim-conversion\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"e0a2d\">\n    <Graphics ZOrder=\"12288\" LineStyle=\"Broken\" LineThickness=\"1.0\">\n      <Point X=\"428.3333333333333\" Y=\"400.3333333333333\" GraphRef=\"b4811\" RelX=\"0.0\" RelY=\"1.0\" />\n      <Point X=\"404.75\" Y=\"459.0\" GraphRef=\"b47da\" RelX=\"0.5\" RelY=\"-1.0\" ArrowHead=\"Arrow\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"e10e0\">\n    <Graphics ZOrder=\"12288\" LineThickness=\"1.0\">\n      <Point X=\"106.75\" Y=\"480.0\" GraphRef=\"fd8ec\" RelX=\"0.5\" RelY=\"1.0\" />\n      <Point X=\"163.0\" Y=\"481.0\" GraphRef=\"becdd\" RelX=\"-1.0\" RelY=\"0.0\" ArrowHead=\"mim-transcription-translation\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"e2599\">\n    <Graphics ZOrder=\"12288\" LineThickness=\"1.0\">\n      <Point X=\"404.75\" Y=\"479.0\" GraphRef=\"b47da\" RelX=\"0.5\" RelY=\"1.0\" />\n      <Point X=\"460.0\" Y=\"479.0\" GraphRef=\"e14ac\" RelX=\"-1.0\" RelY=\"0.0\" ArrowHead=\"mim-transcription-translation\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"e54f1\">\n    <BiopaxRef>bc9</BiopaxRef>\n    <Graphics ZOrder=\"12288\" LineThickness=\"1.0\">\n      <Point X=\"351.5\" Y=\"231.0\" GraphRef=\"c9ca8\" RelX=\"1.0\" RelY=\"0.0\" />\n      <Point X=\"382.37157627426086\" Y=\"250.015508797388\" GraphRef=\"deb70\" RelX=\"0.0\" RelY=\"0.0\" ArrowHead=\"mim-catalysis\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"eaace\">\n    <Graphics ZOrder=\"12288\" LineThickness=\"1.0\">\n      <Point X=\"464.75\" Y=\"529.75\" GraphRef=\"da9ee\" RelX=\"1.0\" RelY=\"0.0\" />\n      <Point X=\"523.0\" Y=\"530.0\" GraphRef=\"b75c3\" RelX=\"-1.0\" RelY=\"0.0\" ArrowHead=\"mim-conversion\" />\n      <Anchor Position=\"0.4678209615030391\" Shape=\"None\" GraphId=\"ff984\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"ec371\">\n    <Graphics ConnectorType=\"Elbow\" ZOrder=\"12288\" LineThickness=\"1.0\">\n      <Point X=\"296.74789915966375\" Y=\"118.94957983193277\" GraphRef=\"e0a11\" RelX=\"-1.0\" RelY=\"0.0\" />\n      <Point X=\"62.0\" Y=\"379.0\" GraphRef=\"ee619\" RelX=\"-0.5\" RelY=\"-1.0\" ArrowHead=\"Arrow\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"ee6b5\">\n    <Graphics ZOrder=\"12288\" LineStyle=\"Broken\" LineThickness=\"1.0\">\n      <Point X=\"774.3333333333333\" Y=\"398.0\" GraphRef=\"f54f5\" RelX=\"0.5\" RelY=\"1.0\" />\n      <Point X=\"797.5\" Y=\"465.0\" GraphRef=\"b2aa3\" RelX=\"-0.5\" RelY=\"-1.0\" ArrowHead=\"Arrow\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"ef46a\">\n    <Graphics ZOrder=\"12288\" LineThickness=\"1.0\">\n      <Point X=\"204.5\" Y=\"336.0\" GraphRef=\"df262\" RelX=\"0.0\" RelY=\"1.0\" />\n      <Point X=\"204.0\" Y=\"379.0\" GraphRef=\"e3012\" RelX=\"0.0\" RelY=\"-1.0\" ArrowHead=\"mim-binding\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"f0f9a\">\n    <Graphics ZOrder=\"12288\" LineThickness=\"1.0\">\n      <Point X=\"282.5\" Y=\"178.0\" GraphRef=\"e68e0\" RelX=\"0.0\" RelY=\"1.0\" />\n      <Point X=\"281.87368421052633\" Y=\"201.0\" GraphRef=\"da414\" RelX=\"0.0\" RelY=\"0.0\" ArrowHead=\"mim-catalysis\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"f3a74\">\n    <Graphics ZOrder=\"12288\" LineThickness=\"1.0\">\n      <Point X=\"382.0\" Y=\"211.0\" GraphRef=\"d6a08\" RelX=\"0.0\" RelY=\"1.0\" />\n      <Point X=\"383.0\" Y=\"316.0\" GraphRef=\"ca3ad\" RelX=\"0.0\" RelY=\"-1.0\" ArrowHead=\"mim-conversion\" />\n      <Anchor Position=\"0.37157627426083817\" Shape=\"None\" GraphId=\"deb70\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"f3f2e\">\n    <Graphics ConnectorType=\"Elbow\" ZOrder=\"12288\" LineThickness=\"1.0\" Color=\"ff0000\">\n      <Point X=\"513.0\" Y=\"40.66666666666666\" GraphRef=\"d9620\" RelX=\"1.0\" RelY=\"0.0\" />\n      <Point X=\"652.3333333333333\" Y=\"273.66666666666663\" GraphRef=\"d6f1e\" RelX=\"0.0\" RelY=\"-1.0\" ArrowHead=\"mim-inhibition\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"f744e\">\n    <Graphics ZOrder=\"12288\" LineThickness=\"1.0\">\n      <Point X=\"682.3333333333333\" Y=\"283.66666666666663\" GraphRef=\"d6f1e\" RelX=\"1.0\" RelY=\"0.0\" />\n      <Point X=\"717.3301602262018\" Y=\"283.41258246936854\" GraphRef=\"c11bc\" RelX=\"0.0\" RelY=\"0.0\" ArrowHead=\"mim-catalysis\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"f75ef\">\n    <Graphics ZOrder=\"12288\" LineStyle=\"Broken\" LineThickness=\"1.0\">\n      <Point X=\"371.0\" Y=\"435.0\" GraphRef=\"d519e\" RelX=\"0.0\" RelY=\"1.0\" />\n      <Point X=\"386.25\" Y=\"459.0\" GraphRef=\"b47da\" RelX=\"-0.5\" RelY=\"-1.0\" ArrowHead=\"Arrow\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"f8c76\">\n    <Graphics ZOrder=\"12288\" LineThickness=\"1.0\">\n      <Point X=\"179.0\" Y=\"326.0\" GraphRef=\"df262\" RelX=\"-1.0\" RelY=\"0.0\" />\n      <Point X=\"129.0\" Y=\"326.0\" GraphRef=\"b298c\" RelX=\"1.0\" RelY=\"0.0\" ArrowHead=\"mim-conversion\" />\n      <Anchor Position=\"0.48\" Shape=\"None\" GraphId=\"a547f\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"d4ea3\">\n    <Graphics ZOrder=\"32769\" LineThickness=\"1.0\">\n      <Point X=\"343.0\" Y=\"326.0\" GraphRef=\"ca3ad\" RelX=\"-1.0\" RelY=\"0.0\" />\n      <Point X=\"230.0\" Y=\"326.0\" GraphRef=\"df262\" RelX=\"1.0\" RelY=\"0.0\" ArrowHead=\"mim-conversion\" />\n      <Anchor Position=\"0.6145833333333334\" Shape=\"None\" GraphId=\"b2c25\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"d5bc3\">\n    <Graphics ConnectorType=\"Elbow\" ZOrder=\"32771\" LineThickness=\"1.0\" Color=\"ff0000\">\n      <Point X=\"422.0\" Y=\"40.66666666666666\" GraphRef=\"d9620\" RelX=\"-1.0\" RelY=\"0.0\" />\n      <Point X=\"155.0\" Y=\"276.66666666666663\" GraphRef=\"d046f\" RelX=\"0.0\" RelY=\"-1.0\" ArrowHead=\"mim-inhibition\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Interaction GraphId=\"a6f8b\">\n    <Graphics ZOrder=\"32772\" LineThickness=\"1.0\">\n      <Point X=\"717.0\" Y=\"329.0\" GraphRef=\"c496e\" RelX=\"0.0\" RelY=\"1.0\" />\n      <Point X=\"681.5\" Y=\"378.0\" GraphRef=\"b4349\" RelX=\"0.0\" RelY=\"-1.0\" ArrowHead=\"mim-binding\" />\n    </Graphics>\n    <Xref Database=\"\" ID=\"\" />\n  </Interaction>\n  <Label TextLabel=\"Zona glomerulosa &#xA;of the adrenal gland\" GraphId=\"a49c5\">\n    <Graphics CenterX=\"400.0\" CenterY=\"585.5\" Width=\"184.0\" Height=\"35.0\" ZOrder=\"28672\" FillColor=\"ffffff\" FontWeight=\"Bold\" FontSize=\"14\" Valign=\"Middle\" Color=\"999999\" />\n  </Label>\n  <Label TextLabel=\"EDHF\" GraphId=\"b2aa3\">\n    <Graphics CenterX=\"808.0\" CenterY=\"475.0\" Width=\"42.0\" Height=\"20.0\" ZOrder=\"28672\" FillColor=\"ffffff\" FontWeight=\"Bold\" FontSize=\"10\" Valign=\"Middle\" />\n  </Label>\n  <Label TextLabel=\"Blood vessels\" GraphId=\"d3e2b\">\n    <Graphics CenterX=\"749.0\" CenterY=\"579.5\" Width=\"184.0\" Height=\"35.0\" ZOrder=\"28672\" FillColor=\"ffffff\" FontWeight=\"Bold\" FontSize=\"14\" Valign=\"Middle\" Color=\"999999\" />\n  </Label>\n  <Label TextLabel=\"Mesangial cells around&#xA;blood vessels in the kidneys\" GraphId=\"e166c\">\n    <Graphics CenterX=\"139.0\" CenterY=\"583.5\" Width=\"222.0\" Height=\"35.0\" ZOrder=\"28672\" FillColor=\"ffffff\" FontWeight=\"Bold\" FontSize=\"14\" Valign=\"Middle\" Color=\"999999\" />\n  </Label>\n  <Shape GraphId=\"c2049\">\n    <Attribute Key=\"org.pathvisio.DoubleLineProperty\" Value=\"Double\" />\n    <Graphics CenterX=\"748.0\" CenterY=\"502.5\" Width=\"232.0\" Height=\"235.0\" ZOrder=\"16384\" FontSize=\"10\" Valign=\"Middle\" ShapeType=\"RoundedRectangle\" LineThickness=\"3.0\" Color=\"c0c0c0\" Rotation=\"0.0\" />\n  </Shape>\n  <Shape GraphId=\"d5695\">\n    <Attribute Key=\"org.pathvisio.DoubleLineProperty\" Value=\"Double\" />\n    <Graphics CenterX=\"450.0\" CenterY=\"503.5\" Width=\"344.0\" Height=\"235.0\" ZOrder=\"16384\" FontSize=\"10\" Valign=\"Middle\" ShapeType=\"RoundedRectangle\" LineThickness=\"3.0\" Color=\"c0c0c0\" Rotation=\"0.0\" />\n  </Shape>\n  <Shape GraphId=\"f975e\">\n    <Attribute Key=\"org.pathvisio.DoubleLineProperty\" Value=\"Double\" />\n    <Graphics CenterX=\"140.0\" CenterY=\"504.0\" Width=\"254.0\" Height=\"232.0\" ZOrder=\"16384\" FontSize=\"10\" Valign=\"Middle\" ShapeType=\"RoundedRectangle\" LineThickness=\"3.0\" Color=\"c0c0c0\" Rotation=\"0.0\" />\n  </Shape>\n  <InfoBox CenterX=\"1.6666666666666667\" CenterY=\"38.333333333333336\" />\n  <Biopax>\n    <bp:PublicationXref xmlns:bp=\"http://www.biopax.org/release/biopax-level3.owl#\" xmlns:rdf=\"http://www.w3.org/1999/02/22-rdf-syntax-ns#\" rdf:id=\"c16\">\n      <bp:ID rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">18417113</bp:ID>\n      <bp:DB rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">PubMed</bp:DB>\n      <bp:TITLE rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">The (pro)renin receptor: a new addition to the renin-angiotensin system?</bp:TITLE>\n      <bp:SOURCE rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Eur J Pharmacol</bp:SOURCE>\n      <bp:YEAR rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">2008</bp:YEAR>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Batenburg WW</bp:AUTHORS>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Jan Danser AH</bp:AUTHORS>\n    </bp:PublicationXref>\n    <bp:PublicationXref xmlns:bp=\"http://www.biopax.org/release/biopax-level3.owl#\" xmlns:rdf=\"http://www.w3.org/1999/02/22-rdf-syntax-ns#\" rdf:id=\"bc9\">\n      <bp:ID rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">18035185</bp:ID>\n      <bp:DB rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">PubMed</bp:DB>\n      <bp:TITLE rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Effects of renin-angiotensin system inhibition on end-organ protection: can we do better?</bp:TITLE>\n      <bp:SOURCE rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Clin Ther</bp:SOURCE>\n      <bp:YEAR rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">2007</bp:YEAR>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Weir MR</bp:AUTHORS>\n    </bp:PublicationXref>\n    <bp:PublicationXref xmlns:bp=\"http://www.biopax.org/release/biopax-level3.owl#\" xmlns:rdf=\"http://www.w3.org/1999/02/22-rdf-syntax-ns#\" rdf:id=\"ce3\">\n      <bp:ID rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">15134803</bp:ID>\n      <bp:DB rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">PubMed</bp:DB>\n      <bp:TITLE rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">The regulation of aldosterone synthase expression.</bp:TITLE>\n      <bp:SOURCE rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Mol Cell Endocrinol</bp:SOURCE>\n      <bp:YEAR rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">2004</bp:YEAR>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Bassett MH</bp:AUTHORS>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">White PC</bp:AUTHORS>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Rainey WE</bp:AUTHORS>\n    </bp:PublicationXref>\n    <bp:openControlledVocabulary xmlns:bp=\"http://www.biopax.org/release/biopax-level3.owl#\">\n      <bp:TERM xmlns:rdf=\"http://www.w3.org/1999/02/22-rdf-syntax-ns#\" rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">ACE inhibitor drug pathway</bp:TERM>\n      <bp:ID xmlns:rdf=\"http://www.w3.org/1999/02/22-rdf-syntax-ns#\" rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">PW:0001228</bp:ID>\n      <bp:Ontology xmlns:rdf=\"http://www.w3.org/1999/02/22-rdf-syntax-ns#\" rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Pathway Ontology</bp:Ontology>\n    </bp:openControlledVocabulary>\n    <bp:PublicationXref xmlns:bp=\"http://www.biopax.org/release/biopax-level3.owl#\" xmlns:rdf=\"http://www.w3.org/1999/02/22-rdf-syntax-ns#\" rdf:id=\"b93\">\n      <bp:ID rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">17464936</bp:ID>\n      <bp:DB rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">PubMed</bp:DB>\n      <bp:TITLE rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">The emerging role of ACE2 in physiology and disease.</bp:TITLE>\n      <bp:SOURCE rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">J Pathol</bp:SOURCE>\n      <bp:YEAR rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">2007</bp:YEAR>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Hamming I</bp:AUTHORS>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Cooper ME</bp:AUTHORS>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Haagmans BL</bp:AUTHORS>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Hooper NM</bp:AUTHORS>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Korstanje R</bp:AUTHORS>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Osterhaus AD</bp:AUTHORS>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Timens W</bp:AUTHORS>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Turner AJ</bp:AUTHORS>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Navis G</bp:AUTHORS>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">van Goor H</bp:AUTHORS>\n    </bp:PublicationXref>\n    <bp:PublicationXref xmlns:bp=\"http://www.biopax.org/release/biopax-level3.owl#\" xmlns:rdf=\"http://www.w3.org/1999/02/22-rdf-syntax-ns#\" rdf:id=\"bd1\">\n      <bp:ID rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">15174896</bp:ID>\n      <bp:DB rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">PubMed</bp:DB>\n      <bp:TITLE rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Pharmacogenetics of antihypertensive drug responses.</bp:TITLE>\n      <bp:SOURCE rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Am J Pharmacogenomics</bp:SOURCE>\n      <bp:YEAR rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">2004</bp:YEAR>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Schwartz GL</bp:AUTHORS>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Turner ST</bp:AUTHORS>\n    </bp:PublicationXref>\n    <bp:PublicationXref xmlns:bp=\"http://www.biopax.org/release/biopax-level3.owl#\" xmlns:rdf=\"http://www.w3.org/1999/02/22-rdf-syntax-ns#\" rdf:id=\"eeb\">\n      <bp:ID rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">16679330</bp:ID>\n      <bp:DB rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">PubMed</bp:DB>\n      <bp:TITLE rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Systematic review and meta-analysis of ethnic differences in risks of adverse reactions to drugs used in cardiovascular medicine.</bp:TITLE>\n      <bp:SOURCE rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">BMJ</bp:SOURCE>\n      <bp:YEAR rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">2006</bp:YEAR>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">McDowell SE</bp:AUTHORS>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Coleman JJ</bp:AUTHORS>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Ferner RE</bp:AUTHORS>\n    </bp:PublicationXref>\n    <bp:PublicationXref xmlns:bp=\"http://www.biopax.org/release/biopax-level3.owl#\" xmlns:rdf=\"http://www.w3.org/1999/02/22-rdf-syntax-ns#\" rdf:id=\"c74\">\n      <bp:ID rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">18449520</bp:ID>\n      <bp:DB rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">PubMed</bp:DB>\n      <bp:TITLE rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Are we poised to target ACE2 for the next generation of antihypertensives?</bp:TITLE>\n      <bp:SOURCE rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">J Mol Med</bp:SOURCE>\n      <bp:YEAR rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">2008</bp:YEAR>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Ferreira AJ</bp:AUTHORS>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Raizada MK</bp:AUTHORS>\n    </bp:PublicationXref>\n    <bp:PublicationXref xmlns:bp=\"http://www.biopax.org/release/biopax-level3.owl#\" xmlns:rdf=\"http://www.w3.org/1999/02/22-rdf-syntax-ns#\" rdf:id=\"af8\">\n      <bp:ID rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">16374430</bp:ID>\n      <bp:DB rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">PubMed</bp:DB>\n      <bp:TITLE rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Renin increases mesangial cell transforming growth factor-beta1 and matrix proteins through receptor-mediated, angiotensin II-independent mechanisms.</bp:TITLE>\n      <bp:SOURCE rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Kidney Int</bp:SOURCE>\n      <bp:YEAR rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">2006</bp:YEAR>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Huang Y</bp:AUTHORS>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Wongamorntham S</bp:AUTHORS>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Kasting J</bp:AUTHORS>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">McQuillan D</bp:AUTHORS>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Owens RT</bp:AUTHORS>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Yu L</bp:AUTHORS>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Noble NA</bp:AUTHORS>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Border W</bp:AUTHORS>\n    </bp:PublicationXref>\n    <bp:PublicationXref xmlns:bp=\"http://www.biopax.org/release/biopax-level3.owl#\" xmlns:rdf=\"http://www.w3.org/1999/02/22-rdf-syntax-ns#\" rdf:id=\"b40\">\n      <bp:ID rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">17376010</bp:ID>\n      <bp:DB rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">PubMed</bp:DB>\n      <bp:TITLE rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Angiotensin II reactivation and aldosterone escape phenomena in renin-angiotensin-aldosterone system blockade: is oral renin inhibition the solution?</bp:TITLE>\n      <bp:SOURCE rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Expert Opin Pharmacother</bp:SOURCE>\n      <bp:YEAR rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">2007</bp:YEAR>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Athyros VG</bp:AUTHORS>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Mikhailidis DP</bp:AUTHORS>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Kakafika AI</bp:AUTHORS>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Tziomalos K</bp:AUTHORS>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Karagiannis A</bp:AUTHORS>\n    </bp:PublicationXref>\n    <bp:openControlledVocabulary xmlns:bp=\"http://www.biopax.org/release/biopax-level3.owl#\">\n      <bp:TERM xmlns:rdf=\"http://www.w3.org/1999/02/22-rdf-syntax-ns#\" rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">hypertension</bp:TERM>\n      <bp:ID xmlns:rdf=\"http://www.w3.org/1999/02/22-rdf-syntax-ns#\" rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">DOID:10763</bp:ID>\n      <bp:Ontology xmlns:rdf=\"http://www.w3.org/1999/02/22-rdf-syntax-ns#\" rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Disease</bp:Ontology>\n    </bp:openControlledVocabulary>\n    <bp:PublicationXref xmlns:bp=\"http://www.biopax.org/release/biopax-level3.owl#\" xmlns:rdf=\"http://www.w3.org/1999/02/22-rdf-syntax-ns#\" rdf:id=\"b5d\">\n      <bp:ID rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">18729003</bp:ID>\n      <bp:DB rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">PubMed</bp:DB>\n      <bp:TITLE rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Mineralocorticoid receptor antagonists and endothelial function.</bp:TITLE>\n      <bp:SOURCE rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Curr Opin Investig Drugs</bp:SOURCE>\n      <bp:YEAR rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">2008</bp:YEAR>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Maron BA</bp:AUTHORS>\n      <bp:AUTHORS rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Leopold JA</bp:AUTHORS>\n    </bp:PublicationXref>\n    <bp:openControlledVocabulary xmlns:bp=\"http://www.biopax.org/release/biopax-level3.owl#\">\n      <bp:TERM xmlns:rdf=\"http://www.w3.org/1999/02/22-rdf-syntax-ns#\" rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">angiotensin signaling pathway</bp:TERM>\n      <bp:ID xmlns:rdf=\"http://www.w3.org/1999/02/22-rdf-syntax-ns#\" rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">PW:0000245</bp:ID>\n      <bp:Ontology xmlns:rdf=\"http://www.w3.org/1999/02/22-rdf-syntax-ns#\" rdf:datatype=\"http://www.w3.org/2001/XMLSchema#string\">Pathway Ontology</bp:Ontology>\n    </bp:openControlledVocabulary>\n  </Biopax>\n</Pathway>"
#>