Get Recent Changes

Retrieve recent changes to pathways at WikiPathways.

getRecentChanges(timestamp)

Arguments

timestamp
(14 digits, YYYYMMDDhhmmss) Limit by time, only history items after the given time, e.g., 20180201000000 for changes since Feb 1st, 2018.

Value

A list of changes, including pathway WPID, name, url, species and revision

Examples

{
getRecentChanges('20180201000000')
}

#> [[1]]
#> [[1]]$id
#> [1] "WP4829"
#> 
#> [[1]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4829"
#> 
#> [[1]]$name
#> [1] "mBDNF and proBDNF regulation of GABA neurotransmission"
#> 
#> [[1]]$species
#> [1] "Homo sapiens"
#> 
#> [[1]]$revision
#> [1] "109219"
#> 
#> 
#> [[2]]
#> [[2]]$id
#> [1] "WP4228"
#> 
#> [[2]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4228"
#> 
#> [[2]]$name
#> [1] "Vitamin B6-dependent and responsive disorders"
#> 
#> [[2]]$species
#> [1] "Homo sapiens"
#> 
#> [[2]]$revision
#> [1] "109214"
#> 
#> 
#> [[3]]
#> [[3]]$id
#> [1] "WP4535"
#> 
#> [[3]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4535"
#> 
#> [[3]]$name
#> [1] "Envelope proteins and their potential roles in EDMD physiopathology"
#> 
#> [[3]]$species
#> [1] "Homo sapiens"
#> 
#> [[3]]$revision
#> [1] "109210"
#> 
#> 
#> [[4]]
#> [[4]]$id
#> [1] "WP4298"
#> 
#> [[4]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4298"
#> 
#> [[4]]$name
#> [1] "Viral Acute Myocarditis"
#> 
#> [[4]]$species
#> [1] "Homo sapiens"
#> 
#> [[4]]$revision
#> [1] "109212"
#> 
#> 
#> [[5]]
#> [[5]]$id
#> [1] "WP4577"
#> 
#> [[5]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4577"
#> 
#> [[5]]$name
#> [1] "Neurodegeneration with brain iron accumulation (NBIA) subtypes pathway"
#> 
#> [[5]]$species
#> [1] "Homo sapiens"
#> 
#> [[5]]$revision
#> [1] "109208"
#> 
#> 
#> [[6]]
#> [[6]]$id
#> [1] "WP2447"
#> 
#> [[6]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP2447"
#> 
#> [[6]]$name
#> [1] "Amyotrophic lateral sclerosis (ALS)"
#> 
#> [[6]]$species
#> [1] "Homo sapiens"
#> 
#> [[6]]$revision
#> [1] "109207"
#> 
#> 
#> [[7]]
#> [[7]]$id
#> [1] "WP4828"
#> 
#> [[7]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4828"
#> 
#> [[7]]$name
#> [1] "Mammalian Sexual Development"
#> 
#> [[7]]$species
#> [1] "Homo sapiens"
#> 
#> [[7]]$revision
#> [1] "109202"
#> 
#> 
#> [[8]]
#> [[8]]$id
#> [1] "WP4827"
#> 
#> [[8]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4827"
#> 
#> [[8]]$name
#> [1] "inflammation and cancer"
#> 
#> [[8]]$species
#> [1] "Anopheles gambiae"
#> 
#> [[8]]$revision
#> [1] "109192"
#> 
#> 
#> [[9]]
#> [[9]]$id
#> [1] "WP4826"
#> 
#> [[9]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4826"
#> 
#> [[9]]$name
#> [1] "inflammation leading to cancer"
#> 
#> [[9]]$species
#> [1] "Homo sapiens"
#> 
#> [[9]]$revision
#> [1] "109191"
#> 
#> 
#> [[10]]
#> [[10]]$id
#> [1] "WP4798"
#> 
#> [[10]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4798"
#> 
#> [[10]]$name
#> [1] ""
#> 
#> [[10]]$species
#> [1] ""
#> 
#> [[10]]$revision
#> [1] 0
#> 
#> 
#> [[11]]
#> [[11]]$id
#> [1] "WP4781"
#> 
#> [[11]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4781"
#> 
#> [[11]]$name
#> [1] ""
#> 
#> [[11]]$species
#> [1] ""
#> 
#> [[11]]$revision
#> [1] 0
#> 
#> 
#> [[12]]
#> [[12]]$id
#> [1] "WP4782"
#> 
#> [[12]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4782"
#> 
#> [[12]]$name
#> [1] ""
#> 
#> [[12]]$species
#> [1] ""
#> 
#> [[12]]$revision
#> [1] 0
#> 
#> 
#> [[13]]
#> [[13]]$id
#> [1] "WP4791"
#> 
#> [[13]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4791"
#> 
#> [[13]]$name
#> [1] ""
#> 
#> [[13]]$species
#> [1] ""
#> 
#> [[13]]$revision
#> [1] 0
#> 
#> 
#> [[14]]
#> [[14]]$id
#> [1] "WP4783"
#> 
#> [[14]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4783"
#> 
#> [[14]]$name
#> [1] ""
#> 
#> [[14]]$species
#> [1] ""
#> 
#> [[14]]$revision
#> [1] 0
#> 
#> 
#> [[15]]
#> [[15]]$id
#> [1] "WP4800"
#> 
#> [[15]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4800"
#> 
#> [[15]]$name
#> [1] ""
#> 
#> [[15]]$species
#> [1] ""
#> 
#> [[15]]$revision
#> [1] 0
#> 
#> 
#> [[16]]
#> [[16]]$id
#> [1] "WP4807"
#> 
#> [[16]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4807"
#> 
#> [[16]]$name
#> [1] ""
#> 
#> [[16]]$species
#> [1] ""
#> 
#> [[16]]$revision
#> [1] 0
#> 
#> 
#> [[17]]
#> [[17]]$id
#> [1] "WP4817"
#> 
#> [[17]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4817"
#> 
#> [[17]]$name
#> [1] ""
#> 
#> [[17]]$species
#> [1] ""
#> 
#> [[17]]$revision
#> [1] 0
#> 
#> 
#> [[18]]
#> [[18]]$id
#> [1] "WP4819"
#> 
#> [[18]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4819"
#> 
#> [[18]]$name
#> [1] ""
#> 
#> [[18]]$species
#> [1] ""
#> 
#> [[18]]$revision
#> [1] 0
#> 
#> 
#> [[19]]
#> [[19]]$id
#> [1] "WP4821"
#> 
#> [[19]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4821"
#> 
#> [[19]]$name
#> [1] ""
#> 
#> [[19]]$species
#> [1] ""
#> 
#> [[19]]$revision
#> [1] 0
#> 
#> 
#> [[20]]
#> [[20]]$id
#> [1] "WP4822"
#> 
#> [[20]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4822"
#> 
#> [[20]]$name
#> [1] ""
#> 
#> [[20]]$species
#> [1] ""
#> 
#> [[20]]$revision
#> [1] 0
#> 
#> 
#> [[21]]
#> [[21]]$id
#> [1] "WP4741"
#> 
#> [[21]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4741"
#> 
#> [[21]]$name
#> [1] ""
#> 
#> [[21]]$species
#> [1] ""
#> 
#> [[21]]$revision
#> [1] 0
#> 
#> 
#> [[22]]
#> [[22]]$id
#> [1] "WP4775"
#> 
#> [[22]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4775"
#> 
#> [[22]]$name
#> [1] ""
#> 
#> [[22]]$species
#> [1] ""
#> 
#> [[22]]$revision
#> [1] 0
#> 
#> 
#> [[23]]
#> [[23]]$id
#> [1] "WP4774"
#> 
#> [[23]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4774"
#> 
#> [[23]]$name
#> [1] ""
#> 
#> [[23]]$species
#> [1] ""
#> 
#> [[23]]$revision
#> [1] 0
#> 
#> 
#> [[24]]
#> [[24]]$id
#> [1] "WP4825"
#> 
#> [[24]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4825"
#> 
#> [[24]]$name
#> [1] "MTOR 1"
#> 
#> [[24]]$species
#> [1] "Homo sapiens"
#> 
#> [[24]]$revision
#> [1] "109187"
#> 
#> 
#> [[25]]
#> [[25]]$id
#> [1] "WP4824"
#> 
#> [[25]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4824"
#> 
#> [[25]]$name
#> [1] "Regucalcin in proximal tubule kidney cells"
#> 
#> [[25]]$species
#> [1] "Rattus norvegicus"
#> 
#> [[25]]$revision
#> [1] "109158"
#> 
#> 
#> [[26]]
#> [[26]]$id
#> [1] "WP4823"
#> 
#> [[26]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4823"
#> 
#> [[26]]$name
#> [1] "Genes controlling renal nephrogenesis"
#> 
#> [[26]]$species
#> [1] "Homo sapiens"
#> 
#> [[26]]$revision
#> [1] "109163"
#> 
#> 
#> [[27]]
#> [[27]]$id
#> [1] "WP4820"
#> 
#> [[27]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4820"
#> 
#> [[27]]$name
#> [1] "GDNF/RET signalling axis"
#> 
#> [[27]]$species
#> [1] "Mus musculus"
#> 
#> [[27]]$revision
#> [1] "109145"
#> 
#> 
#> [[28]]
#> [[28]]$id
#> [1] "WP4818"
#> 
#> [[28]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4818"
#> 
#> [[28]]$name
#> [1] "Conversion of Angiotensinogen to Angiotensin II "
#> 
#> [[28]]$species
#> [1] "Homo sapiens"
#> 
#> [[28]]$revision
#> [1] "109140"
#> 
#> 
#> [[29]]
#> [[29]]$id
#> [1] "WP4816"
#> 
#> [[29]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4816"
#> 
#> [[29]]$name
#> [1] "TGF-beta Receptor Signaling"
#> 
#> [[29]]$species
#> [1] "Homo sapiens"
#> 
#> [[29]]$revision
#> [1] "109129"
#> 
#> 
#> [[30]]
#> [[30]]$id
#> [1] "WP4815"
#> 
#> [[30]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4815"
#> 
#> [[30]]$name
#> [1] "Glycosaminoglycan degradation"
#> 
#> [[30]]$species
#> [1] "Homo sapiens"
#> 
#> [[30]]$revision
#> [1] "109128"
#> 
#> 
#> [[31]]
#> [[31]]$id
#> [1] "WP4814"
#> 
#> [[31]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4814"
#> 
#> [[31]]$name
#> [1] "Somatic Sex determination"
#> 
#> [[31]]$species
#> [1] "Homo sapiens"
#> 
#> [[31]]$revision
#> [1] "109119"
#> 
#> 
#> [[32]]
#> [[32]]$id
#> [1] "WP4813"
#> 
#> [[32]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4813"
#> 
#> [[32]]$name
#> [1] "Regucalcin in kidney cells"
#> 
#> [[32]]$species
#> [1] "Homo sapiens"
#> 
#> [[32]]$revision
#> [1] "109117"
#> 
#> 
#> [[33]]
#> [[33]]$id
#> [1] "WP4812"
#> 
#> [[33]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4812"
#> 
#> [[33]]$name
#> [1] "practice"
#> 
#> [[33]]$species
#> [1] "Homo sapiens"
#> 
#> [[33]]$revision
#> [1] "109118"
#> 
#> 
#> [[34]]
#> [[34]]$id
#> [1] "WP4811"
#> 
#> [[34]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4811"
#> 
#> [[34]]$name
#> [1] "Kisspeptin/Kisspeptin Receptor System in the Ovary"
#> 
#> [[34]]$species
#> [1] "Homo sapiens"
#> 
#> [[34]]$revision
#> [1] "109107"
#> 
#> 
#> [[35]]
#> [[35]]$id
#> [1] "WP4810"
#> 
#> [[35]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4810"
#> 
#> [[35]]$name
#> [1] "test4"
#> 
#> [[35]]$species
#> [1] "Homo sapiens"
#> 
#> [[35]]$revision
#> [1] "109099"
#> 
#> 
#> [[36]]
#> [[36]]$id
#> [1] "WP4809"
#> 
#> [[36]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4809"
#> 
#> [[36]]$name
#> [1] "Academy"
#> 
#> [[36]]$species
#> [1] "Homo sapiens"
#> 
#> [[36]]$revision
#> [1] "109096"
#> 
#> 
#> [[37]]
#> [[37]]$id
#> [1] "WP4808"
#> 
#> [[37]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4808"
#> 
#> [[37]]$name
#> [1] "Endochondral Ossification with Skeletal Dysplasias"
#> 
#> [[37]]$species
#> [1] "Homo sapiens"
#> 
#> [[37]]$revision
#> [1] "109123"
#> 
#> 
#> [[38]]
#> [[38]]$id
#> [1] "WP4768"
#> 
#> [[38]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4768"
#> 
#> [[38]]$name
#> [1] ""
#> 
#> [[38]]$species
#> [1] ""
#> 
#> [[38]]$revision
#> [1] 0
#> 
#> 
#> [[39]]
#> [[39]]$id
#> [1] "WP4769"
#> 
#> [[39]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4769"
#> 
#> [[39]]$name
#> [1] ""
#> 
#> [[39]]$species
#> [1] ""
#> 
#> [[39]]$revision
#> [1] 0
#> 
#> 
#> [[40]]
#> [[40]]$id
#> [1] "WP4806"
#> 
#> [[40]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4806"
#> 
#> [[40]]$name
#> [1] "EGFR Tyrosine Kinase Inhibitor Resistance"
#> 
#> [[40]]$species
#> [1] "Homo sapiens"
#> 
#> [[40]]$revision
#> [1] "109041"
#> 
#> 
#> [[41]]
#> [[41]]$id
#> [1] "WP4318"
#> 
#> [[41]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4318"
#> 
#> [[41]]$name
#> [1] "Mitochondrial fatty acid synthesis and respiration"
#> 
#> [[41]]$species
#> [1] "Homo sapiens"
#> 
#> [[41]]$revision
#> [1] "109037"
#> 
#> 
#> [[42]]
#> [[42]]$id
#> [1] "WP4805"
#> 
#> [[42]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4805"
#> 
#> [[42]]$name
#> [1] "mypath"
#> 
#> [[42]]$species
#> [1] "Bos taurus"
#> 
#> [[42]]$revision
#> [1] "109029"
#> 
#> 
#> [[43]]
#> [[43]]$id
#> [1] "WP4804"
#> 
#> [[43]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4804"
#> 
#> [[43]]$name
#> [1] "Cholesterol Biosynthesis with Skeletal Dysplasias"
#> 
#> [[43]]$species
#> [1] "Homo sapiens"
#> 
#> [[43]]$revision
#> [1] "109130"
#> 
#> 
#> [[44]]
#> [[44]]$id
#> [1] "WP4803"
#> 
#> [[44]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4803"
#> 
#> [[44]]$name
#> [1] "Ciliopathies"
#> 
#> [[44]]$species
#> [1] "Homo sapiens"
#> 
#> [[44]]$revision
#> [1] "109093"
#> 
#> 
#> [[45]]
#> [[45]]$id
#> [1] "WP4779"
#> 
#> [[45]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4779"
#> 
#> [[45]]$name
#> [1] ""
#> 
#> [[45]]$species
#> [1] ""
#> 
#> [[45]]$revision
#> [1] 0
#> 
#> 
#> [[46]]
#> [[46]]$id
#> [1] "WP4778"
#> 
#> [[46]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4778"
#> 
#> [[46]]$name
#> [1] ""
#> 
#> [[46]]$species
#> [1] ""
#> 
#> [[46]]$revision
#> [1] 0
#> 
#> 
#> [[47]]
#> [[47]]$id
#> [1] "WP4780"
#> 
#> [[47]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4780"
#> 
#> [[47]]$name
#> [1] ""
#> 
#> [[47]]$species
#> [1] ""
#> 
#> [[47]]$revision
#> [1] 0
#> 
#> 
#> [[48]]
#> [[48]]$id
#> [1] "WP4777"
#> 
#> [[48]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4777"
#> 
#> [[48]]$name
#> [1] ""
#> 
#> [[48]]$species
#> [1] ""
#> 
#> [[48]]$revision
#> [1] 0
#> 
#> 
#> [[49]]
#> [[49]]$id
#> [1] "WP4776"
#> 
#> [[49]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4776"
#> 
#> [[49]]$name
#> [1] ""
#> 
#> [[49]]$species
#> [1] ""
#> 
#> [[49]]$revision
#> [1] 0
#> 
#> 
#> [[50]]
#> [[50]]$id
#> [1] "WP4801"
#> 
#> [[50]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4801"
#> 
#> [[50]]$name
#> [1] "Test run of glucose metabolism "
#> 
#> [[50]]$species
#> [1] "Homo sapiens"
#> 
#> [[50]]$revision
#> [1] "109002"
#> 
#> 
#> [[51]]
#> [[51]]$id
#> [1] "WP4766"
#> 
#> [[51]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4766"
#> 
#> [[51]]$name
#> [1] ""
#> 
#> [[51]]$species
#> [1] ""
#> 
#> [[51]]$revision
#> [1] 0
#> 
#> 
#> [[52]]
#> [[52]]$id
#> [1] "WP4799"
#> 
#> [[52]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4799"
#> 
#> [[52]]$name
#> [1] "Downregulation of ACE2 expression by SARS-CoV infection and SARS-CoV Spike protein."
#> 
#> [[52]]$species
#> [1] "Homo sapiens"
#> 
#> [[52]]$revision
#> [1] "109161"
#> 
#> 
#> [[53]]
#> [[53]]$id
#> [1] "WP4797"
#> 
#> [[53]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4797"
#> 
#> [[53]]$name
#> [1] "WikiPahtway introduction"
#> 
#> [[53]]$species
#> [1] "Homo sapiens"
#> 
#> [[53]]$revision
#> [1] "108968"
#> 
#> 
#> [[54]]
#> [[54]]$id
#> [1] "WP4796"
#> 
#> [[54]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4796"
#> 
#> [[54]]$name
#> [1] "test2"
#> 
#> [[54]]$species
#> [1] "Homo sapiens"
#> 
#> [[54]]$revision
#> [1] "109095"
#> 
#> 
#> [[55]]
#> [[55]]$id
#> [1] "WP3604"
#> 
#> [[55]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3604"
#> 
#> [[55]]$name
#> [1] "Biochemical Pathways Part I"
#> 
#> [[55]]$species
#> [1] "Homo sapiens"
#> 
#> [[55]]$revision
#> [1] "109082"
#> 
#> 
#> [[56]]
#> [[56]]$id
#> [1] "WP4795"
#> 
#> [[56]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4795"
#> 
#> [[56]]$name
#> [1] "Target genes for floral homeotic tetrameric MADS complexes"
#> 
#> [[56]]$species
#> [1] "Arabidopsis thaliana"
#> 
#> [[56]]$revision
#> [1] "108950"
#> 
#> 
#> [[57]]
#> [[57]]$id
#> [1] "WP4794"
#> 
#> [[57]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4794"
#> 
#> [[57]]$name
#> [1] "cancer"
#> 
#> [[57]]$species
#> [1] "Homo sapiens"
#> 
#> [[57]]$revision
#> [1] "108946"
#> 
#> 
#> [[58]]
#> [[58]]$id
#> [1] "WP4793"
#> 
#> [[58]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4793"
#> 
#> [[58]]$name
#> [1] "HCC_test_01"
#> 
#> [[58]]$species
#> [1] "Homo sapiens"
#> 
#> [[58]]$revision
#> [1] "108940"
#> 
#> 
#> [[59]]
#> [[59]]$id
#> [1] "WP4747"
#> 
#> [[59]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4747"
#> 
#> [[59]]$name
#> [1] "Netrin-UNC5B signaling Pathway"
#> 
#> [[59]]$species
#> [1] "Homo sapiens"
#> 
#> [[59]]$revision
#> [1] "108877"
#> 
#> 
#> [[60]]
#> [[60]]$id
#> [1] "WP4792"
#> 
#> [[60]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4792"
#> 
#> [[60]]$name
#> [1] "Purine metabolism"
#> 
#> [[60]]$species
#> [1] "Homo sapiens"
#> 
#> [[60]]$revision
#> [1] "108876"
#> 
#> 
#> [[61]]
#> [[61]]$id
#> [1] "WP4189"
#> 
#> [[61]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4189"
#> 
#> [[61]]$name
#> [1] "Mevalonate arm of cholesterol biosynthesis pathway with inhibitors"
#> 
#> [[61]]$species
#> [1] "Homo sapiens"
#> 
#> [[61]]$revision
#> [1] "108848"
#> 
#> 
#> [[62]]
#> [[62]]$id
#> [1] "WP4790"
#> 
#> [[62]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4790"
#> 
#> [[62]]$name
#> [1] "FGF23 signalling in Hypophosphatemic rickets"
#> 
#> [[62]]$species
#> [1] "Homo sapiens"
#> 
#> [[62]]$revision
#> [1] "108845"
#> 
#> 
#> [[63]]
#> [[63]]$id
#> [1] "WP4789"
#> 
#> [[63]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4789"
#> 
#> [[63]]$name
#> [1] "MED and Pseudoachondroplasia genes"
#> 
#> [[63]]$species
#> [1] "Homo sapiens"
#> 
#> [[63]]$revision
#> [1] "109081"
#> 
#> 
#> [[64]]
#> [[64]]$id
#> [1] "WP4674"
#> 
#> [[64]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4674"
#> 
#> [[64]]$name
#> [1] "Head and Neck Squamous Cell Carcinoma"
#> 
#> [[64]]$species
#> [1] "Homo sapiens"
#> 
#> [[64]]$revision
#> [1] "109135"
#> 
#> 
#> [[65]]
#> [[65]]$id
#> [1] "WP4"
#> 
#> [[65]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4"
#> 
#> [[65]]$name
#> [1] "Sandbox Pathway Test"
#> 
#> [[65]]$species
#> [1] "Homo sapiens"
#> 
#> [[65]]$revision
#> [1] "108828"
#> 
#> 
#> [[66]]
#> [[66]]$id
#> [1] "WP4788"
#> 
#> [[66]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4788"
#> 
#> [[66]]$name
#> [1] "Autosomal recessive Osteopetrosis pathways"
#> 
#> [[66]]$species
#> [1] "Homo sapiens"
#> 
#> [[66]]$revision
#> [1] "108987"
#> 
#> 
#> [[67]]
#> [[67]]$id
#> [1] "WP4787"
#> 
#> [[67]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4787"
#> 
#> [[67]]$name
#> [1] "Osteoblast differentiation"
#> 
#> [[67]]$species
#> [1] "Homo sapiens"
#> 
#> [[67]]$revision
#> [1] "108819"
#> 
#> 
#> [[68]]
#> [[68]]$id
#> [1] "WP4786"
#> 
#> [[68]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4786"
#> 
#> [[68]]$name
#> [1] "Type I collagen synthesis in the context of Osteogenesis imperfecta"
#> 
#> [[68]]$species
#> [1] "Homo sapiens"
#> 
#> [[68]]$revision
#> [1] "108833"
#> 
#> 
#> [[69]]
#> [[69]]$id
#> [1] "WP4757"
#> 
#> [[69]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4757"
#> 
#> [[69]]$name
#> [1] ""
#> 
#> [[69]]$species
#> [1] ""
#> 
#> [[69]]$revision
#> [1] 0
#> 
#> 
#> [[70]]
#> [[70]]$id
#> [1] "WP1424"
#> 
#> [[70]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP1424"
#> 
#> [[70]]$name
#> [1] "Globo Sphingolipid Metabolism"
#> 
#> [[70]]$species
#> [1] "Homo sapiens"
#> 
#> [[70]]$revision
#> [1] "108809"
#> 
#> 
#> [[71]]
#> [[71]]$id
#> [1] "WP2023"
#> 
#> [[71]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP2023"
#> 
#> [[71]]$name
#> [1] "Cell Differentiation - Index expanded"
#> 
#> [[71]]$species
#> [1] "Homo sapiens"
#> 
#> [[71]]$revision
#> [1] "108801"
#> 
#> 
#> [[72]]
#> [[72]]$id
#> [1] "WP1544"
#> 
#> [[72]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP1544"
#> 
#> [[72]]$name
#> [1] "MicroRNAs in cardiomyocyte hypertrophy"
#> 
#> [[72]]$species
#> [1] "Homo sapiens"
#> 
#> [[72]]$revision
#> [1] "108800"
#> 
#> 
#> [[73]]
#> [[73]]$id
#> [1] "WP1984"
#> 
#> [[73]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP1984"
#> 
#> [[73]]$name
#> [1] "Integrated Breast Cancer Pathway"
#> 
#> [[73]]$species
#> [1] "Homo sapiens"
#> 
#> [[73]]$revision
#> [1] "108807"
#> 
#> 
#> [[74]]
#> [[74]]$id
#> [1] "WP4142"
#> 
#> [[74]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4142"
#> 
#> [[74]]$name
#> [1] "Metabolism of Spingolipids in ER and Golgi apparatus"
#> 
#> [[74]]$species
#> [1] "Homo sapiens"
#> 
#> [[74]]$revision
#> [1] "108860"
#> 
#> 
#> [[75]]
#> [[75]]$id
#> [1] "WP4785"
#> 
#> [[75]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4785"
#> 
#> [[75]]$name
#> [1] "Somitogenesis in the context of spondylocostal dysostosis"
#> 
#> [[75]]$species
#> [1] "Homo sapiens"
#> 
#> [[75]]$revision
#> [1] "108794"
#> 
#> 
#> [[76]]
#> [[76]]$id
#> [1] "WP4784"
#> 
#> [[76]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4784"
#> 
#> [[76]]$name
#> [1] "Proteoglycan biosynthesis"
#> 
#> [[76]]$species
#> [1] "Homo sapiens"
#> 
#> [[76]]$revision
#> [1] "108790"
#> 
#> 
#> [[77]]
#> [[77]]$id
#> [1] "WP4259"
#> 
#> [[77]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4259"
#> 
#> [[77]]$name
#> [1] "Disorders of Folate Metabolism and Transport"
#> 
#> [[77]]$species
#> [1] "Homo sapiens"
#> 
#> [[77]]$revision
#> [1] "108786"
#> 
#> 
#> [[78]]
#> [[78]]$id
#> [1] "WP1946"
#> 
#> [[78]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP1946"
#> 
#> [[78]]$name
#> [1] "Cori Cycle"
#> 
#> [[78]]$species
#> [1] "Homo sapiens"
#> 
#> [[78]]$revision
#> [1] "108783"
#> 
#> 
#> [[79]]
#> [[79]]$id
#> [1] "WP3594"
#> 
#> [[79]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3594"
#> 
#> [[79]]$name
#> [1] "Circadian rhythm related genes"
#> 
#> [[79]]$species
#> [1] "Homo sapiens"
#> 
#> [[79]]$revision
#> [1] "108782"
#> 
#> 
#> [[80]]
#> [[80]]$id
#> [1] "WP4773"
#> 
#> [[80]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4773"
#> 
#> [[80]]$name
#> [1] ""
#> 
#> [[80]]$species
#> [1] ""
#> 
#> [[80]]$revision
#> [1] 0
#> 
#> 
#> [[81]]
#> [[81]]$id
#> [1] "WP4701"
#> 
#> [[81]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4701"
#> 
#> [[81]]$name
#> [1] "Monoterpene Biosynthesis Pathway Genes in Mint"
#> 
#> [[81]]$species
#> [1] "Vitis vinifera"
#> 
#> [[81]]$revision
#> [1] "108774"
#> 
#> 
#> [[82]]
#> [[82]]$id
#> [1] "WP422"
#> 
#> [[82]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP422"
#> 
#> [[82]]$name
#> [1] "MAPK Cascade"
#> 
#> [[82]]$species
#> [1] "Homo sapiens"
#> 
#> [[82]]$revision
#> [1] "108756"
#> 
#> 
#> [[83]]
#> [[83]]$id
#> [1] "WP673"
#> 
#> [[83]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP673"
#> 
#> [[83]]$name
#> [1] "ErbB Signaling Pathway"
#> 
#> [[83]]$species
#> [1] "Homo sapiens"
#> 
#> [[83]]$revision
#> [1] "108898"
#> 
#> 
#> [[84]]
#> [[84]]$id
#> [1] "WP4772"
#> 
#> [[84]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4772"
#> 
#> [[84]]$name
#> [1] ""
#> 
#> [[84]]$species
#> [1] ""
#> 
#> [[84]]$revision
#> [1] 0
#> 
#> 
#> [[85]]
#> [[85]]$id
#> [1] "WP4771"
#> 
#> [[85]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4771"
#> 
#> [[85]]$name
#> [1] ""
#> 
#> [[85]]$species
#> [1] ""
#> 
#> [[85]]$revision
#> [1] 0
#> 
#> 
#> [[86]]
#> [[86]]$id
#> [1] "WP4770"
#> 
#> [[86]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4770"
#> 
#> [[86]]$name
#> [1] ""
#> 
#> [[86]]$species
#> [1] ""
#> 
#> [[86]]$revision
#> [1] 0
#> 
#> 
#> [[87]]
#> [[87]]$id
#> [1] "WP4467"
#> 
#> [[87]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4467"
#> 
#> [[87]]$name
#> [1] ""
#> 
#> [[87]]$species
#> [1] ""
#> 
#> [[87]]$revision
#> [1] 0
#> 
#> 
#> [[88]]
#> [[88]]$id
#> [1] "WP4518"
#> 
#> [[88]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4518"
#> 
#> [[88]]$name
#> [1] "Gamma-Glutamyl Cycle for the biosynthesis and degradation of glutathione, including diseases"
#> 
#> [[88]]$species
#> [1] "Homo sapiens"
#> 
#> [[88]]$revision
#> [1] "108661"
#> 
#> 
#> [[89]]
#> [[89]]$id
#> [1] "WP2029"
#> 
#> [[89]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP2029"
#> 
#> [[89]]$name
#> [1] "Cell Differentiation - Index"
#> 
#> [[89]]$species
#> [1] "Homo sapiens"
#> 
#> [[89]]$revision
#> [1] "108731"
#> 
#> 
#> [[90]]
#> [[90]]$id
#> [1] "WP4746"
#> 
#> [[90]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4746"
#> 
#> [[90]]$name
#> [1] "Thyroid hormones production and their peripheral downstream signalling effects regarding congenital hypothyroidism"
#> 
#> [[90]]$species
#> [1] "Homo sapiens"
#> 
#> [[90]]$revision
#> [1] "108781"
#> 
#> 
#> [[91]]
#> [[91]]$id
#> [1] "WP4761"
#> 
#> [[91]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4761"
#> 
#> [[91]]$name
#> [1] ""
#> 
#> [[91]]$species
#> [1] ""
#> 
#> [[91]]$revision
#> [1] 0
#> 
#> 
#> [[92]]
#> [[92]]$id
#> [1] "WP4762"
#> 
#> [[92]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4762"
#> 
#> [[92]]$name
#> [1] ""
#> 
#> [[92]]$species
#> [1] ""
#> 
#> [[92]]$revision
#> [1] 0
#> 
#> 
#> [[93]]
#> [[93]]$id
#> [1] "WP4754"
#> 
#> [[93]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4754"
#> 
#> [[93]]$name
#> [1] "IL-18 signaling pathway"
#> 
#> [[93]]$species
#> [1] "Homo sapiens"
#> 
#> [[93]]$revision
#> [1] "108627"
#> 
#> 
#> [[94]]
#> [[94]]$id
#> [1] "WP4767"
#> 
#> [[94]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4767"
#> 
#> [[94]]$name
#> [1] "FGFR3 signalling in chondrocyte proliferation and terminal differentiation"
#> 
#> [[94]]$species
#> [1] "Homo sapiens"
#> 
#> [[94]]$revision
#> [1] "109131"
#> 
#> 
#> [[95]]
#> [[95]]$id
#> [1] "WP339"
#> 
#> [[95]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP339"
#> 
#> [[95]]$name
#> [1] "ESC Pluripotency Pathways"
#> 
#> [[95]]$species
#> [1] "Mus musculus"
#> 
#> [[95]]$revision
#> [1] "108619"
#> 
#> 
#> [[96]]
#> [[96]]$id
#> [1] "WP531"
#> 
#> [[96]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP531"
#> 
#> [[96]]$name
#> [1] "DNA Mismatch Repair"
#> 
#> [[96]]$species
#> [1] "Homo sapiens"
#> 
#> [[96]]$revision
#> [1] "108607"
#> 
#> 
#> [[97]]
#> [[97]]$id
#> [1] "WP4752"
#> 
#> [[97]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4752"
#> 
#> [[97]]$name
#> [1] "Base Excision Repair"
#> 
#> [[97]]$species
#> [1] "Homo sapiens"
#> 
#> [[97]]$revision
#> [1] "108597"
#> 
#> 
#> [[98]]
#> [[98]]$id
#> [1] "WP4753"
#> 
#> [[98]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4753"
#> 
#> [[98]]$name
#> [1] "Nucleotide Excision Repair"
#> 
#> [[98]]$species
#> [1] "Homo sapiens"
#> 
#> [[98]]$revision
#> [1] "108596"
#> 
#> 
#> [[99]]
#> [[99]]$id
#> [1] "WP4759"
#> 
#> [[99]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4759"
#> 
#> [[99]]$name
#> [1] ""
#> 
#> [[99]]$species
#> [1] ""
#> 
#> [[99]]$revision
#> [1] 0
#> 
#> 
#> [[100]]
#> [[100]]$id
#> [1] "WP4737"
#> 
#> [[100]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4737"
#> 
#> [[100]]$name
#> [1] ""
#> 
#> [[100]]$species
#> [1] ""
#> 
#> [[100]]$revision
#> [1] 0
#> 
#> 
#> [[101]]
#> [[101]]$id
#> [1] "WP4764"
#> 
#> [[101]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4764"
#> 
#> [[101]]$name
#> [1] ""
#> 
#> [[101]]$species
#> [1] ""
#> 
#> [[101]]$revision
#> [1] 0
#> 
#> 
#> [[102]]
#> [[102]]$id
#> [1] "WP4765"
#> 
#> [[102]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4765"
#> 
#> [[102]]$name
#> [1] ""
#> 
#> [[102]]$species
#> [1] ""
#> 
#> [[102]]$revision
#> [1] 0
#> 
#> 
#> [[103]]
#> [[103]]$id
#> [1] "WP3888"
#> 
#> [[103]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP3888"
#> 
#> [[103]]$name
#> [1] "VEGFA-VEGFR2 Signaling Pathway"
#> 
#> [[103]]$species
#> [1] "Homo sapiens"
#> 
#> [[103]]$revision
#> [1] "108912"
#> 
#> 
#> [[104]]
#> [[104]]$id
#> [1] "WP4224"
#> 
#> [[104]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4224"
#> 
#> [[104]]$name
#> [1] "Purine metabolism and related disorders"
#> 
#> [[104]]$species
#> [1] "Homo sapiens"
#> 
#> [[104]]$revision
#> [1] "108454"
#> 
#> 
#> [[105]]
#> [[105]]$id
#> [1] "WP4742"
#> 
#> [[105]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4742"
#> 
#> [[105]]$name
#> [1] "Ketogenesis and Ketolysis"
#> 
#> [[105]]$species
#> [1] "Homo sapiens"
#> 
#> [[105]]$revision
#> [1] "108452"
#> 
#> 
#> [[106]]
#> [[106]]$id
#> [1] "WP4763"
#> 
#> [[106]]$url
#> [1] "https://www.wikipathways.org/index.php/Pathway:WP4763"
#> 
#> [[106]]$name
#> [1] "Regulation of Pgc1a expression by a Gsk3b-Tfeb signaling axis in skeletal muscle"
#> 
#> [[106]]$species
#> [1] "Mus musculus"
#> 
#> [[106]]$revision
#> [1] "108519"
#> 
#>