Get Pathway Names of Recent Changes

Retrieve names of recently changed pathways at WikiPathways.

getRecentChangesNames(timestamp)

Arguments

timestamp
(14 digits, YYYYMMDDhhmmss) Limit by time, only history items after the given time, e.g., 20180201000000 for changes since Feb 1st, 2018.

Value

A list of pathway names. Note: pathway deletions will be listed as blank names.

Examples

{
getRecentChangesNames('20180201000000')
}

#>   [1] "mBDNF and proBDNF regulation of GABA neurotransmission"                                                            
#>   [2] "Vitamin B6-dependent and responsive disorders"                                                                     
#>   [3] "Envelope proteins and their potential roles in EDMD physiopathology"                                               
#>   [4] "Viral Acute Myocarditis"                                                                                           
#>   [5] "Neurodegeneration with brain iron accumulation (NBIA) subtypes pathway"                                            
#>   [6] "Amyotrophic lateral sclerosis (ALS)"                                                                               
#>   [7] "Mammalian Sexual Development"                                                                                      
#>   [8] "inflammation and cancer"                                                                                           
#>   [9] "inflammation leading to cancer"                                                                                    
#>  [10] ""                                                                                                                  
#>  [11] ""                                                                                                                  
#>  [12] ""                                                                                                                  
#>  [13] ""                                                                                                                  
#>  [14] ""                                                                                                                  
#>  [15] ""                                                                                                                  
#>  [16] ""                                                                                                                  
#>  [17] ""                                                                                                                  
#>  [18] ""                                                                                                                  
#>  [19] ""                                                                                                                  
#>  [20] ""                                                                                                                  
#>  [21] ""                                                                                                                  
#>  [22] ""                                                                                                                  
#>  [23] ""                                                                                                                  
#>  [24] "MTOR 1"                                                                                                            
#>  [25] "Regucalcin in proximal tubule kidney cells"                                                                        
#>  [26] "Genes controlling renal nephrogenesis"                                                                             
#>  [27] "GDNF/RET signalling axis"                                                                                          
#>  [28] "Conversion of Angiotensinogen to Angiotensin II "                                                                  
#>  [29] "TGF-beta Receptor Signaling"                                                                                       
#>  [30] "Glycosaminoglycan degradation"                                                                                     
#>  [31] "Somatic Sex determination"                                                                                         
#>  [32] "Regucalcin in kidney cells"                                                                                        
#>  [33] "practice"                                                                                                          
#>  [34] "Kisspeptin/Kisspeptin Receptor System in the Ovary"                                                                
#>  [35] "test4"                                                                                                             
#>  [36] "Academy"                                                                                                           
#>  [37] "Endochondral Ossification with Skeletal Dysplasias"                                                                
#>  [38] ""                                                                                                                  
#>  [39] ""                                                                                                                  
#>  [40] "EGFR Tyrosine Kinase Inhibitor Resistance"                                                                         
#>  [41] "Mitochondrial fatty acid synthesis and respiration"                                                                
#>  [42] "mypath"                                                                                                            
#>  [43] "Cholesterol Biosynthesis with Skeletal Dysplasias"                                                                 
#>  [44] "Ciliopathies"                                                                                                      
#>  [45] ""                                                                                                                  
#>  [46] ""                                                                                                                  
#>  [47] ""                                                                                                                  
#>  [48] ""                                                                                                                  
#>  [49] ""                                                                                                                  
#>  [50] "Test run of glucose metabolism "                                                                                   
#>  [51] ""                                                                                                                  
#>  [52] "Downregulation of ACE2 expression by SARS-CoV infection and SARS-CoV Spike protein."                               
#>  [53] "WikiPahtway introduction"                                                                                          
#>  [54] "test2"                                                                                                             
#>  [55] "Biochemical Pathways Part I"                                                                                       
#>  [56] "Target genes for floral homeotic tetrameric MADS complexes"                                                        
#>  [57] "cancer"                                                                                                            
#>  [58] "HCC_test_01"                                                                                                       
#>  [59] "Netrin-UNC5B signaling Pathway"                                                                                    
#>  [60] "Purine metabolism"                                                                                                 
#>  [61] "Mevalonate arm of cholesterol biosynthesis pathway with inhibitors"                                                
#>  [62] "FGF23 signalling in Hypophosphatemic rickets"                                                                      
#>  [63] "MED and Pseudoachondroplasia genes"                                                                                
#>  [64] "Head and Neck Squamous Cell Carcinoma"                                                                             
#>  [65] "Sandbox Pathway Test"                                                                                              
#>  [66] "Autosomal recessive Osteopetrosis pathways"                                                                        
#>  [67] "Osteoblast differentiation"                                                                                        
#>  [68] "Type I collagen synthesis in the context of Osteogenesis imperfecta"                                               
#>  [69] ""                                                                                                                  
#>  [70] "Globo Sphingolipid Metabolism"                                                                                     
#>  [71] "Cell Differentiation - Index expanded"                                                                             
#>  [72] "MicroRNAs in cardiomyocyte hypertrophy"                                                                            
#>  [73] "Integrated Breast Cancer Pathway"                                                                                  
#>  [74] "Metabolism of Spingolipids in ER and Golgi apparatus"                                                              
#>  [75] "Somitogenesis in the context of spondylocostal dysostosis"                                                         
#>  [76] "Proteoglycan biosynthesis"                                                                                         
#>  [77] "Disorders of Folate Metabolism and Transport"                                                                      
#>  [78] "Cori Cycle"                                                                                                        
#>  [79] "Circadian rhythm related genes"                                                                                    
#>  [80] ""                                                                                                                  
#>  [81] "Monoterpene Biosynthesis Pathway Genes in Mint"                                                                    
#>  [82] "MAPK Cascade"                                                                                                      
#>  [83] "ErbB Signaling Pathway"                                                                                            
#>  [84] ""                                                                                                                  
#>  [85] ""                                                                                                                  
#>  [86] ""                                                                                                                  
#>  [87] ""                                                                                                                  
#>  [88] "Gamma-Glutamyl Cycle for the biosynthesis and degradation of glutathione, including diseases"                      
#>  [89] "Cell Differentiation - Index"                                                                                      
#>  [90] "Thyroid hormones production and their peripheral downstream signalling effects regarding congenital hypothyroidism"
#>  [91] ""                                                                                                                  
#>  [92] ""                                                                                                                  
#>  [93] "IL-18 signaling pathway"                                                                                           
#>  [94] "FGFR3 signalling in chondrocyte proliferation and terminal differentiation"                                        
#>  [95] "ESC Pluripotency Pathways"                                                                                         
#>  [96] "DNA Mismatch Repair"                                                                                               
#>  [97] "Base Excision Repair"                                                                                              
#>  [98] "Nucleotide Excision Repair"                                                                                        
#>  [99] ""                                                                                                                  
#> [100] ""                                                                                                                  
#> [101] ""                                                                                                                  
#> [102] ""                                                                                                                  
#> [103] "VEGFA-VEGFR2 Signaling Pathway"                                                                                    
#> [104] "Purine metabolism and related disorders"                                                                           
#> [105] "Ketogenesis and Ketolysis"                                                                                         
#> [106] "Regulation of Pgc1a expression by a Gsk3b-Tfeb signaling axis in skeletal muscle"