Find Pathway WPIDs By Xref

Retrieve list of pathway WPIDs containing the query Xref by identifier and system code.

findPathwayIdsByXref(identifier, systemCode)

Arguments

identifier
(character) The official ID specified by a data source or system

systemCode
(character) The BridgeDb code associated with the data source or system, e.g., En (Ensembl), L (Entrez), Ch (HMDB), etc. See column two of https://github.com/bridgedb/BridgeDb/blob/master/org.bridgedb.bio/resources/org/bridgedb/bio/datasources.txt.

Value

A list of WPIDs

Details

Note: there will be multiple listings of the same pathway if the Xref is present mutiple times.

Examples

{
findPathwayIdsByXref('ENSG00000232810','En')
}

#>   [1] "WP2813" "WP4329" "WP4329" "WP2882" "WP4066" "WP4066" "WP3980" "WP3980"
#>   [9] "WP4063" "WP3599" "WP2873" "WP3893" "WP3893" "WP262"  "WP4136" "WP3865"
#>  [17] "WP3941" "WP3624" "WP2431" "WP2431" "WP2431" "WP2431" "WP3863" "WP4673"
#>  [25] "WP3287" "WP4008" "WP4341" "WP4666" "WP4666" "WP2059" "WP2571" "WP1584"
#>  [33] "WP4298" "WP4481" "WP530"  "WP530"  "WP530"  "WP530"  "WP530"  "WP530" 
#>  [41] "WP530"  "WP3380" "WP3380" "WP3380" "WP3380" "WP3380" "WP3380" "WP3380"
#>  [49] "WP3380" "WP3380" "WP3380" "WP3380" "WP3380" "WP3380" "WP3380" "WP3380"
#>  [57] "WP3380" "WP3380" "WP3380" "WP3380" "WP3380" "WP3380" "WP3380" "WP3380"
#>  [65] "WP3380" "WP3380" "WP3380" "WP3380" "WP3380" "WP3380" "WP3380" "WP3380"
#>  [73] "WP3380" "WP3380" "WP3380" "WP3380" "WP3380" "WP3380" "WP3380" "WP3380"
#>  [81] "WP2751" "WP4396" "WP4396" "WP4396" "WP4066" "WP4066" "WP3980" "WP3398"
#>  [89] "WP3398" "WP3398" "WP3398" "WP3398" "WP3398" "WP3398" "WP3398" "WP4496"
#>  [97] "WP3611" "WP3611" "WP3611" "WP4493" "WP4063" "WP4478" "WP231"  "WP231" 
#> [105] "WP231"  "WP231"  "WP231"  "WP231"  "WP3851" "WP2036" "WP2036" "WP2036"
#> [113] "WP75"   "WP236"  "WP254"  "WP314"  "WP560"  "WP560"  "WP1449" "WP3617"
#> [121] "WP3617" "WP2328" "WP129"  "WP2513" "WP2795" "WP3297" "WP706"  "WP15"  
#> [129] "WP15"   "WP1533" "WP1533" "WP176"  "WP176"  "WP382"  "WP727"  "WP2586"
#> [137] "WP4480" "WP4304" "WP4304" "WP2864" "WP4482" "WP4754" "WP4754" "WP1544"
#> [145] "WP4747" "WP4816" "WP4816" "WP2447" "WP2447"