Find Pathway Names By Literature

Retrieve list of pathway names containing the query citation.

findPathwayNamesByLiterature(query)

Arguments

query
The character string to search for, e.g., a PMID, title keyword or author name.

Value

A list of lists

Details

Note: there will be multiple listings of the same pathway name if copies exist for multiple species.

Examples

{
findPathwayNamesByLiterature('19649250')
findPathwayNamesByLiterature('smith')
findPathwayNamesByLiterature('cancer')
}

#>   [1] "E3 ubiquitin ligases ubiquitinate target proteins"                                                                          
#>   [2] "Signaling by FGFR1"                                                                                                         
#>   [3] "Irinotecan Pathway"                                                                                                         
#>   [4] "Irinotecan Pathway"                                                                                                         
#>   [5] "Irinotecan Pathway"                                                                                                         
#>   [6] "Irinotecan Pathway"                                                                                                         
#>   [7] "Irinotecan Pathway"                                                                                                         
#>   [8] "Irinotecan Pathway"                                                                                                         
#>   [9] "RAF/MAP kinase cascade"                                                                                                     
#>  [10] "Energy dependent regulation of mTOR by LKB1-AMPK"                                                                           
#>  [11] "Hedgehog ligand biogenesis"                                                                                                 
#>  [12] "Signaling by FGFR2"                                                                                                         
#>  [13] "mTOR signalling"                                                                                                            
#>  [14] "Metastatic brain tumor"                                                                                                     
#>  [15] "Signaling by Rho GTPases"                                                                                                   
#>  [16] "Signaling by FGFR4"                                                                                                         
#>  [17] "ATM Signaling Pathway"                                                                                                      
#>  [18] "NRF2-ARE regulation"                                                                                                        
#>  [19] "Signaling by FGFR3"                                                                                                         
#>  [20] "DNA IR-Double Strand Breaks (DSBs) and cellular response via ATM"                                                           
#>  [21] "Integrin-mediated Cell Adhesion"                                                                                            
#>  [22] "Bladder Cancer"                                                                                                             
#>  [23] "ATM Signaling Pathway"                                                                                                      
#>  [24] "TP53 Regulates Metabolic Genes"                                                                                             
#>  [25] "Overview of leukocyte-intrinsic Hippo pathway functions"                                                                    
#>  [26] "Transcriptional activation by NRF2"                                                                                         
#>  [27] "Keap1-Nrf2"                                                                                                                 
#>  [28] "Keap1-Nrf2"                                                                                                                 
#>  [29] "Keap1-Nrf2"                                                                                                                 
#>  [30] "Retinol metabolism"                                                                                                         
#>  [31] "Phytochemical activity on NRF2 transcriptional activation"                                                                  
#>  [32] "Integrin-mediated Cell Adhesion"                                                                                            
#>  [33] "Keap1-Nrf2"                                                                                                                 
#>  [34] "Keap1-Nrf2"                                                                                                                 
#>  [35] "Keap1-Nrf2"                                                                                                                 
#>  [36] "Metastatic brain tumor"                                                                                                     
#>  [37] "Parkinsons Disease Pathway"                                                                                                 
#>  [38] "Bladder Cancer"                                                                                                             
#>  [39] "DNA IR-damage and cellular response via ATR"                                                                                
#>  [40] "Parkinsons Disease Pathway"                                                                                                 
#>  [41] "DNA Damage Response"                                                                                                        
#>  [42] "miRNA Regulation of DNA Damage Response"                                                                                    
#>  [43] "DNA Damage Response"                                                                                                        
#>  [44] "DNA damage response"                                                                                                        
#>  [45] "HDR through Homologous Recombination (HRR) or Single Strand Annealing (SSA)"                                                
#>  [46] "YAP1- and WWTR1 (TAZ)-stimulated gene expression"                                                                           
#>  [47] "DNA Damage/Telomere Stress Induced Senescence"                                                                              
#>  [48] "Signaling by ERBB2"                                                                                                         
#>  [49] "Signaling by PTK6"                                                                                                          
#>  [50] "NCAM signaling for neurite out-growth"                                                                                      
#>  [51] "Other interleukin signaling"                                                                                                
#>  [52] "Interleukin-3, Interleukin-5 and GM-CSF signaling"                                                                          
#>  [53] "Signaling by SCF-KIT"                                                                                                       
#>  [54] "Signaling by NTRK1 (TRKA)"                                                                                                  
#>  [55] "RNA interference"                                                                                                           
#>  [56] "Signaling by MET"                                                                                                           
#>  [57] "Signaling by VEGF"                                                                                                          
#>  [58] "Signaling by EGFR"                                                                                                          
#>  [59] "Fanconi Anemia Pathway"                                                                                                     
#>  [60] "Generic Transcription Pathway"                                                                                              
#>  [61] "Phase II - Conjugation of compounds"                                                                                        
#>  [62] "TCF dependent signaling in response to WNT"                                                                                 
#>  [63] "Signaling by NTRK3 (TRKC)"                                                                                                  
#>  [64] "Signaling by Type 1 Insulin-like Growth Factor 1 Receptor (IGF1R)"                                                          
#>  [65] "Oxidative Stress Induced Senescence"                                                                                        
#>  [66] "Senescence-Associated Secretory Phenotype (SASP)"                                                                           
#>  [67] "Signaling by NTRK2 (TRKB)"                                                                                                  
#>  [68] "ESR-mediated signaling"                                                                                                     
#>  [69] "PIP3 activates AKT signaling"                                                                                               
#>  [70] "Signaling by ERBB4"                                                                                                         
#>  [71] "PI Metabolism"                                                                                                              
#>  [72] "Signaling by Hippo"                                                                                                         
#>  [73] "TP53 Regulates Transcription of DNA Repair Genes"                                                                           
#>  [74] "Signaling by Leptin"                                                                                                        
#>  [75] "Pre-NOTCH Expression and Processing"                                                                                        
#>  [76] "Oncogene Induced Senescence"                                                                                                
#>  [77] "RET signaling"                                                                                                              
#>  [78] "Gastrin-CREB signalling pathway via PKC and MAPK"                                                                           
#>  [79] "Signaling by NOTCH3"                                                                                                        
#>  [80] "Interleukin-2 family signaling"                                                                                             
#>  [81] "Fc epsilon receptor (FCERI) signaling"                                                                                      
#>  [82] "Transcriptional regulation by RUNX1"                                                                                        
#>  [83] "Signaling by Insulin receptor"                                                                                              
#>  [84] "DAP12 interactions"                                                                                                         
#>  [85] "O-linked glycosylation"                                                                                                     
#>  [86] "Transcriptional regulation by RUNX2"                                                                                        
#>  [87] "Signaling by PDGF"                                                                                                          
#>  [88] "RAB GEFs exchange GTP for GDP on RABs"                                                                                      
#>  [89] "Aflatoxin B1 metabolism"                                                                                                    
#>  [90] "Aflatoxin B1 metabolism"                                                                                                    
#>  [91] "Regulation of sister chromatid separation at the metaphase-anaphase transition"                                             
#>  [92] "Type 2 papillary renal cell carcinoma"                                                                                      
#>  [93] "exRNA mechanism of action and biogenesis"                                                                                   
#>  [94] "Aflatoxin B1 metabolism"                                                                                                    
#>  [95] "Aflatoxin B1 metabolism"                                                                                                    
#>  [96] "Aflatoxin B1 metabolism"                                                                                                    
#>  [97] "Aflatoxin B1 metabolism"                                                                                                    
#>  [98] "Glycerophospholipid Biosynthetic Pathway"                                                                                   
#>  [99] "Glycerophospholipid Biosynthetic Pathway"                                                                                   
#> [100] "Cancer immunotherapy by PD-1 blockade"                                                                                      
#> [101] "Chromosomal and microsatellite instability in colorectal cancer "                                                           
#> [102] "Epithelial to mesenchymal transition in colorectal cancer"                                                                  
#> [103] "Apoptosis-related network due to altered Notch3 in ovarian cancer"                                                          
#> [104] "Apoptosis-related network due to altered Notch3 in ovarian cancer"                                                          
#> [105] "Small cell lung cancer"                                                                                                     
#> [106] "Mammary gland development pathway - Embryonic development (Stage 1 of 4)"                                                   
#> [107] "TP53 Regulates Transcription of Cell Cycle Genes"                                                                           
#> [108] "Regulation of TP53 Expression and Degradation"                                                                              
#> [109] "Interleukin-15 signaling"                                                                                                   
#> [110] "Processing of Capped Intron-Containing Pre-mRNA"                                                                            
#> [111] "Asparagine N-linked glycosylation"                                                                                          
#> [112] "TGF Beta Signaling Pathway"                                                                                                 
#> [113] "DNA methylation"                                                                                                            
#> [114] "Transforming Growth Factor (TGF) Beta Signaling Pathway"                                                                    
#> [115] "Mitotic G1-G1/S phases"                                                                                                     
#> [116] "Beta-catenin independent WNT signaling"                                                                                     
#> [117] "Abnormal conversion of 2-oxoglutarate to 2-hydroxyglutarate"                                                                
#> [118] "TGF Beta Signaling Pathway"                                                                                                 
#> [119] "Protein ubiquitination"                                                                                                     
#> [120] "Pentose phosphate pathway"                                                                                                  
#> [121] "p53 pathway"                                                                                                                
#> [122] "Transforming Growth Factor (TGF) Beta Signaling Pathway"                                                                    
#> [123] "Inositol phosphate metabolism"                                                                                              
#> [124] "Regulation of TP53 Activity through Phosphorylation"                                                                        
#> [125] "Nucleotide Excision Repair"                                                                                                 
#> [126] "Mismatch Repair"                                                                                                            
#> [127] "TET1,2,3 and TDG demethylate DNA"                                                                                           
#> [128] "Caspase activation via Dependence Receptors in the absence of ligand"                                                       
#> [129] "Transforming Growth Factor (TGF) Beta Signaling Pathway"                                                                    
#> [130] "Transforming Growth Factor (TGF) Beta Signaling Pathway"                                                                    
#> [131] "Ebola Virus Pathway on Host"                                                                                                
#> [132] "NADPH regeneration"                                                                                                         
#> [133] "TGF Beta Signaling Pathway"                                                                                                 
#> [134] "Integrated Cancer pathway"                                                                                                  
#> [135] "Regulation of RUNX1 Expression and Activity"                                                                                
#> [136] "Signaling by Erythropoietin"                                                                                                
#> [137] "Regulation of Hypoxia-inducible Factor (HIF) by oxygen"                                                                     
#> [138] "RUNX1 and FOXP3 control the development of regulatory T lymphocytes (Tregs)"                                                
#> [139] "PluriNetWork"                                                                                                               
#> [140] "S Phase"                                                                                                                    
#> [141] "Rett syndrome causing genes"                                                                                                
#> [142] "Interactions between immune cells and microRNAs in tumor microenvironment"                                                  
#> [143] "MET in type 1 papillary renal cell carcinoma"                                                                               
#> [144] "Oxysterols derived from cholesterol"                                                                                        
#> [145] "Senescence and Autophagy in Cancer"                                                                                         
#> [146] "TGF-beta Receptor Signaling"                                                                                                
#> [147] "Pathways in clear cell renal cell carcinoma"                                                                                
#> [148] "Integrated Cancer Pathway"                                                                                                  
#> [149] "Zinc homeostasis"                                                                                                           
#> [150] "PI3K-Akt Signaling Pathway"                                                                                                 
#> [151] "Tamoxifen metabolism"                                                                                                       
#> [152] "Tamoxifen metabolism"                                                                                                       
#> [153] "Folate Metabolism"                                                                                                          
#> [154] "Selenium Micronutrient Network"                                                                                             
#> [155] "Vitamin B12 Metabolism"                                                                                                     
#> [156] "Selenium Micronutrient Network"                                                                                             
#> [157] "Vitamin B12 Metabolism"                                                                                                     
#> [158] "Folate Metabolism"                                                                                                          
#> [159] "Mevalonate pathway"                                                                                                         
#> [160] "Fragile X Syndrome "                                                                                                        
#> [161] "Integrated Breast Cancer Pathway"                                                                                           
#> [162] "Breast cancer pathway"                                                                                                      
#> [163] "Endometrial cancer"                                                                                                         
#> [164] "FABP4 in ovarian cancer"                                                                                                    
#> [165] "Senescence and Autophagy in Cancer"                                                                                         
#> [166] "Deregulation of Rab and Rab Effector Genes in Bladder Cancer"                                                               
#> [167] "Deregulation of Rab and Rab Effector Genes in Bladder Cancer"                                                               
#> [168] "Non-small cell lung cancer"                                                                                                 
#> [169] "Fluoropyrimidine Activity"                                                                                                  
#> [170] "Fluoropyrimidine Activity"                                                                                                  
#> [171] "Tamoxifen metabolism"                                                                                                       
#> [172] "RAC1/PAK1/p38/MMP2 Pathway"                                                                                                 
#> [173] "ErbB Signaling Pathway"                                                                                                     
#> [174] "Integrated Breast Cancer Pathway"                                                                                           
#> [175] "ErbB Signaling Pathway"                                                                                                     
#> [176] "TGF-beta Receptor Signaling"                                                                                                
#> [177] "Neurodegeneration with brain iron accumulation (NBIA) subtypes pathway"                                                     
#> [178] "Transport of bile salts and organic acids, metal ions and amine compounds"                                                  
#> [179] "MicroRNA for Targeting Cancer Growth and Vascularization in Glioblastoma"                                                   
#> [180] "MFAP5 effect on permeability and motility of endothelial cells via cytoskeleton rearrangement"                              
#> [181] "Modified nucleosides derived from t-RNA as urinary cancer markers"                                                          
#> [182] "Hereditary leiomyomatosis and renal cell carcinoma pathway"                                                                 
#> [183] "Mitotic G2-G2/M phases"                                                                                                     
#> [184] "RUNX1 regulates genes involved in megakaryocyte differentiation and platelet function"                                      
#> [185] "Integrin alphaIIb beta3 signaling"                                                                                          
#> [186] "Sphingolipid metabolism"                                                                                                    
#> [187] "Transcriptional regulation by RUNX3"                                                                                        
#> [188] "Hedgehog 'on' state"                                                                                                        
#> [189] "Unfolded Protein Response (UPR)"                                                                                            
#> [190] "RHO GTPases Activate Formins"                                                                                               
#> [191] "Genotoxicity pathway"                                                                                                       
#> [192] "Sulfur amino acid metabolism"                                                                                               
#> [193] "RHO GTPases Activate WASPs and WAVEs"                                                                                       
#> [194] "WNT ligand biogenesis and trafficking"                                                                                      
#> [195] "SUMOylation of transcription factors"                                                                                       
#> [196] "Gap junction trafficking and regulation"                                                                                    
#> [197] "Miscellaneous transport and binding events"                                                                                 
#> [198] "Signaling by NOTCH4"                                                                                                        
#> [199] "Nuclear Receptors Meta-Pathway"                                                                                             
#> [200] "Visual phototransduction"                                                                                                   
#> [201] "Chromatin organization"                                                                                                     
#> [202] "RAF-independent MAPK1/3 activation"                                                                                         
#> [203] "Netrin-1 signaling"                                                                                                         
#> [204] "Abacavir transport and metabolism"                                                                                          
#> [205] "Amino acid synthesis and interconversion (transamination)"                                                                  
#> [206] "Iron uptake and transport"                                                                                                  
#> [207] "Metabolism of polyamines"                                                                                                   
#> [208] "Activation of anterior HOX genes in hindbrain development during early embryogenesis"                                       
#> [209] "Degradation of the extracellular matrix"                                                                                    
#> [210] "Nucleobase catabolism"                                                                                                      
#> [211] "Activated PKN1 stimulates transcription of AR (androgen receptor) regulated genes KLK2 and KLK3"                            
#> [212] "Metabolism of folate and pterines"                                                                                          
#> [213] "Neurotransmitter Release Cycle"                                                                                             
#> [214] "Signaling by ROBO receptors"                                                                                                
#> [215] "Surfactant metabolism"                                                                                                      
#> [216] "Transcriptional Regulation by MECP2"                                                                                        
#> [217] "RHO GTPases activate PKNs"                                                                                                  
#> [218] "DNA Double Strand Break Response"                                                                                           
#> [219] "MAPK6/MAPK4 signaling"                                                                                                      
#> [220] "Signaling by NOTCH2"                                                                                                        
#> [221] "tRNA processing in the nucleus"                                                                                             
#> [222] "Glycosaminoglycan metabolism"                                                                                               
#> [223] "Iron uptake and transport"                                                                                                  
#> [224] "Signaling by NOTCH1"                                                                                                        
#> [225] "PTEN Regulation"                                                                                                            
#> [226] "Neurotransmitter release cycle"                                                                                             
#> [227] "miRNAs involved in DNA damage response"                                                                                     
#> [228] "Antimicrobial peptides"                                                                                                     
#> [229] "NRF2 pathway"                                                                                                               
#> [230] "Neddylation"                                                                                                                
#> [231] "Transcriptional regulation by the AP-2 (TFAP2) family of transcription factors"                                             
#> [232] "Phase I - Functionalization of compounds"                                                                                   
#> [233] "Pancreatic adenocarcinoma pathway"                                                                                          
#> [234] "Vitamin D Receptor Pathway"                                                                                                 
#> [235] "miRNA targets in ECM and membrane receptors"                                                                                
#> [236] "miRNAs involved in DNA damage response"                                                                                     
#> [237] "Apoptosis Modulation and Signaling"                                                                                         
#> [238] "Thyroxine (Thyroid Hormone) Production"                                                                                     
#> [239] "Influenza A virus infection"                                                                                                
#> [240] "FAS pathway and Stress induction of HSP regulation"                                                                         
#> [241] "Quercetin and Nf-kB/ AP-1 Induced Cell Apoptosis"                                                                           
#> [242] "Estrogen metabolism"                                                                                                        
#> [243] "Selenium Metabolism and Selenoproteins"                                                                                     
#> [244] "Selenium metabolism Selenoproteins"                                                                                         
#> [245] "Estrogen metabolism"                                                                                                        
#> [246] "Selenium metabolism Selenoproteins"                                                                                         
#> [247] "Estrogen metabolism"                                                                                                        
#> [248] "Thyroxine (Thyroid Hormone) Production"                                                                                     
#> [249] "Serotonin HTR1 Group and FOS Pathway"                                                                                       
#> [250] "Quercetin and Nf-kB/ AP-1 Induced Cell Apoptosis"                                                                           
#> [251] "Eicosanoid metabolism via Cyclo Oxygenases (COX)"                                                                           
#> [252] "Eicosanoid metabolism via Lipo Oxygenases (LOX)"                                                                            
#> [253] "Fas Ligand (FasL) pathway and Stress induction of Heat Shock Proteins (HSP) regulation"                                     
#> [254] "Selenium Metabolism and Selenoproteins"                                                                                     
#> [255] "Lipid Metabolism Pathway"                                                                                                   
#> [256] "NRF2 pathway"                                                                                                               
#> [257] "Malate-aspartate shuttle"                                                                                                   
#> [258] "Estrogen metabolism"                                                                                                        
#> [259] "Estrogen metabolism"                                                                                                        
#> [260] "Estrogen metabolism"                                                                                                        
#> [261] "Selenium metabolism Selenoproteins"                                                                                         
#> [262] "Estrogen metabolism"                                                                                                        
#> [263] "Elongation of (very) long chain fatty acids"                                                                                
#> [264] "Apoptosis Modulation and Signaling"                                                                                         
#> [265] "Eicosanoid metabolism via Cyclo Oxygenases (COX)"                                                                           
#> [266] "Eicosanoid metabolism via Lipo Oxygenases (LOX)"                                                                            
#> [267] "PI3K-AKT-mTOR signaling pathway and therapeutic opportunities"                                                              
#> [268] "miRNA regulation of prostate cancer signaling pathways"                                                                     
#> [269] "miRNA regulation of p53 pathway in prostate cancer"                                                                         
#> [270] "Cancer immunotherapy by CTLA4 blockade"                                                                                     
#> [271] "mir-193a and MVP in colon cancer metastasis"                                                                                
#> [272] "LncRNA involvement in canonical Wnt signaling and colorectal cancer"                                                        
#> [273] "PDGFR-beta pathway"                                                                                                         
#> [274] "Influenza A virus infection"                                                                                                
#> [275] "Genetic alterations of lung cancer"                                                                                         
#> [276] "Netrin-UNC5B signaling Pathway"                                                                                             
#> [277] "EGFR Tyrosine Kinase Inhibitor Resistance"                                                                                  
#> [278] "Transcriptional regulation of pluripotent stem cells"                                                                       
#> [279] "Serotonin Receptor 4/6/7 and NR3C Signaling"                                                                                
#> [280] "Signaling by NODAL"                                                                                                         
#> [281] "GPVI-mediated activation cascade"                                                                                           
#> [282] "Regulation of RUNX2 expression and activity"                                                                                
#> [283] "NoRC negatively regulates rRNA expression"                                                                                  
#> [284] "Degradation of beta-catenin by the destruction complex"                                                                     
#> [285] "BMAL1:CLOCK,NPAS2 activates circadian gene expression"                                                                      
#> [286] "Glycerophospholipid biosynthesis"                                                                                           
#> [287] "Fatty acyl-CoA biosynthesis"                                                                                                
#> [288] "SUMOylation of intracellular receptors"                                                                                     
#> [289] "Signaling by BMP"                                                                                                           
#> [290] "Regulated Necrosis"                                                                                                         
#> [291] "RHO GTPases activate IQGAPs"                                                                                                
#> [292] "Interleukin-20 family signaling"                                                                                            
#> [293] "TNF signaling"                                                                                                              
#> [294] "POU5F1 (OCT4), SOX2, NANOG activate genes related to proliferation"                                                         
#> [295] "Gastric cancer network 2"                                                                                                   
#> [296] "Vasopressin regulates renal water homeostasis via Aquaporins"                                                               
#> [297] "Transcriptional regulation of white adipocyte differentiation"                                                              
#> [298] "Extracellular matrix organization"                                                                                          
#> [299] "DNA Damage Reversal"                                                                                                        
#> [300] "Nanoparticle triggered autophagic cell death"                                                                               
#> [301] "Structural Pathway of Interleukin 1 (IL-1)"                                                                                 
#> [302] "Deubiquitination"                                                                                                           
#> [303] "Interleukin-4 and Interleukin-13 signaling"                                                                                 
#> [304] "Gene Silencing by RNA"                                                                                                      
#> [305] "Transcriptional activity of SMAD2/SMAD3:SMAD4 heterotrimer"                                                                 
#> [306] "Gamma carboxylation, hypusine formation and arylsulfatase activation"                                                       
#> [307] "Metabolism of steroid hormones"                                                                                             
#> [308] "Transcriptional Regulation by E2F6"                                                                                         
#> [309] "Signaling by Activin"                                                                                                       
#> [310] "Liver steatosis AOP"                                                                                                        
#> [311] "SUMOylation of ubiquitinylation proteins"                                                                                   
#> [312] "Regulation of TP53 Activity through Association with Co-factors"                                                            
#> [313] "Serotonin Receptor 2 and STAT3 Signaling"                                                                                   
#> [314] "Class I MHC mediated antigen processing &amp; presentation"                                                                 
#> [315] "Signaling by MST1"                                                                                                          
#> [316] "mir-124 predicted interactions with cell cycle and differentiation "                                                        
#> [317] "Nicotinate metabolism"                                                                                                      
#> [318] "ATM Signaling Pathway"                                                                                                      
#> [319] "Serotonin Receptor 2 and ELK-SRF/GATA4 signaling"                                                                           
#> [320] "Interleukin-17 signaling"                                                                                                   
#> [321] "MyD88 dependent cascade initiated on endosome"                                                                              
#> [322] "TP53 Regulates Transcription of Cell Death Genes"                                                                           
#> [323] "Cellular hexose transport"                                                                                                  
#> [324] "tRNA modification in the nucleus and cytosol"                                                                               
#> [325] "Peroxisomal lipid metabolism"                                                                                               
#> [326] "Transport of inorganic cations/anions and amino acids/oligopeptides"                                                        
#> [327] "Toll-like Receptor Cascades"                                                                                                
#> [328] "Class A/1 (Rhodopsin-like receptors)"                                                                                       
#> [329] "Serotonin Receptor 2 and STAT3 Signaling"                                                                                   
#> [330] "miR-517 relationship with ARCN1 and USP1"                                                                                   
#> [331] "Codeine and Morphine Metabolism"                                                                                            
#> [332] "H19 action Rb-E2F1 signaling and CDK-Beta-catenin activity"                                                                 
#> [333] "Serotonin Receptor 2 and STAT3 Signaling"                                                                                   
#> [334] "Serotonin Receptor 2 and ELK-SRF/GATA4 signaling"                                                                           
#> [335] "Serotonin Receptor 4/6/7 and NR3C Signaling"                                                                                
#> [336] "Prostaglandin Synthesis and Regulation"                                                                                     
#> [337] "Prader-Willi and Angelman Syndrome"                                                                                         
#> [338] "Codeine and Morphine Metabolism"                                                                                            
#> [339] "Aryl Hydrocarbon Receptor"                                                                                                  
#> [340] "Melatonin metabolism and effects"                                                                                           
#> [341] "Toll-like Receptor Signaling related to MyD88"                                                                              
#> [342] "Aryl Hydrocarbon Receptor Netpath"                                                                                          
#> [343] "Structural Pathway of Interleukin 1 (IL-1)"                                                                                 
#> [344] "Renin - Angiotensin System"                                                                                                 
#> [345] "Gastric Cancer Network 2"                                                                                                   
#> [346] "Glycosylation and related congenital defects"                                                                               
#> [347] "TCA Cycle Nutrient Utilization and Invasiveness of Ovarian Cancer"                                                          
#> [348] "TCA Cycle Nutrient Utilization and Invasiveness of Ovarian Cancer"                                                          
#> [349] "MFAP5-mediated ovarian cancer cell motility and invasiveness"                                                               
#> [350] "Gastric Cancer Network 1"                                                                                                   
#> [351] "Gastric Cancer Network 1"                                                                                                   
#> [352] "miR-509-3p alteration of YAP1/ECM axis"                                                                                     
#> [353] "Simplified Interaction Map Between LOXL4 and Oxidative Stress Pathway"                                                      
#> [354] "Alzheimers Disease"                                                                                                         
#> [355] "Tumor suppressor activity of SMARCB1"                                                                                       
#> [356] "Regulation of Pgc1a expression by a Gsk3b-Tfeb signaling axis in skeletal muscle"                                           
#> [357] "Thyroid hormones production and their peripheral downstream signalling effects regarding congenital hypothyroidism"         
#> [358] "Head and Neck Squamous Cell Carcinoma"                                                                                      
#> [359] "Sialic acid metabolism"                                                                                                     
#> [360] "Inhibition of exosome biogenesis and secretion by Manumycin A in CRPC cells"                                                
#> [361] "Threonine catabolism"                                                                                                       
#> [362] "SUMOylation of transcription cofactors"                                                                                     
#> [363] "Regulation of Insulin-like Growth Factor (IGF) transport and uptake by Insulin-like Growth Factor Binding Proteins (IGFBPs)"
#> [364] "MAP kinase activation"                                                                                                      
#> [365] "Neural Crest Cell Migration during Development"                                                                             
#> [366] "Cell migration and invasion through p75NTR"                                                                                 
#> [367] "Neural Crest Cell Migration in Cancer"                                                                                      
#> [368] "Regulation of mRNA stability by proteins that bind AU-rich elements"                                                        
#> [369] "ATF4 activates genes"                                                                                                       
#> [370] "PI3K Cascade"                                                                                                               
#> [371] "SUMOylation of DNA damage response and repair proteins"