Find Pathway URLs By Xref

Retrieve list of pathway URLs containing the query Xref by identifier and system code.

findPathwayUrlsByXref(identifier, systemCode)

Arguments

identifier
(character) The official ID specified by a data source or system

systemCode
(character) The BridgeDb code associated with the data source or system, e.g., En (Ensembl), L (Entrez), Ch (HMDB), etc. See column two of https://github.com/bridgedb/BridgeDb/blob/master/org.bridgedb.bio/resources/org/bridgedb/bio/datasources.txt.

Value

A list of lists

Details

Note: there will be multiple listings of the same pathway if the Xref is present mutiple times.

Examples

{
findPathwayUrlsByXref('ENSG00000232810','En')
}

#>   [1] "https://www.wikipathways.org/index.php/Pathway:WP2813"
#>   [2] "https://www.wikipathways.org/index.php/Pathway:WP4329"
#>   [3] "https://www.wikipathways.org/index.php/Pathway:WP4329"
#>   [4] "https://www.wikipathways.org/index.php/Pathway:WP2882"
#>   [5] "https://www.wikipathways.org/index.php/Pathway:WP4066"
#>   [6] "https://www.wikipathways.org/index.php/Pathway:WP4066"
#>   [7] "https://www.wikipathways.org/index.php/Pathway:WP3980"
#>   [8] "https://www.wikipathways.org/index.php/Pathway:WP3980"
#>   [9] "https://www.wikipathways.org/index.php/Pathway:WP4063"
#>  [10] "https://www.wikipathways.org/index.php/Pathway:WP3599"
#>  [11] "https://www.wikipathways.org/index.php/Pathway:WP2873"
#>  [12] "https://www.wikipathways.org/index.php/Pathway:WP3893"
#>  [13] "https://www.wikipathways.org/index.php/Pathway:WP3893"
#>  [14] "https://www.wikipathways.org/index.php/Pathway:WP262" 
#>  [15] "https://www.wikipathways.org/index.php/Pathway:WP4136"
#>  [16] "https://www.wikipathways.org/index.php/Pathway:WP3865"
#>  [17] "https://www.wikipathways.org/index.php/Pathway:WP3941"
#>  [18] "https://www.wikipathways.org/index.php/Pathway:WP3624"
#>  [19] "https://www.wikipathways.org/index.php/Pathway:WP2431"
#>  [20] "https://www.wikipathways.org/index.php/Pathway:WP2431"
#>  [21] "https://www.wikipathways.org/index.php/Pathway:WP2431"
#>  [22] "https://www.wikipathways.org/index.php/Pathway:WP2431"
#>  [23] "https://www.wikipathways.org/index.php/Pathway:WP3863"
#>  [24] "https://www.wikipathways.org/index.php/Pathway:WP4673"
#>  [25] "https://www.wikipathways.org/index.php/Pathway:WP3287"
#>  [26] "https://www.wikipathways.org/index.php/Pathway:WP4008"
#>  [27] "https://www.wikipathways.org/index.php/Pathway:WP4341"
#>  [28] "https://www.wikipathways.org/index.php/Pathway:WP4666"
#>  [29] "https://www.wikipathways.org/index.php/Pathway:WP4666"
#>  [30] "https://www.wikipathways.org/index.php/Pathway:WP2059"
#>  [31] "https://www.wikipathways.org/index.php/Pathway:WP2571"
#>  [32] "https://www.wikipathways.org/index.php/Pathway:WP1584"
#>  [33] "https://www.wikipathways.org/index.php/Pathway:WP4298"
#>  [34] "https://www.wikipathways.org/index.php/Pathway:WP4481"
#>  [35] "https://www.wikipathways.org/index.php/Pathway:WP530" 
#>  [36] "https://www.wikipathways.org/index.php/Pathway:WP530" 
#>  [37] "https://www.wikipathways.org/index.php/Pathway:WP530" 
#>  [38] "https://www.wikipathways.org/index.php/Pathway:WP530" 
#>  [39] "https://www.wikipathways.org/index.php/Pathway:WP530" 
#>  [40] "https://www.wikipathways.org/index.php/Pathway:WP530" 
#>  [41] "https://www.wikipathways.org/index.php/Pathway:WP530" 
#>  [42] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [43] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [44] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [45] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [46] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [47] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [48] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [49] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [50] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [51] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [52] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [53] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [54] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [55] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [56] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [57] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [58] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [59] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [60] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [61] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [62] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [63] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [64] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [65] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [66] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [67] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [68] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [69] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [70] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [71] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [72] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [73] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [74] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [75] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [76] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [77] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [78] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [79] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [80] "https://www.wikipathways.org/index.php/Pathway:WP3380"
#>  [81] "https://www.wikipathways.org/index.php/Pathway:WP2751"
#>  [82] "https://www.wikipathways.org/index.php/Pathway:WP4396"
#>  [83] "https://www.wikipathways.org/index.php/Pathway:WP4396"
#>  [84] "https://www.wikipathways.org/index.php/Pathway:WP4396"
#>  [85] "https://www.wikipathways.org/index.php/Pathway:WP4066"
#>  [86] "https://www.wikipathways.org/index.php/Pathway:WP4066"
#>  [87] "https://www.wikipathways.org/index.php/Pathway:WP3980"
#>  [88] "https://www.wikipathways.org/index.php/Pathway:WP3398"
#>  [89] "https://www.wikipathways.org/index.php/Pathway:WP3398"
#>  [90] "https://www.wikipathways.org/index.php/Pathway:WP3398"
#>  [91] "https://www.wikipathways.org/index.php/Pathway:WP3398"
#>  [92] "https://www.wikipathways.org/index.php/Pathway:WP3398"
#>  [93] "https://www.wikipathways.org/index.php/Pathway:WP3398"
#>  [94] "https://www.wikipathways.org/index.php/Pathway:WP3398"
#>  [95] "https://www.wikipathways.org/index.php/Pathway:WP3398"
#>  [96] "https://www.wikipathways.org/index.php/Pathway:WP4496"
#>  [97] "https://www.wikipathways.org/index.php/Pathway:WP3611"
#>  [98] "https://www.wikipathways.org/index.php/Pathway:WP3611"
#>  [99] "https://www.wikipathways.org/index.php/Pathway:WP3611"
#> [100] "https://www.wikipathways.org/index.php/Pathway:WP4493"
#> [101] "https://www.wikipathways.org/index.php/Pathway:WP4063"
#> [102] "https://www.wikipathways.org/index.php/Pathway:WP4478"
#> [103] "https://www.wikipathways.org/index.php/Pathway:WP231" 
#> [104] "https://www.wikipathways.org/index.php/Pathway:WP231" 
#> [105] "https://www.wikipathways.org/index.php/Pathway:WP231" 
#> [106] "https://www.wikipathways.org/index.php/Pathway:WP231" 
#> [107] "https://www.wikipathways.org/index.php/Pathway:WP231" 
#> [108] "https://www.wikipathways.org/index.php/Pathway:WP231" 
#> [109] "https://www.wikipathways.org/index.php/Pathway:WP3851"
#> [110] "https://www.wikipathways.org/index.php/Pathway:WP2036"
#> [111] "https://www.wikipathways.org/index.php/Pathway:WP2036"
#> [112] "https://www.wikipathways.org/index.php/Pathway:WP2036"
#> [113] "https://www.wikipathways.org/index.php/Pathway:WP75"  
#> [114] "https://www.wikipathways.org/index.php/Pathway:WP236" 
#> [115] "https://www.wikipathways.org/index.php/Pathway:WP254" 
#> [116] "https://www.wikipathways.org/index.php/Pathway:WP314" 
#> [117] "https://www.wikipathways.org/index.php/Pathway:WP560" 
#> [118] "https://www.wikipathways.org/index.php/Pathway:WP560" 
#> [119] "https://www.wikipathways.org/index.php/Pathway:WP1449"
#> [120] "https://www.wikipathways.org/index.php/Pathway:WP3617"
#> [121] "https://www.wikipathways.org/index.php/Pathway:WP3617"
#> [122] "https://www.wikipathways.org/index.php/Pathway:WP2328"
#> [123] "https://www.wikipathways.org/index.php/Pathway:WP129" 
#> [124] "https://www.wikipathways.org/index.php/Pathway:WP2513"
#> [125] "https://www.wikipathways.org/index.php/Pathway:WP2795"
#> [126] "https://www.wikipathways.org/index.php/Pathway:WP3297"
#> [127] "https://www.wikipathways.org/index.php/Pathway:WP706" 
#> [128] "https://www.wikipathways.org/index.php/Pathway:WP15"  
#> [129] "https://www.wikipathways.org/index.php/Pathway:WP15"  
#> [130] "https://www.wikipathways.org/index.php/Pathway:WP1533"
#> [131] "https://www.wikipathways.org/index.php/Pathway:WP1533"
#> [132] "https://www.wikipathways.org/index.php/Pathway:WP176" 
#> [133] "https://www.wikipathways.org/index.php/Pathway:WP176" 
#> [134] "https://www.wikipathways.org/index.php/Pathway:WP382" 
#> [135] "https://www.wikipathways.org/index.php/Pathway:WP727" 
#> [136] "https://www.wikipathways.org/index.php/Pathway:WP2586"
#> [137] "https://www.wikipathways.org/index.php/Pathway:WP4480"
#> [138] "https://www.wikipathways.org/index.php/Pathway:WP4304"
#> [139] "https://www.wikipathways.org/index.php/Pathway:WP4304"
#> [140] "https://www.wikipathways.org/index.php/Pathway:WP2864"
#> [141] "https://www.wikipathways.org/index.php/Pathway:WP4482"
#> [142] "https://www.wikipathways.org/index.php/Pathway:WP4754"
#> [143] "https://www.wikipathways.org/index.php/Pathway:WP4754"
#> [144] "https://www.wikipathways.org/index.php/Pathway:WP1544"
#> [145] "https://www.wikipathways.org/index.php/Pathway:WP4747"
#> [146] "https://www.wikipathways.org/index.php/Pathway:WP4816"
#> [147] "https://www.wikipathways.org/index.php/Pathway:WP4816"
#> [148] "https://www.wikipathways.org/index.php/Pathway:WP2447"
#> [149] "https://www.wikipathways.org/index.php/Pathway:WP2447"