Find Pathways By Text

Retrieve pathways containing the query text.

findPathwaysByText(query)

Arguments

query
A character string to search for, e.g., "cancer"

Value

A dataframe of pathway attributes in addition to query result score

Details

The score is from a lucene index search engine, ranging from 0 to 1 with higher scores for better matches.

Examples

{
findPathwaysByText('cancer')
}

#>           score     id                                                   url
#>   1: 1.48345970 WP1984 https://www.wikipathways.org/index.php/Pathway:WP1984
#>   2: 1.48345970 WP3260 https://www.wikipathways.org/index.php/Pathway:WP3260
#>   3: 1.29852510 WP2868 https://www.wikipathways.org/index.php/Pathway:WP2868
#>   4: 1.23793510 WP4262 https://www.wikipathways.org/index.php/Pathway:WP4262
#>   5: 1.20841570 WP3252 https://www.wikipathways.org/index.php/Pathway:WP3252
#>  ---                                                                        
#> 115: 0.02568957  WP566  https://www.wikipathways.org/index.php/Pathway:WP566
#> 116: 0.02201963 WP3265 https://www.wikipathways.org/index.php/Pathway:WP3265
#> 117: 0.02201963 WP2355 https://www.wikipathways.org/index.php/Pathway:WP2355
#> 118: 0.01834969 WP4857 https://www.wikipathways.org/index.php/Pathway:WP4857
#> 119: 0.01834969 WP4754 https://www.wikipathways.org/index.php/Pathway:WP4754
#>                                                                   name
#>   1:                                  Integrated Breast Cancer Pathway
#>   2:                                  Integrated Breast Cancer Pathway
#>   3: TCA Cycle Nutrient Utilization and Invasiveness of Ovarian Cancer
#>   4:                                             Breast cancer pathway
#>   5: TCA Cycle Nutrient Utilization and Invasiveness of Ovarian Cancer
#>  ---                                                                  
#> 115:                                           Canonical wnt signaling
#> 116:                                   Corticotropin-releasing hormone
#> 117:                 Corticotropin-releasing hormone signaling pathway
#> 118:                                                Endothelin Pathway
#> 119:                                           IL-18 signaling pathway
#>           species revision
#>   1: Homo sapiens   113594
#>   2:   Bos taurus   113595
#>   3: Homo sapiens   112051
#>   4: Homo sapiens   108103
#>   5:   Bos taurus   108105
#>  ---                      
#> 115:  Danio rerio    96484
#> 116:   Bos taurus   108255
#> 117: Homo sapiens   108254
#> 118: Homo sapiens   111629
#> 119: Homo sapiens   115147